Wald-type tests for moderators in psychmeta meta-analyses

This function computes Wald-type pairwise comparisons for each level of categorical moderators for an ma_psychmeta object, as well as an ombnibus one-way ANOVA test (equal variance not assumed).

Currently, samples across moderator levels are assumed to be indepdent.

# S3 method for ma_psychmeta
anova(
  object,
  ...,
  analyses = "all",
  moderators = NULL,
  L = NULL,
  ma_obj2 = NULL,
  ma_method = c("bb", "ic", "ad"),
  correction_type = c("ts", "vgx", "vgy"),
  conf_level = NULL
)

Arguments

object
A psychmeta meta-analysis object.

...
Additional arguments.

analyses
Which analyses to to test moderators for? Can be either "all" to test moderators for all meta-analyses in the object (default) or a list containing one or more of the arguments construct, construct_pair, pair_id, k_min, and N_min. See filter_ma() for details. Note that analysis_id should not be used. If k_min is not supplied, it is set to 2.

moderators
A character vector of moderators to test. If NULL, all categorical moderators are tested.

L
A named list with with elements specifying set of linear contrasts for each variable in moderators. (Not yet implemented.)

ma_obj2
A second psychmeta meta-analysis object to compare to object (Not yet implemented.)

ma_method
Meta-analytic methods to be included. Valid options are: "bb", "ic", and "ad"

correction_type
Types of meta-analytic corrections to be incldued. Valid options are: "ts", "vgx", and "vgy"

conf_level
Confidence level to define the width of confidence intervals (defaults to value set when object was fit)

Value

An object of class anova.ma_psychmeta. A tibble with a row for each construct pair in object and a column for each moderator tested. Cells lists of contrasts tested.

Note

Currently, only simple (single) categorical moderators (one-way ANOVA) are supported.

Examples

ma_obj <- ma_r(rxyi, n, construct_x = x_name, construct_y = y_name,
moderators = moderator, data = data_r_meas_multi)

#>  **** Running ma_r: Meta-analysis of correlations **** 

anova(ma_obj)

#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.

#>   pair_id construct_x construct_y moderator F value df_num df_denom level_1
#> 1       1           X           Y moderator    23.6      1     36.5       1
#> 2       2           X           Z moderator    43.6      1     37.8       1
#> 3       3           Y           Z moderator    28.7      1     37.6       1
#>   level_2 mean_1 mean_2  diff CI_LL_95 CI_UL_95
#> 1       2  0.397  0.245 0.152   0.0887    0.216
#> 2       2  0.422  0.236 0.186   0.1292    0.243
#> 3       2  0.405  0.228 0.178   0.1106    0.245