Scalar formula to estimate the correlation between a composite and another variable or between two composite variables

This function estimates the correlation between a set of X variables and a set of Y variables using a scalar formula.

composite_r_scalar(
  mean_rxy,
  k_vars_x = NULL,
  mean_intercor_x = NULL,
  k_vars_y = NULL,
  mean_intercor_y = NULL
)

Arguments

mean_rxy
Mean correlation between sets of X and Y variables.

k_vars_x
Number of X variables.

mean_intercor_x
Mean correlation among X variables.

k_vars_y
Number of Y variables.

mean_intercor_y
Mean correlation among Y variables.

Value

A vector of composite correlations

Details

The formula to estimate a correlation between one composite variable and one external variable is:

$$\rho_{Xy}=\frac{\bar{\rho}_{x_{i}y}}{\sqrt{\frac{1}{k_{x}}+\frac{k_{x}-1}{k_{x}}\bar{\rho}_{x_{i}x_{j}}}}$$

and the formula to estimate the correlation between two composite variables is:

$$\rho_{XY}=\frac{\bar{\rho}_{x_{i}y_{j}}}{\sqrt{\frac{1}{k_{x}}+\frac{k-1}{k_{x}}\bar{\rho}_{x_{i}x_{j}}}\sqrt{\frac{1}{k_{y}}+\frac{k_{y}-1}{k_{y}}\bar{\rho}_{y_{i}y_{j}}}}$$

where \(\bar{\rho}_{x_{i}y}\) and \(\bar{\rho}_{x_{i}y{j}}\) are mean correlations between the x variables and the y variable(s), \(\bar{\rho}_{x_{i}x_{j}}\) is the mean correlation among x variables, \(\bar{\rho}_{y_{i}y_{j}}\) is the mean correlation among y variables, \({k}_{x}\) is the number of x variables, and \({k}_{y}\) is the number of y variables.

References

Ghiselli, E. E., Campbell, J. P., & Zedeck, S. (1981). Measurement theory for the behavioral sciences. San Francisco, CA: Freeman. p. 163-164.

Schmidt, F. L., & Hunter, J. E. (2015). Methods of meta-analysis: Correcting error and bias in research findings (3rd ed.). Thousand Oaks, CA: Sage. https://doi.org/10/b6mg. pp. 441 - 447.

Examples

## Composite correlation between 4 variables and an outside variable with which
## the four variables correlate .3 on average:
composite_r_scalar(mean_rxy = .3, k_vars_x = 4, mean_intercor_x = .4)

#> [1] 0.4045199

## Correlation between two composites:
composite_r_scalar(mean_rxy = .3, k_vars_x = 2, mean_intercor_x = .5,
                   k_vars_y = 2, mean_intercor_y = .2)

#> [1] 0.4472136