Matrix formula to estimate the reliability of a weighted or unweighted composite variable

This function computes the reliability of a variable that is a weighted or unweighted composite of other variables.

composite_rel_matrix(rel_vec, r_mat, sd_vec, wt_vec = rep(1, length(rel_vec)))

Arguments

rel_vec
Vector of reliabilities associated with variables in the composite to be formed.

r_mat
Correlation matrix from which the composite is to be computed.

sd_vec
Vector of standard deviations associated with variables in the composite to be formed.

wt_vec
Weights to be used in forming the composite (by default, all variables receive equal weight).

Value

The estimated reliability of the composite variable.

Details

This function treats measure-specific variance as reliable.

The Mosier composite formula is computed as:

$$\rho_{XX}=\frac{\mathbf{w}^{T}\left(\mathbf{r}\circ\mathbf{s}\right)+\mathbf{w}^{T}\mathbf{S}\mathbf{w}-\mathbf{w}^{T}\mathbf{s}}{\mathbf{w}^{T}\mathbf{S}\mathbf{w}}$$

where \(\rho_{XX}\) is a composite reliability estimate, \(\mathbf{r}\) is a vector of reliability estimates, \(\mathbf{w}\) is a vector of weights, \(\mathbf{S}\) is a covariance matrix, and \(\mathbf{s}\) is a vector of variances (i.e., the diagonal elements of \(\mathbf{S}\)).

References

Mosier, C. I. (1943). On the reliability of a weighted composite. Psychometrika, 8(3), 161–168. https://doi.org/10.1007/BF02288700

Schmidt, F. L., & Hunter, J. E. (2015). Methods of meta-analysis: Correcting error and bias in research findings (3rd ed.). Thousand Oaks, CA: Sage. https://doi.org/10/b6mg. pp. 441 - 447.

Examples

composite_rel_matrix(rel_vec = c(.8, .8),
r_mat = matrix(c(1, .4, .4, 1), 2, 2), sd_vec = c(1, 1))

#> [1] 0.8571429