Matrix formula to estimate the u ratio of a composite variable

This function estimates the u ratio of a composite variable when at least one matrix of correlations (restricted or unrestricted) among the variables is available.

composite_u_matrix(
  ri_mat = NULL,
  ra_mat = NULL,
  u_vec,
  wt_vec = rep(1, length(u_vec)),
  sign_r_vec = 1
)

Arguments

ri_mat
Range-restricted correlation matrix from which the composite is to be computed (if NULL, ri_mat is estimated from ra_mat).

ra_mat
Unrestricted correlation matrix from which the composite is to be computed (if NULL, ra_mat is estimated from ri_mat).

u_vec
Vector of u ratios associated with variables in the composite to be formed.

wt_vec
Weights to be used in forming the composite (by default, all variables receive equal weight).

sign_r_vec
The signs of the relationships between the variables in the composite and the variable by which range restriction was induced.

Value

The estimated u ratio of the composite variable.

Details

This is computed as:

$$u_{composite}=\sqrt{\frac{\left(\mathbf{w}\circ\mathbf{u}\right)^{T}\mathbf{R}_{i}\left(\mathbf{w}\circ\mathbf{u}\right)}{\mathbf{w}^{T}\mathbf{R}_{a}\mathbf{w}}}$$

where \(u_{composite}\) is the composite u ratio, \(\mathbf{u}\) is a vector of u ratios, \(\mathbf{R}_{i}\) is a range-restricted correlation matrix, \(\mathbf{R}_{a}\) is an unrestricted correlation matrix, and \(\mathbf{w}\) is a vector of weights.

Examples

composite_u_matrix(ri_mat = matrix(c(1, .3, .3, 1), 2, 2), u_vec = c(.8, .8))

#> [1] 0.7321765