Comprehensive \(d_{Mod}\) calculator

This is a general-purpose function to compute \(d_{Mod}\) effect sizes from raw data and to perform bootstrapping. It subsumes the functionalities of the compute_dmod_par (for parametric computations) and compute_dmod_npar (for non-parametric computations) functions and automates the generation of regression equations and descriptive statistics for computing \(d_{Mod}\) effect sizes. Please see documentation for compute_dmod_par and compute_dmod_npar for details about how the effect sizes are computed.

compute_dmod(
  data,
  group,
  predictors,
  criterion,
  referent_id,
  focal_id_vec = NULL,
  conf_level = 0.95,
  rescale_cdf = TRUE,
  parametric = TRUE,
  bootstrap = TRUE,
  boot_iter = 1000,
  stratify = FALSE,
  empirical_ci = FALSE,
  cross_validate_wts = FALSE
)

Arguments

data
Data frame containing the data to be analyzed (if not a data frame, must be an object convertible to a data frame via the as.data.frame function). The data set must contain a criterion variable, at least one predictor variable, and a categorical variable that identifies the group to which each case (i.e., row) in the data set belongs.

group
Name or column-index number of the variable that identifies group membership in the data set.

predictors
Name(s) or column-index number(s) of the predictor variable(s) in the data set. No predictor can be a factor-type variable. If multiple predictors are specified, they will be combined into a regression-weighted composite that will be carried forward to compute \(d_{Mod}\) effect sizes.

Note: If weights other than regression weights should be used to combine the predictors into a composite, the user must manually compute such a composite, include the composite in the dat data set, and identify the composite variable in predictors.

criterion
Name or column-index number of the criterion variable in the data set. The criterion cannot be a factor-type variable.

referent_id
Label used to identify the referent group in the group variable.

focal_id_vec
Label(s) to identify the focal group(s) in the group variable. If NULL (the default), the specified referent group will be compared to all other groups.

conf_level
Confidence level (between 0 and 1) to be used in generating confidence intervals. Default is .95

rescale_cdf
Logical argument that indicates whether parametric \(d_{Mod}\) results should be rescaled to account for using a cumulative density < 1 in the computations (TRUE; default) or not (FALSE).

parametric
Logical argument that indicates whether \(d_{Mod}\) should be computed using an assumed normal distribution (TRUE; default) or observed frequencies (FALSE).

bootstrap
Logical argument that indicates whether \(d_{Mod}\) should be bootstrapped (TRUE; default) or not (FALSE).

boot_iter
Number of bootstrap iterations to compute (default = 1000).

stratify
Logical argument that indicates whether the random bootstrap sampling should be stratified (TRUE) or unstratified (FALSE; default).

empirical_ci
Logical argument that indicates whether the bootstrapped confidence invervals should be computed from the observed empirical distributions (TRUE) or computed using bootstrapped means and standard errors via the normal-theory approach (FALSE).

cross_validate_wts
Only relevant when multiple predictors are specified and bootstrapping is performed. Logical argument that indicates whether regression weights derived from the full sample should be used to combine predictors in the bootstrapped samples (TRUE) or if a new set of weights should be derived during each iteration of the bootstrapping procedure (FALSE; default).

Value

If bootstrapping is selected, the list will include:

point_estimate A matrix of effect sizes (\(d_{Mod_{Signed}}\), \(d_{Mod_{Unsigned}}\), \(d_{Mod_{Under}}\), \(d_{Mod_{Over}}\)), proportions of under- and over-predicted criterion scores, minimum and maximum differences, and the scores associated with minimum and maximum differences. All of these values are computed using the full data set.

bootstrap_mean A matrix of the same statistics as the point_estimate matrix, but the values in this matrix are the means of the results from bootstrapped samples.

bootstrap_se A matrix of the same statistics as the point_estimate matrix, but the values in this matrix are bootstrapped standard errors (i.e., the standard deviations of the results from bootstrapped samples).

bootstrap_CI_Lo A matrix of the same statistics as the point_estimate matrix, but the values in this matrix are the lower confidence bounds of the results from bootstrapped samples.

bootstrap_CI_Hi A matrix of the same statistics as the point_estimate matrix, but the values in this matrix are the upper confidence bounds of the results from bootstrapped samples.

If no bootstrapping is performed, the output will be limited to the point_estimate matrix.

References

Nye, C. D., & Sackett, P. R. (2016). New effect sizes for tests of categorical moderation and differential prediction. Organizational Research Methods, https://doi.org/10.1177/1094428116644505.

Examples

# Generate some hypothetical data for a referent group and three focal groups:
set.seed(10)
refDat <- MASS::mvrnorm(n = 1000, mu = c(.5, .2),
                        Sigma = matrix(c(1, .5, .5, 1), 2, 2), empirical = TRUE)
foc1Dat <- MASS::mvrnorm(n = 1000, mu = c(-.5, -.2),
                         Sigma = matrix(c(1, .5, .5, 1), 2, 2), empirical = TRUE)
foc2Dat <- MASS::mvrnorm(n = 1000, mu = c(0, 0),
                         Sigma = matrix(c(1, .3, .3, 1), 2, 2), empirical = TRUE)
foc3Dat <- MASS::mvrnorm(n = 1000, mu = c(-.5, -.2),
                         Sigma = matrix(c(1, .3, .3, 1), 2, 2), empirical = TRUE)
colnames(refDat) <- colnames(foc1Dat) <- colnames(foc2Dat) <- colnames(foc3Dat) <- c("X", "Y")
dat <- rbind(cbind(G = 1, refDat), cbind(G = 2, foc1Dat),
             cbind(G = 3, foc2Dat), cbind(G = 4, foc3Dat))

# Compute point estimates of parametric d_mod effect sizes:
compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y",
     referent_id = 1, focal_id_vec = 2:4,
     conf_level = .95, rescale_cdf = TRUE, parametric = TRUE,
     bootstrap = FALSE)

#> 
#> Call:
#> compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y", 
#>     referent_id = 1, focal_id_vec = 2:4, conf_level = 0.95, rescale_cdf = TRUE, 
#>     parametric = TRUE, bootstrap = FALSE)
#> 
#> Point Estimates:
#>   focal_group d_signed d_unsigned d_under d_over prop_under prop_over     d_min
#> 1           2  -0.1000      0.100  -0.100 0.0000      1.000     0.000 -1.00e-01
#> 2           3  -0.0503      0.164  -0.107 0.0567      0.599     0.401  8.69e-16
#> 3           4  -0.0996      0.179  -0.139 0.0395      0.691     0.309  7.08e-16
#>    d_max d_min_score d_max_score
#> 1 -0.100          NA          NA
#> 2 -0.731        0.25       -3.41
#> 3 -0.727        0.00       -3.63

# Compute point estimates of non-parametric d_mod effect sizes:
compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y",
     referent_id = 1, focal_id_vec = 2:4,
     conf_level = .95, rescale_cdf = TRUE, parametric = FALSE,
     bootstrap = FALSE)

#> 
#> Call:
#> compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y", 
#>     referent_id = 1, focal_id_vec = 2:4, conf_level = 0.95, rescale_cdf = TRUE, 
#>     parametric = FALSE, bootstrap = FALSE)
#> 
#> Point Estimates:
#>   focal_group d_signed d_unsigned d_under d_over prop_under prop_over     d_min
#> 1           2    -0.10      0.100  -0.100 0.0000      1.000     0.000 -1.00e-01
#> 2           3    -0.05      0.165  -0.107 0.0574      0.609     0.391  4.27e-05
#> 3           4    -0.10      0.179  -0.140 0.0396      0.690     0.310 -3.64e-04
#>    d_max d_min_score d_max_score
#> 1 -0.100          NA          NA
#> 2 -0.731     0.25021       -3.41
#> 3 -0.727    -0.00182       -3.63

# Compute unstratified bootstrapped estimates of parametric d_mod effect sizes:
compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y",
     referent_id = 1, focal_id_vec = 2:4,
     conf_level = .95, rescale_cdf = TRUE, parametric = TRUE,
     boot_iter = 10, bootstrap = TRUE, stratify = FALSE, empirical_ci = FALSE)

#> 
#> Call:
#> compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y", 
#>     referent_id = 1, focal_id_vec = 2:4, conf_level = 0.95, rescale_cdf = TRUE, 
#>     parametric = TRUE, bootstrap = TRUE, boot_iter = 10, stratify = FALSE, 
#>     empirical_ci = FALSE)
#> 
#> Point Estimates:
#>   focal_group d_signed d_unsigned d_under d_over prop_under prop_over     d_min
#> 1           2  -0.1000      0.100  -0.100 0.0000      1.000     0.000 -1.00e-01
#> 2           3  -0.0503      0.164  -0.107 0.0567      0.599     0.401  8.69e-16
#> 3           4  -0.0996      0.179  -0.139 0.0395      0.691     0.309  7.08e-16
#>    d_max d_min_score d_max_score
#> 1 -0.100          NA          NA
#> 2 -0.731        0.25       -3.41
#> 3 -0.727        0.00       -3.63
#> 
#> Mean Bootstrapped Values:
#>   focal_group d_signed d_unsigned d_under   d_over prop_under prop_over
#> 1           2  -0.0811     0.0819 -0.0815 0.000419      0.947    0.0526
#> 2           3  -0.0581     0.1611 -0.1096 0.051477      0.620    0.3803
#> 3           4  -0.0952     0.1695 -0.1323 0.037132      0.695    0.3053
#>       d_min  d_max d_min_score d_max_score
#> 1 -1.01e-02 -0.166     -1.7922        0.56
#> 2  9.66e-08 -0.647      0.2911       -3.09
#> 3  3.36e-08 -0.509      0.0128       -2.92
#> 
#> Bootstrapped Standard Errors:
#>   focal_group d_signed d_unsigned d_under  d_over prop_under prop_over    d_min
#> 1           2   0.0375     0.0364  0.0369 0.00072     0.1250    0.1250 1.74e-02
#> 2           3   0.0484     0.0304  0.0342 0.02151     0.1051    0.1051 5.80e-07
#> 3           4   0.0475     0.0485  0.0449 0.01702     0.0653    0.0653 2.85e-07
#>   d_max d_min_score d_max_score
#> 1 0.091       2.158       2.976
#> 2 0.162       0.257       0.351
#> 3 0.498       0.191       2.122
#> 
#> Bootstrapped Lower-Bound Confidence Limit:
#>   focal_group d_signed d_unsigned d_under    d_over prop_under prop_over
#> 1           2   -0.155     0.0107  -0.154 -0.000991      0.702    -0.192
#> 2           3   -0.153     0.1016  -0.177  0.009314      0.414     0.174
#> 3           4   -0.188     0.0743  -0.220  0.003774      0.567     0.177
#>       d_min  d_max d_min_score d_max_score
#> 1 -4.42e-02 -0.345      -6.021       -5.27
#> 2 -1.04e-06 -0.965      -0.213       -3.78
#> 3 -5.25e-07 -1.485      -0.362       -7.07
#> 
#> Bootstrapped Upper-Bound Confidence Limit:
#>   focal_group d_signed d_unsigned  d_under  d_over prop_under prop_over
#> 1           2 -0.00764      0.153 -0.00916 0.00183      1.192     0.298
#> 2           3  0.03674      0.221 -0.04259 0.09364      0.826     0.586
#> 3           4 -0.00218      0.265 -0.04437 0.07049      0.823     0.433
#>      d_min   d_max d_min_score d_max_score
#> 1 2.41e-02  0.0119       2.437        6.39
#> 2 1.23e-06 -0.3292       0.795       -2.40
#> 3 5.92e-07  0.4682       0.388        1.24

# Compute unstratified bootstrapped estimates of non-parametric d_mod effect sizes:
compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y",
     referent_id = 1, focal_id_vec = 2:4,
     conf_level = .95, rescale_cdf = TRUE, parametric = FALSE,
     boot_iter = 10, bootstrap = TRUE, stratify = FALSE, empirical_ci = FALSE)

#> 
#> Call:
#> compute_dmod(data = dat, group = "G", predictors = "X", criterion = "Y", 
#>     referent_id = 1, focal_id_vec = 2:4, conf_level = 0.95, rescale_cdf = TRUE, 
#>     parametric = FALSE, bootstrap = TRUE, boot_iter = 10, stratify = FALSE, 
#>     empirical_ci = FALSE)
#> 
#> Point Estimates:
#>   focal_group d_signed d_unsigned d_under d_over prop_under prop_over     d_min
#> 1           2    -0.10      0.100  -0.100 0.0000      1.000     0.000 -1.00e-01
#> 2           3    -0.05      0.165  -0.107 0.0574      0.609     0.391  4.27e-05
#> 3           4    -0.10      0.179  -0.140 0.0396      0.690     0.310 -3.64e-04
#>    d_max d_min_score d_max_score
#> 1 -0.100          NA          NA
#> 2 -0.731     0.25021       -3.41
#> 3 -0.727    -0.00182       -3.63
#> 
#> Mean Bootstrapped Values:
#>   focal_group d_signed d_unsigned d_under  d_over prop_under prop_over
#> 1           2  -0.1002      0.105  -0.103 0.00245      0.963     0.037
#> 2           3  -0.0526      0.166  -0.109 0.05668      0.611     0.389
#> 3           4  -0.0944      0.189  -0.141 0.04704      0.665     0.335
#>       d_min  d_max d_min_score d_max_score
#> 1 -4.18e-02 -0.193      -0.491      -0.726
#> 2 -2.07e-04 -0.674       0.250      -3.183
#> 3 -4.84e-05 -0.620      -0.056      -2.878
#> 
#> Bootstrapped Standard Errors:
#>   focal_group d_signed d_unsigned d_under  d_over prop_under prop_over    d_min
#> 1           2   0.0412     0.0366  0.0382 0.00758     0.1051    0.1051 0.043406
#> 2           3   0.0400     0.0243  0.0254 0.02129     0.0810    0.0810 0.000391
#> 3           4   0.0491     0.0374  0.0396 0.01842     0.0822    0.0822 0.000568
#>   d_max d_min_score d_max_score
#> 1 0.089       2.925       3.309
#> 2 0.119       0.239       0.296
#> 3 0.425       0.215       2.109
#> 
#> Bootstrapped Lower-Bound Confidence Limit:
#>   focal_group d_signed d_unsigned d_under  d_over prop_under prop_over
#> 1           2   -0.181     0.0334  -0.178 -0.0124      0.757    -0.169
#> 2           3   -0.131     0.1182  -0.159  0.0150      0.452     0.230
#> 3           4   -0.191     0.1153  -0.219  0.0109      0.504     0.174
#>       d_min  d_max d_min_score d_max_score
#> 1 -0.126891 -0.368      -6.224       -7.21
#> 2 -0.000973 -0.907      -0.219       -3.76
#> 3 -0.001162 -1.453      -0.478       -7.01
#> 
#> Bootstrapped Upper-Bound Confidence Limit:
#>   focal_group d_signed d_unsigned d_under d_over prop_under prop_over    d_min
#> 1           2 -0.01940      0.177 -0.0277 0.0173      1.169     0.243 0.043257
#> 2           3  0.02591      0.214 -0.0595 0.0984      0.770     0.548 0.000559
#> 3           4  0.00182      0.262 -0.0639 0.0832      0.826     0.496 0.001065
#>     d_max d_min_score d_max_score
#> 1 -0.0188       5.242        5.76
#> 2 -0.4406       0.719       -2.60
#> 3  0.2118       0.366        1.26