Function for computing parametric \(d_{Mod}\) effect sizes for any number of focal groups

This function computes \(d_{Mod}\) effect sizes from user-defined descriptive statistics and regression coefficients. If one has access to a raw data set, the dMod function may be used as a wrapper to this function so that the regression equations and descriptive statistics can be computed automatically within the program.

compute_dmod_par(
  referent_int,
  referent_slope,
  focal_int,
  focal_slope,
  focal_mean_x,
  focal_sd_x,
  referent_sd_y,
  focal_min_x,
  focal_max_x,
  focal_names = NULL,
  rescale_cdf = TRUE
)

Arguments

referent_int
Referent group's intercept.

referent_slope
Referent group's slope.

focal_int
Focal groups' intercepts.

focal_slope
Focal groups' slopes.

focal_mean_x
Focal groups' predictor-score means.

focal_sd_x
Focal groups' predictor-score standard deviations.

referent_sd_y
Referent group's criterion standard deviation.

focal_min_x
Focal groups' minimum predictor scores.

focal_max_x
Focal groups' maximum predictor scores.

focal_names
Focal-group names. If NULL (the default), the focal groups will be given numeric labels ranging from 1 through the number of groups.

rescale_cdf
Logical argument that indicates whether parametric \(d_{Mod}\) results should be rescaled to account for using a cumulative density < 1 in the computations (TRUE; default) or not (FALSE).

Value

A matrix of effect sizes (\(d_{Mod_{Signed}}\), \(d_{Mod_{Unsigned}}\), \(d_{Mod_{Under}}\), \(d_{Mod_{Over}}\)), proportions of under- and over-predicted criterion scores, minimum and maximum differences (i.e., \(d_{Mod_{Under}}\) and \(d_{Mod_{Over}}\)), and the scores associated with minimum and maximum differences. Note that if the regression lines are parallel and infinite focal_min_x and focal_max_x values were specified, the extrema will be defined using the scores 3 focal-group SDs above and below the corresponding focal-group means.

Details

The \(d_{Mod_{Signed}}\) effect size (i.e., the average of differences in prediction over the range of predictor scores) is computed as $$d_{Mod_{Signed}}=\frac{1}{SD_{Y_{1}}}\intop f_{2}(X)\left[X\left(b_{1_{1}}-b_{1_{2}}\right)+b_{0_{1}}-b_{0_{2}}\right] dX,$$ where

\(SD_{Y_{1}}\) is the referent group's criterion standard deviation;

\(f_{2}(X)\) is the normal-density function for the distribution of focal-group predictor scores;

\(b_{1_{1}}\) and \(b_{1_{0}}\) are the slopes of the regression of \(Y\) on \(X\) for the referent and focal groups, respectively;

\(b_{0_{1}}\) and \(b_{0_{0}}\) are the intercepts of the regression of \(Y\) on \(X\) for the referent and focal groups, respectively; and

the integral spans all \(X\) scores within the operational range of predictor scores for the focal group.

The \(d_{Mod_{Under}}\) and \(d_{Mod_{Over}}\) effect sizes are computed using the same equation as \(d_{Mod_{Signed}}\), but \(d_{Mod_{Under}}\) is the weighted average of all scores in the area of underprediction (i.e., the differences in prediction with negative signs) and \(d_{Mod_{Over}}\) is the weighted average of all scores in the area of overprediction (i.e., the differences in prediction with negative signs).

The \(d_{Mod_{Unsigned}}\) effect size (i.e., the average of absolute differences in prediction over the range of predictor scores) is computed as $$d_{Mod_{Unsigned}}=\frac{1}{SD_{Y_{1}}}\intop f_{2}(X)\left|X\left(b_{1_{1}}-b_{1_{2}}\right)+b_{0_{1}}-b_{0_{2}}\right|dX.$$

The \(d_{Min}\) effect size (i.e., the smallest absolute difference in prediction observed over the range of predictor scores) is computed as $$d_{Min}=\frac{1}{SD_{Y_{1}}}Min\left[\left|X\left(b_{1_{1}}-b_{1_{2}}\right)+b_{0_{1}}-b_{0_{2}}\right|\right].$$

The \(d_{Max}\) effect size (i.e., the largest absolute difference in prediction observed over the range of predictor scores)is computed as $$d_{Max}=\frac{1}{SD_{Y_{1}}}Max\left[\left|X\left(b_{1_{1}}-b_{1_{2}}\right)+b_{0_{1}}-b_{0_{2}}\right|\right].$$ Note: When \(d_{Min}\) and \(d_{Max}\) are computed in this package, the output will display the signs of the differences (rather than the absolute values of the differences) to aid in interpretation.

If \(d_{Mod}\) effect sizes are to be rescaled to compensate for a cumulative density less than 1 (see the rescale_cdf argument), the result of each effect size involving integration will be divided by the ratio of the cumulative density of the observed range of scores (i.e., the range bounded by the focal_min_x and focal_max_x arguments) to the cumulative density of scores bounded by -Inf and Inf.

References

Nye, C. D., & Sackett, P. R. (2016). New effect sizes for tests of categorical moderation and differential prediction. Organizational Research Methods, https://doi.org/10.1177/1094428116644505.

Examples

compute_dmod_par(referent_int = -.05, referent_slope = .5,
                 focal_int = c(.05, 0, -.05), focal_slope = c(.5, .3, .3),
                 focal_mean_x = c(-.5, 0, -.5), focal_sd_x = rep(1, 3),
                 referent_sd_y = 1,
                 focal_min_x = rep(-Inf, 3), focal_max_x = rep(Inf, 3),
                 focal_names = NULL, rescale_cdf = TRUE)

#> 
#> Call:
#> compute_dmod_par(referent_int = -0.05, referent_slope = 0.5, 
#>     focal_int = c(0.05, 0, -0.05), focal_slope = c(0.5, 0.3, 
#>         0.3), focal_mean_x = c(-0.5, 0, -0.5), focal_sd_x = rep(1, 
#>         3), referent_sd_y = 1, focal_min_x = rep(-Inf, 3), focal_max_x = rep(Inf, 
#>         3), focal_names = NULL, rescale_cdf = TRUE)
#> 
#> Point Estimates:
#>   focal_group d_signed d_unsigned d_under d_over prop_under prop_over d_min
#> 1           1    -0.10      0.100  -0.100 0.0000      1.000     0.000  -0.1
#> 2           2    -0.05      0.165  -0.107 0.0573      0.599     0.401   0.0
#> 3           3    -0.10      0.179  -0.140 0.0396      0.691     0.309   0.0
#>   d_max d_min_score d_max_score
#> 1 -0.10          NA          NA
#> 2 -0.65        0.25        -3.0
#> 3 -0.70        0.00        -3.5