Confidence limits for noncentral chi square parameters (function and documentation from package 'MBESS' version 4.4.3) Function to determine the noncentral parameter that leads to the observed Chi.Square-value, so that a confidence interval for the population noncentral chi-squrae value can be formed.

Confidence limits for noncentral chi square parameters (function and documentation from package 'MBESS' version 4.4.3) Function to determine the noncentral parameter that leads to the observed Chi.Square-value, so that a confidence interval for the population noncentral chi-squrae value can be formed.

conf.limits.nc.chisq(
  Chi.Square = NULL,
  conf.level = 0.95,
  df = NULL,
  alpha.lower = NULL,
  alpha.upper = NULL,
  tol = 1e-09,
  Jumping.Prop = 0.1
)

Arguments

Chi.Square
the observed chi-square value

conf.level
the desired degree of confidence for the interval

df
the degrees of freedom

alpha.lower
Type I error for the lower confidence limit

alpha.upper
Type I error for the upper confidence limit

tol
tolerance for iterative convergence

Jumping.Prop
Value used in the iterative scheme to determine the noncentral parameters necessary for confidence interval construction using noncentral chi square-distributions (0 < Jumping.Prop < 1)

Value

Lower.LimitValue of the distribution with Lower.Limit noncentral value that has at its specified quantile Chi.Square

Prob.Less.LowerProportion of cases falling below Lower.Limit

Upper.LimitValue of the distribution with Upper.Limit noncentral value that has at its specified quantile Chi.Square

Prob.Greater.UpperProportion of cases falling above Upper.Limit

Details

If the function fails (or if a function relying upon this function fails), adjust the Jumping.Prop (to a smaller value).

Author

Ken Kelley (University of Notre Dame; KKelley@ND.edu), Keke Lai (University of California--Merced)