Control for psychmeta meta-analyses

Control for psychmeta meta-analyses

control_psychmeta(
  error_type = c("mean", "sample"),
  conf_level = 0.95,
  cred_level = 0.8,
  conf_method = c("t", "norm"),
  cred_method = c("t", "norm"),
  var_unbiased = TRUE,
  pairwise_ads = FALSE,
  moderated_ads = FALSE,
  residual_ads = TRUE,
  check_dependence = TRUE,
  collapse_method = c("composite", "average", "stop"),
  intercor = control_intercor(),
  clean_artifacts = TRUE,
  impute_artifacts = TRUE,
  impute_method = c("bootstrap_mod", "bootstrap_full", "simulate_mod", "simulate_full",
    "wt_mean_mod", "wt_mean_full", "unwt_mean_mod", "unwt_mean_full", "replace_unity",
    "stop"),
  seed = 42,
  use_all_arts = TRUE,
  estimate_pa = FALSE,
  decimals = 2,
  hs_override = FALSE,
  ...
)

Arguments

error_type
Method to be used to estimate error variances: "mean" uses the mean effect size to estimate error variances and "sample" uses the sample-specific effect sizes.

conf_level
Confidence level to define the width of the confidence interval (default = .95).

cred_level
Credibility level to define the width of the credibility interval (default = .80).

conf_method
Distribution to be used to compute the width of confidence intervals. Available options are "t" for t distribution or "norm" for normal distribution.

cred_method
Distribution to be used to compute the width of credibility intervals. Available options are "t" for t distribution or "norm" for normal distribution.

var_unbiased
Logical scalar determining whether variances should be unbiased (TRUE) or maximum-likelihood (FALSE).

pairwise_ads
Logical value that determines whether to compute artifact distributions in a construct-pair-wise fashion (TRUE) or separately by construct (FALSE, default).

moderated_ads
Logical value that determines whether to compute artifact distributions separately for each moderator combination (TRUE) or for overall analyses only (FALSE, default).

residual_ads
Logical argument that determines whether to use residualized variances (TRUE) or observed variances (FALSE) of artifact distributions to estimate sd_rho.

check_dependence
Logical scalar that determines whether database should be checked for violations of independence (TRUE) or not (FALSE).

collapse_method
Character argument that determines how to collapse dependent studies. Options are "composite" (default), "average," and "stop."

intercor
The intercorrelation(s) among variables to be combined into a composite. Can be a scalar, a named vector with element named according to the names of constructs, or output from the control_intercor function. Default scalar value is .5.

clean_artifacts
If TRUE, multiple instances of the same construct (or construct-measure pair, if measure is provided) in the database are compared and reconciled with each other in the case that any of the matching entries within a study have different artifact values. When impute_method is anything other than "stop", this method is always implemented to prevent discrepancies among imputed values.

impute_artifacts
If TRUE, artifact imputation will be performed (see impute_method for imputation procedures). Default is FALSE for artifact-distribution meta-analyses and TRUE otherwise. When imputation is performed, clean_artifacts is treated as TRUE so as to resolve all discrepancies among artifact entries before and after imputation.

impute_method
Method to use for imputing artifacts. Choices are:

bootstrap_mod
Select random values from the most specific moderator categories available (default).

bootstrap_full
Select random values from the full vector of artifacts.

simulate_mod
Generate random values from the distribution with the mean and variance of observed artifacts from the most specific moderator categories available. (uses rnorm for u ratios and rbeta for reliability values).

simulate_full
Generate random values from the distribution with the mean and variance of all observed artifacts (uses rnorm for u ratios and rbeta for reliability values).

wt_mean_mod
Replace missing values with the sample-size weighted mean of the distribution of artifacts from the most specific moderator categories available (not recommended).

wt_mean_full
Replace missing values with the sample-size weighted mean of the full distribution of artifacts (not recommended).

unwt_mean_mod
Replace missing values with the unweighted mean of the distribution of artifacts from the most specific moderator categories available (not recommended).

unwt_mean_full
Replace missing values with the unweighted mean of the full distribution of artifacts (not recommended).

replace_unity
Replace missing values with 1 (not recommended).

stop
Stop evaluations when missing artifacts are encountered.

If an imputation method ending in "mod" is selected but no moderators are provided, the "mod" suffix will internally be replaced with "full".

seed
Seed value to use for imputing artifacts in a reproducible way. Default value is 42.

use_all_arts
Logical scalar that determines whether artifact values from studies without valid effect sizes should be used in artifact distributions (TRUE; default) or not (FALSE).

estimate_pa
Logical scalar that determines whether the unrestricted subgroup proportions associated with univariate-range-restricted effect sizes should be estimated by rescaling the range-restricted subgroup proportions as a function of the range-restriction correction (TRUE) or not (FALSE; default).

decimals
Number of decimal places to which interactive artifact distributions should be rounded (default is 2 decimal places).

hs_override
When TRUE, this will override settings for wt_type (will set to "sample_size"), error_type (will set to "mean"), correct_bias (will set to TRUE), conf_method (will set to "norm"), cred_method (will set to "norm"), var_unbiased (will set to FALSE), residual_ads (will be set to FALSE), and use_all_arts (will set to FALSE).

...
Further arguments to be passed to functions called within the meta-analysis.

Value

A list of control arguments in the package environment.

Examples

control_psychmeta()

#> $error_type
#> [1] "mean"
#> 
#> $conf_level
#> [1] 0.95
#> 
#> $cred_level
#> [1] 0.8
#> 
#> $conf_method
#> [1] "t"
#> 
#> $cred_method
#> [1] "t"
#> 
#> $var_unbiased
#> [1] TRUE
#> 
#> $pairwise_ads
#> [1] FALSE
#> 
#> $moderated_ads
#> [1] FALSE
#> 
#> $residual_ads
#> [1] TRUE
#> 
#> $check_dependence
#> [1] TRUE
#> 
#> $collapse_method
#> [1] "composite"
#> 
#> $intercor
#> [1] 0.5
#> attr(,"class")
#> [1] "control_intercor"
#> 
#> $clean_artifacts
#> [1] TRUE
#> 
#> $impute_artifacts
#> [1] TRUE
#> 
#> $impute_method
#> [1] "bootstrap_mod"
#> 
#> $seed
#> [1] 42
#> 
#> $use_all_arts
#> [1] TRUE
#> 
#> $estimate_pa
#> [1] FALSE
#> 
#> $decimals
#> [1] 2
#> 
#> $hs_override
#> [1] FALSE
#>