Correct d values for measurement error and/or range restriction

This function is a wrapper for the correct_r() function to correct d values for statistical and psychometric artifacts.

correct_d(
  correction = c("meas", "uvdrr_g", "uvdrr_y", "uvirr_g", "uvirr_y", "bvdrr", "bvirr"),
  d,
  ryy = 1,
  uy = 1,
  rGg = 1,
  pi = NULL,
  pa = NULL,
  uy_observed = TRUE,
  ryy_restricted = TRUE,
  ryy_type = "alpha",
  k_items_y = NA,
  sign_rgz = 1,
  sign_ryz = 1,
  n1 = NULL,
  n2 = NA,
  conf_level = 0.95,
  correct_bias = FALSE
)

Arguments

correction
Type of correction to be applied. Options are "meas", "uvdrr_g", "uvdrr_y", "uvirr_g", "uvirr_y", "bvdrr", "bvirr"

d
Vector of d values.

ryy
Vector of reliability coefficients for Y (the continuous variable).

uy
Vector of u ratios for Y (the continuous variable).

rGg
Vector of reliabilities for the group variable (i.e., the correlations between observed group membership and latent group membership).

pi
Proportion of cases in one of the groups in the observed data (not necessary if n1 and n2 reflect this proportionality).

pa
Proportion of cases in one of the groups in the population.

uy_observed
Logical vector in which each entry specifies whether the corresponding uy value is an observed-score u ratio (TRUE) or a true-score u ratio. All entries are TRUE by default.

ryy_restricted
Logical vector in which each entry specifies whether the corresponding rxx value is an incumbent reliability (TRUE) or an applicant reliability. All entries are TRUE by default.

ryy_type
String vector identifying the types of reliability estimates supplied (e.g., "alpha", "retest", "interrater_r", "splithalf"). See the documentation for ma_r() for a full list of acceptable reliability types.

k_items_y
Numeric vector identifying the number of items in each scale.

sign_rgz
Vector of signs of the relationships between grouping variables and the selection mechanism.

sign_ryz
Vector of signs of the relationships between Y variables and the selection mechanism.

n1
Optional vector of sample sizes associated with group 1 (or the total sample size, if n2 is NULL).

n2
Optional vector of sample sizes associated with group 2.

conf_level
Confidence level to define the width of the confidence interval (default = .95).

correct_bias
Logical argument that determines whether to correct error-variance estimates for small-sample bias in correlations (TRUE) or not (FALSE). For sporadic corrections (e.g., in mixed artifact-distribution meta-analyses), this should be set to FALSE (the default).

Value

Data frame(s) of observed d values (dgyi), operational range-restricted d values corrected for measurement error in Y only (dgpi), operational range-restricted d values corrected for measurement error in the grouping only (dGyi), and range-restricted true-score d values (dGpi), range-corrected observed-score d values (dgya), operational range-corrected d values corrected for measurement error in Y only (dgpa), operational range-corrected d values corrected for measurement error in the grouping only (dGya), and range-corrected true-score d values (dGpa).

References

Alexander, R. A., Carson, K. P., Alliger, G. M., & Carr, L. (1987). Correcting doubly truncated correlations: An improved approximation for correcting the bivariate normal correlation when truncation has occurred on both variables. Educational and Psychological Measurement, 47(2), 309–315. https://doi.org/10.1177/0013164487472002

Dahlke, J. A., & Wiernik, B. M. (2019). Not restricted to selection research: Accounting for indirect range restriction in organizational research. Organizational Research Methods. Advance online publication. https://doi.org/10.1177/1094428119859398

Hunter, J. E., Schmidt, F. L., & Le, H. (2006). Implications of direct and indirect range restriction for meta-analysis methods and findings. Journal of Applied Psychology, 91(3), 594–612. https://doi.org/10.1037/0021-9010.91.3.594

Le, H., Oh, I.-S., Schmidt, F. L., & Wooldridge, C. D. (2016). Correction for range restriction in meta-analysis revisited: Improvements and implications for organizational research. Personnel Psychology, 69(4), 975–1008. https://doi.org/10.1111/peps.12122

Schmidt, F. L., & Hunter, J. E. (2015). Methods of meta-analysis: Correcting error and bias in research findings (3rd ed.). SAGE Publications. https://doi.org/10.4135/9781483398105. pp. 43–44, 140–141.

Examples

## Correction for measurement error only
correct_d(correction = "meas", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error:
#> ---------------------------------------------------------------------------------------
#>     dgy   dgp   dGy   dGp
#> 1 0.458 0.516 0.512 0.578

correct_d(correction = "meas", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1 0.595      0.3    0.912 300         183

## Correction for direct range restriction in the continuous variable
correct_d(correction = "uvdrr_y", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>    dgyi  dgpi  dGyi  dGpi  dgya  dgpa  dGya  dGpa
#> 1 0.458 0.512 0.512 0.578 0.655 0.684 0.732 0.776

correct_d(correction = "uvdrr_y", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1 0.799    0.403     1.22 300         106

## Correction for direct range restriction in the grouping variable
correct_d(correction = "uvdrr_g", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>    dgyi  dgpi  dGyi  dGpi dgya  dgpa  dGya  dGpa
#> 1 0.458 0.516 0.516 0.578  0.5 0.563 0.547 0.619

correct_d(correction = "uvdrr_g", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1 0.623    0.314    0.954 300         168

## Correction for indirect range restriction in the continuous variable
correct_d(correction = "uvdrr_y", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>    dgyi  dgpi  dGyi  dGpi  dgya  dgpa  dGya  dGpa
#> 1 0.458 0.512 0.512 0.578 0.655 0.684 0.732 0.776

correct_d(correction = "uvdrr_y", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1 0.799    0.403     1.22 300         106

## Correction for indirect range restriction in the grouping variable
correct_d(correction = "uvirr_g", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error and Univariate Indirect Range Restriction:
#> ---------------------------------------------------------------------------------------
#>    dgyi  dgpi  dGyi  dGpi  dgya  dgpa  dGya  dGpa
#> 1 0.458 0.516 0.512 0.578 0.509 0.584 0.567 0.642

correct_d(correction = "uvirr_g", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error and Univariate Indirect Range Restriction:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1 0.639    0.322     0.98 300         159

## Correction for indirect range restriction in the continuous variable
correct_d(correction = "uvdrr_y", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>    dgyi  dgpi  dGyi  dGpi  dgya  dgpa  dGya  dGpa
#> 1 0.458 0.512 0.512 0.578 0.655 0.684 0.732 0.776

correct_d(correction = "uvdrr_y", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error and Univariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1 0.799    0.403     1.22 300         106

## Correction for direct range restriction in both variables
correct_d(correction = "bvdrr", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error and Bivariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>    dgyi  dgpi  dGyi  dGpi  dgya  dgpa  dGya  dGpa
#> 1 0.458 0.516 0.512 0.578 0.692 0.734 0.764 0.811

correct_d(correction = "bvdrr", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error and Bivariate Direct Range Restriction:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1 0.817    0.419     1.23 300        90.9

## Correction for indirect range restriction in both variables
correct_d(correction = "bvirr", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = .7, pa = .5)

#> d Values Corrected for Measurement Error and Bivariate Indirect Range Restriction:
#> ---------------------------------------------------------------------------------------
#>    dgyi  dgpi  dGyi  dGpi dgya dgpa dGya dGpa
#> 1 0.458 0.516 0.512 0.578 0.95 1.01 1.06 1.13

correct_d(correction = "bvirr", d = .5, ryy = .8, uy = .7,
          rGg = .9, pi = NULL, pa = .5, n1 = 100, n2 = 200)

#> d Values Corrected for Measurement Error and Bivariate Indirect Range Restriction:
#> ---------------------------------------------------------------------------------------
#>   value CI_LL_95 CI_UL_95   n n_effective
#> 1  1.01    0.786     1.25 300        74.5