Correct for small-sample bias in Cohen's d values

Corrects a vector of Cohen's d values for small-sample bias, as Cohen's d has a slight positive bias. The bias-corrected d value is often called Hedges's g.

correct_d_bias(d, n)

Arguments

d
Vector of Cohen's d values.

n
Vector of sample sizes.

Value

Vector of g values (d values corrected for small-sample bias).

Details

The bias correction is: $$g = d_{c} = d_{obs} \times J$$

where $$J = \frac{\Gamma(\frac{n - 2}{2})}{\sqrt{\frac{n - 2}{2}} \times \Gamma(\frac{n - 3}{2})}$$

and \(d_{obs}\) is the observed effect size, \(g = d_{c}\) is the corrected (unbiased) estimate, \(n\) is the total sample size, and \(\Gamma()\) is the gamma function.

Historically, using the gamma function was computationally intensive, so an approximation for \(J\) was used (Borenstein et al., 2009): $$J = 1 - 3 / (4 * (n - 2) - 1)$$

This approximation is no longer necessary with modern computers.

References

Hedges, L. V., & Olkin, I. (1985). Statistical methods for meta-analysis. Academic Press. p. 104

Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Introduction to meta-analysis. Wiley. p. 27.

Examples

correct_d_bias(d = .3, n = 30)

#> [1] 0.2924263

correct_d_bias(d = .3, n = 300)

#> [1] 0.2992493

correct_d_bias(d = .3, n = 3000)

#> [1] 0.299925