Correct correlations for scale coarseness

Correct correlations for scale coarseness

correct_r_coarseness(
  r,
  kx = NULL,
  ky = NULL,
  n = NULL,
  dist_x = "norm",
  dist_y = "norm",
  bin_value_x = c("median", "mean", "index"),
  bin_value_y = c("median", "mean", "index"),
  width_x = 3,
  width_y = 3,
  lbound_x = NULL,
  ubound_x = NULL,
  lbound_y = NULL,
  ubound_y = NULL,
  index_values_x = NULL,
  index_values_y = NULL
)

Arguments

r
Observed correlation.

kx, ky
Number of scale points used to measure the x and y variables. Set to NULL to treat as continuously measured.

n
Optional sample size.

dist_x, dist_y
Assumed latent distribution of the x and y variables.

bin_value_x, bin_value_y
Are the scale points used to measure the of the x and y variables assumed to represent bin medians, means, or index values?

width_x, width_y
For symmetrically distributed variables, how many standard deviations above/below the latent mean should be be used for the latent variable range to make the correction? (Note: Setting width > 3 produces erratic results.) The latent variable range can alternatively be set using lbound and ubound.

lbound_x, lbound_y
What lower bound of the range for the latent x and y variables should be used to make the correction? (Note: For normally distributed variables, setting lbound < -3 produces erratic results.)

ubound_x, ubound_y
What upper bound of the range for the latent x and y variables should be used to make the correction? (Note: For normally distributed variables, setting ubound > 3 produces erratic results.)

index_values_x, index_values_y
Optional. If bin_value = "index", the bin index values. If unspecified, values 1:k are used.

Value

Vector of correlations corrected for scale coarseness (if n is supplied, corrected error variance and adjusted sample size is also reported).

References

Aguinis, H., Pierce, C. A., & Culpepper, S. A. (2009). Scale coarseness as a methodological artifact: Correcting correlation coefficients attenuated from using coarse scales. Organizational Research Methods, 12(4), 623–652. https://doi.org/10.1177/1094428108318065

Schmidt, F. L., & Hunter, J. E. (2015). Methods of meta-analysis: Correcting error and bias in research findings (3rd ed.). Thousand Oaks, CA: SAGE. https://doi.org/10/b6mg. pp. 287-288.

Peters, C. C., & Van Voorhis, W. R. (1940). Statistical procedures and their mathematical bases. New York, NY: Mcgraw-Hill. https://doi.org/10.1037/13596-000. pp. 393–399.

Examples

correct_r_coarseness(r = .35, kx = 5, ky = 4, n = 100)

#>   r_corrected var_e_corrected n_adj
#> 1        0.35     0.007777841   100

correct_r_coarseness(r = .35, kx = 5, n = 100)

#>   r_corrected var_e_corrected n_adj
#> 1        0.35     0.007777841   100

correct_r_coarseness(r = .35, kx = 5, ky = 4, n = 100, dist_x="unif", dist_y="norm")

#>   r_corrected var_e_corrected    n_adj
#> 1   0.3572173     0.008101918 94.93757