Generate an artifact distribution object for a dichotomous grouping variable.

This function generates artifact-distribution objects containing either interactive or Taylor series artifact distributions for dichotomous group-membership variables. Use this to create objects that can be supplied to the ma_r_ad and ma_d_ad functions to apply psychometric corrections to barebones meta-analysis objects via artifact distribution methods.

Allows consolidation of observed and estimated artifact information by cross-correcting artifact distributions and forming weighted artifact summaries.

create_ad_group(
  ad_type = c("tsa", "int"),
  rGg = NULL,
  n_rGg = NULL,
  wt_rGg = n_rGg,
  pi = NULL,
  pa = NULL,
  n_pi = NULL,
  n_pa = NULL,
  wt_p = n_pi,
  mean_rGg = NULL,
  var_rGg = NULL,
  k_rGg = NULL,
  mean_n_rGg = NULL,
  var_unbiased = TRUE,
  ...
)

Arguments

ad_type
Type of artifact distribution to be computed: Either "tsa" for Taylor series approximation or "int" for interactive.

rGg
Vector of incumbent reliability estimates.

n_rGg
Vector of sample sizes associated with the elements of rGg.

wt_rGg
Vector of weights associated with the elements of rGg (by default, sample sizes will be used as weights if provided).

pi
Vector of incumbent/sample proportions of members in one of the two groups being compared (one or both of pi/pa can be vectors - if both are vectors, they must be of equal length).

pa
Vector of applicant/population proportions of members in one of the two groups being compared (one or both of pi/pa can be vectors - if both are vectors, they must be of equal length).

n_pi
Vector of sample sizes associated with the elements in pi.

n_pa
Vector of sample sizes associated with the elements in pa.

wt_p
Vector of weights associated with the collective element pairs in pi and pa.

mean_rGg
Vector that can be used to supply the means of externally computed distributions of correlations between observed and latent group membership.

var_rGg
Vector that can be used to supply the variances of externally computed distributions of correlations between observed and latent group membership.

k_rGg
Vector that can be used to supply the number of studies included in externally computed distributions of correlations between observed and latent group membership.

mean_n_rGg
Vector that can be used to supply the mean sample sizes of externally computed distributions of correlations between observed and latent group membership.

var_unbiased
Logical scalar determining whether variance should be unbiased (TRUE) or maximum-likelihood (FALSE).

...
Further arguments.

Value

Artifact distribution object (matrix of artifact-distribution means and variances) for use in artifact-distribution meta-analyses.

Examples

## Example artifact distribution for a dichotomous grouping variable:
create_ad_group(rGg = c(.8, .9, .95), n_rGg = c(100, 200, 250),
                mean_rGg = .9, var_rGg = .05,
                k_rGg = 5, mean_n_rGg = 100,
                pi = c(.6, .55, .3), pa = .5, n_pi = c(100, 200, 250), n_pa = c(300, 300, 300),
                var_unbiased = TRUE)

#> Taylor-Series Artifact Distributions
#> ------------------------------------
#>           mean   var var_res
#> qxa_irr  0.933 0.021   0.020
#> qxa_drr  0.932 0.022   0.021
#> qxi_irr  0.927 0.025   0.024
#> qxi_drr  0.927 0.025   0.024
#> rxxa_irr 0.871 0.068   0.066
#> rxxa_drr 0.870 0.072   0.069
#> rxxi_irr 0.860 0.082   0.079
#> rxxi_drr 0.860 0.082   0.079
#> ux       0.957 0.002   0.000
#> ut       0.950 0.003   0.000
#> 

                
create_ad_group(ad_type = "int", rGg = c(.8, .9, .95), n_rGg = c(100, 200, 250),
                mean_rGg = .9, var_rGg = .05,
                k_rGg = 5, mean_n_rGg = 100,
                pi = c(.6, .55, .3), pa = .5, n_pi = c(100, 200, 250), n_pa = c(300, 300, 300),
                var_unbiased = TRUE)

#> Interactive Artifact Distributions
#> ----------------------------------
#> 
#> qxa Artifact Distribution - Indirect Range Restriction
#>   Value Weight
#> 1 0.819  0.182
#> 2 0.909  0.364
#> 3 0.954  0.455
#> 
#> qxa Artifact Distribution - Direct Range Restriction
#>   Value Weight
#> 1 0.813  0.182
#> 2 0.907  0.364
#> 3 0.954  0.455
#> 
#> qxi Artifact Distribution - Indirect Range Restriction
#>   Value Weight
#> 1  0.80  0.182
#> 2  0.90  0.364
#> 3  0.95  0.455
#> 
#> qxi Artifact Distribution - Direct Range Restriction
#>   Value Weight
#> 1  0.80  0.182
#> 2  0.90  0.364
#> 3  0.95  0.455
#> 
#> ux Artifact Distribution
#>   Value Weight
#> 1 0.917  0.455
#> 2 0.980  0.182
#> 3 0.995  0.364
#> 
#> ut Artifact Distribution
#>   Value Weight
#> 1 0.901  0.455
#> 2 0.976  0.182
#> 3 0.994  0.364
#>