Create a tibble of artifact distributions by construct

Create a tibble of artifact distributions by construct

create_ad_tibble(
  ad_type = c("tsa", "int"),
  n = NULL,
  sample_id = NULL,
  construct_x = NULL,
  facet_x = NULL,
  measure_x = NULL,
  construct_y = NULL,
  facet_y = NULL,
  measure_y = NULL,
  rxx = NULL,
  rxx_restricted = TRUE,
  rxx_type = "alpha",
  k_items_x = NA,
  ryy = NULL,
  ryy_restricted = TRUE,
  ryy_type = "alpha",
  k_items_y = NA,
  ux = NULL,
  ux_observed = TRUE,
  uy = NULL,
  uy_observed = TRUE,
  estimate_rxxa = TRUE,
  estimate_rxxi = TRUE,
  estimate_ux = TRUE,
  estimate_ut = TRUE,
  moderators = NULL,
  cat_moderators = TRUE,
  moderator_type = c("simple", "hierarchical", "none"),
  construct_order = NULL,
  supplemental_ads = NULL,
  data = NULL,
  control = control_psychmeta(),
  ...
)

create_ad_list(
  ad_type = c("tsa", "int"),
  n = NULL,
  sample_id = NULL,
  construct_x = NULL,
  facet_x = NULL,
  measure_x = NULL,
  construct_y = NULL,
  facet_y = NULL,
  measure_y = NULL,
  rxx = NULL,
  rxx_restricted = TRUE,
  rxx_type = "alpha",
  k_items_x = NA,
  ryy = NULL,
  ryy_restricted = TRUE,
  ryy_type = "alpha",
  k_items_y = NA,
  ux = NULL,
  ux_observed = TRUE,
  uy = NULL,
  uy_observed = TRUE,
  estimate_rxxa = TRUE,
  estimate_rxxi = TRUE,
  estimate_ux = TRUE,
  estimate_ut = TRUE,
  moderators = NULL,
  cat_moderators = TRUE,
  moderator_type = c("simple", "hierarchical", "none"),
  construct_order = NULL,
  supplemental_ads = NULL,
  data = NULL,
  control = control_psychmeta(),
  ...
)

Arguments

ad_type
Type of artifact distributions to be computed: Either "tsa" for Taylor series approximation or "int" for interactive.

n
Vector or column name of sample sizes.

sample_id
Optional vector of identification labels for samples/studies in the meta-analysis.

construct_x, construct_y
Vector of construct names for constructs initially designated as "X" or "Y".

facet_x, facet_y
Vector of facet names for constructs initially designated as "X" or "Y". Facet names "global", "overall", and "total" are reserved to indicate observations that represent effect sizes that have already been composited or that represent construct-level measurements rather than facet-level measurements. To avoid double-compositing, any observation with one of these reserved names will only be eligible for auto-compositing with other such observations and will not be combined with narrow facets.

measure_x, measure_y
Vector of names for measures associated with constructs initially designated as "X" or "Y".

rxx
Vector or column name of reliability estimates for X.

rxx_restricted
Logical vector or column name determining whether each element of rxx is an incumbent reliability (TRUE) or an applicant reliability (FALSE).

rxx_type, ryy_type
String vector identifying the types of reliability estimates supplied. See documentation of ma_r for a full list of acceptable values.

k_items_x, k_items_y
Numeric vector identifying the number of items in each scale.

ryy
Vector or column name of reliability estimates for Y.

ryy_restricted
Logical vector or column name determining whether each element of ryy is an incumbent reliability (TRUE) or an applicant reliability (FALSE).

ux
Vector or column name of u ratios for X.

ux_observed
Logical vector or column name determining whether each element of ux is an observed-score u ratio (TRUE) or a true-score u ratio (FALSE).

uy
Vector or column name of u ratios for Y.

uy_observed
Logical vector or column name determining whether each element of uy is an observed-score u ratio (TRUE) or a true-score u ratio (FALSE).

estimate_rxxa
Logical argument to estimate rxxa values from other artifacts (TRUE) or to only used supplied rxxa values (FALSE). TRUE by default.

estimate_rxxi
Logical argument to estimate rxxi values from other artifacts (TRUE) or to only used supplied rxxi values (FALSE). TRUE by default.

estimate_ux
Logical argument to estimate ux values from other artifacts (TRUE) or to only used supplied ux values (FALSE). TRUE by default.

estimate_ut
Logical argument to estimate ut values from other artifacts (TRUE) or to only used supplied ut values (FALSE). TRUE by default.

moderators
Matrix or column names of moderator variables to be used in the meta-analysis (can be a vector in the case of one moderator).

cat_moderators
Logical scalar or vector identifying whether variables in the moderators argument are categorical variables (TRUE) or continuous variables (FALSE).

moderator_type
Type of moderator analysis: "none" means that no moderators are to be used, "simple" means that moderators are to be examined one at a time, and "hierarchical" means that all possible combinations and subsets of moderators are to be examined.

construct_order
Vector indicating the order in which variables should be arranged, with variables listed earlier in the vector being preferred for designation as X.

supplemental_ads
Named list (named according to the constructs included in the meta-analysis) of supplemental artifact distribution information from studies not included in the meta-analysis. This is a list of lists, where the elements of a list associated with a construct are named like the arguments of the create_ad() function.

data
Data frame containing columns whose names may be provided as arguments to vector arguments.

control
Output from the control_psychmeta() function or a list of arguments controlled by the control_psychmeta() function. Ellipsis arguments will be screened for internal inclusion in control.

...
Additional arguments

Value

A tibble of artifact distributions

Examples

## Examples to create Taylor series artifact distributions:
# Overall artifact distributions (not pairwise, not moderated)
create_ad_tibble(ad_type = "tsa",
                 n = n, rxx = rxxi, ryy = ryyi,
                 construct_x = x_name, construct_y = y_name,
                 sample_id = sample_id, moderators = moderator,
                 data = data_r_meas_multi,
                 control = control_psychmeta(pairwise_ads = FALSE,
                                             moderated_ads = FALSE))

#> Tibble of artifact distributions 
#> -------------------------------- 
#> # A tibble: 3 x 3
#>   construct_x analysis_type ad_x             
#> * <chr>       <chr>         <list>           
#> 1 X           Overall       <ad_tsa [10 × 3]>
#> 2 Y           Overall       <ad_tsa [10 × 3]>
#> 3 Z           Overall       <ad_tsa [10 × 3]>

# Overall artifact distributions by moderator combination
create_ad_tibble(ad_type = "tsa",
                 n = n, rxx = rxxi, ryy = ryyi,
                 construct_x = x_name, construct_y = y_name,
                 sample_id = sample_id, moderators = moderator,
                 data = data_r_meas_multi,
                 control = control_psychmeta(pairwise_ads = FALSE,
                                             moderated_ads = TRUE))

#> Tibble of artifact distributions 
#> -------------------------------- 
#> # A tibble: 9 x 4
#>   construct_x analysis_type    moderator  ad_x             
#> * <chr>       <fct>            <fct>      <list>           
#> 1 X           Overall          All Levels <ad_tsa [10 × 3]>
#> 2 X           Simple Moderator 1          <ad_tsa [10 × 3]>
#> 3 X           Simple Moderator 2          <ad_tsa [10 × 3]>
#> 4 Y           Overall          All Levels <ad_tsa [10 × 3]>
#> 5 Y           Simple Moderator 1          <ad_tsa [10 × 3]>
#> 6 Y           Simple Moderator 2          <ad_tsa [10 × 3]>
#> 7 Z           Overall          All Levels <ad_tsa [10 × 3]>
#> 8 Z           Simple Moderator 1          <ad_tsa [10 × 3]>
#> 9 Z           Simple Moderator 2          <ad_tsa [10 × 3]>

# Pairwise artifact distributions (not moderated)
create_ad_tibble(ad_type = "tsa",
                 n = n, rxx = rxxi, ryy = ryyi,
                 construct_x = x_name, construct_y = y_name,
                 sample_id = sample_id, moderators = moderator,
                 data = data_r_meas_multi,
                 control = control_psychmeta(pairwise_ads = TRUE,
                                               moderated_ads = FALSE))

#> Tibble of artifact distributions 
#> -------------------------------- 
#> # A tibble: 3 x 5
#>   construct_x construct_y analysis_type ad_x              ad_y             
#> * <chr>       <chr>       <chr>         <list>            <list>           
#> 1 X           Y           Overall       <ad_tsa [10 × 3]> <ad_tsa [10 × 3]>
#> 2 X           Z           Overall       <ad_tsa [10 × 3]> <ad_tsa [10 × 3]>
#> 3 Y           Z           Overall       <ad_tsa [10 × 3]> <ad_tsa [10 × 3]>

# Pairwise artifact distributions by moderator combination
create_ad_tibble(ad_type = "tsa",
                 n = n, rxx = rxxi, ryy = ryyi,
                 construct_x = x_name, construct_y = y_name,
                 sample_id = sample_id, moderators = moderator,
                 data = data_r_meas_multi,
                 control = control_psychmeta(pairwise_ads = TRUE,
                                             moderated_ads = TRUE))

#> Tibble of artifact distributions 
#> -------------------------------- 
#> # A tibble: 9 x 6
#>   construct_x construct_y analysis_type    moderator ad_x          ad_y         
#> * <chr>       <chr>       <fct>            <fct>     <list>        <list>       
#> 1 X           Y           Overall          All Leve… <ad_tsa [10 … <ad_tsa [10 …
#> 2 X           Y           Simple Moderator 1         <ad_tsa [10 … <ad_tsa [10 …
#> 3 X           Y           Simple Moderator 2         <ad_tsa [10 … <ad_tsa [10 …
#> 4 X           Z           Overall          All Leve… <ad_tsa [10 … <ad_tsa [10 …
#> 5 X           Z           Simple Moderator 1         <ad_tsa [10 … <ad_tsa [10 …
#> 6 X           Z           Simple Moderator 2         <ad_tsa [10 … <ad_tsa [10 …
#> 7 Y           Z           Overall          All Leve… <ad_tsa [10 … <ad_tsa [10 …
#> 8 Y           Z           Simple Moderator 1         <ad_tsa [10 … <ad_tsa [10 …
#> 9 Y           Z           Simple Moderator 2         <ad_tsa [10 … <ad_tsa [10 …