Construct a credibility interval

Function to construct a credibility interval around a mean effect size.

credibility(mean, sd, k = NULL, cred_level = 0.8, cred_method = c("t", "norm"))

Arguments

mean
Mean effect size.

sd
Residual/true standard deviation of effect sizes, after accounting for variance from artifacts.

k
Number of studies in the meta-analysis.

cred_level
Credibility level that defines the width of the credibility interval (default = .80).

cred_method
Distribution to be used to compute the width of credibility intervals. Available options are "t" for t distribution or "norm" for normal distribution.

Value

A matrix of credibility intervals of the specified width.

Details

$$CR=mean_{es}\pm quantile\times SD_{es}$$

Examples

credibility(mean = .3, sd = .15, cred_level = .8, cred_method = "norm")

#>       CR_LL_80  CR_UL_80
#> [1,] 0.1077673 0.4922327

credibility(mean = .3, sd = .15, cred_level = .8, k = 10)

#> Warning: Only one value can be specified for cred_method: First value used

#>        CR_LL_80  CR_UL_80
#> [1,] 0.09254569 0.5074543

credibility(mean = c(.3, .5), sd = c(.15, .2), cred_level = .8, k = 10)

#> Warning: Only one value can be specified for cred_method: First value used

#>        CR_LL_80  CR_UL_80
#> [1,] 0.09254569 0.5074543
#> [2,] 0.22339425 0.7766057