Function to convert a meta-analysis of correlations to a meta-analysis of d values or vice-versa (does one table)

Function to convert a meta-analysis of correlations to a meta-analysis of d values or vice-versa (does one table)

.convert_metatab(
  ma_table,
  p_vec = rep(0.5, nrow(ma_table)),
  conf_level = 0.95,
  cred_level = 0.8,
  conf_method = "t",
  cred_method = "t"
)

Arguments

ma_table
Meta-analysis table.

p_vec
Vector of proportions associated with the rows of ma_table.

conf_level
Confidence level to define the width of the confidence interval (default = .95).

cred_level
Credibility level to define the width of the credibility interval (default = .80).

conf_method
Distribution to be used to compute the width of confidence intervals. Available options are "t" for t distribution or "norm" for normal distribution.

cred_method
Distribution to be used to compute the width of credibility intervals. Available options are "t" for t distribution or "norm" for normal distribution.

Value

Meta-analysis table converted to a new metric