Computation of heterogeneity indices from meta-analytic results

Computation of heterogeneity indices from meta-analytic results

.heterogeneity(
  mean_es,
  var_es,
  var_pre,
  var_res,
  var_e = NA,
  var_art = NA,
  wt_vec,
  N,
  k,
  es_vec,
  vare_vec,
  es_failsafe = NULL,
  conf_level = 0.95,
  es_type = "es",
  wt_type,
  ma_method,
  var_unbiased,
  var_res_ci_method
)

Arguments

mean_es
The mean effect size.

var_es
The observed variances of effect sizes.

var_pre
The total predicted variance of effect sizes due to sampling error and other artifacts.

var_res
The estimated residual variance of effect sizes.

var_e
The predicted variance of effect sizes due to sampling error.

var_art
The predicted variance of effect sizes predicted from other artifacts.

wt_vec
The vector of weights used in the meta-analysis.

N
The total sample size of the meta-analysis.

k
The number of effect sizes included in the meta-analysis.

es_vec
The vector of effect sizes used in the meta-analysis.

vare_vec
The vector of sampling-error variances used in the meta-analysis.

es_failsafe
Failsafe value of the effect size for use in file-drawer analyses.

conf_level
Confidence level to define the width of the confidence interval (default = .95).

es_type
Name of effect-size type.

wt_type
Weighting method.

ma_method
What artifact correction method is used. Options are "bb", "ic", and "ad".

var_unbiased
Are variances calculated using the unbiased (`TRUE`) or maximum likelihood (`FALSE`) estimator?

var_res_ci_method
Method to use to estimate a confidence interval for `var_res`. See heterogeneity for details.

Value

A list of heterogeneity statistics.