Estimation of applicant and incumbent reliabilities and of true- and observed-score u ratios

Functions to estimate the values of artifacts from other artifacts. These functions allow for reliability estimates to be corrected/attenuated for range restriction and allow u ratios to be converted between observed-score and true-score metrics. Some functions also allow for the extrapolation of an artifact from other available information.

Available functions include:

estimate_rxxa
Estimate the applicant reliability of variable X from X's incumbent reliability value and X's observed-score or true-score u ratio.

estimate_rxxa_u
Estimate the applicant reliability of variable X from X's observed-score and true-score u ratios.

estimate_rxxi
Estimate the incumbent reliability of variable X from X's applicant reliability value and X's observed-score or true-score u ratio.

estimate_rxxi_u
Estimate the incumbent reliability of variable X from X's observed-score and true-score u ratios.

estimate_ux
Estimate the true-score u ratio for variable X from X's reliability coefficient and X's observed-score u ratio.

estimate_uy
Estimate the observed-score u ratio for variable X from X's reliability coefficient and X's true-score u ratio.

estimate_ryya
Estimate the applicant reliability of variable Y from Y's incumbent reliability value, Y's correlation with X, and X's u ratio.

estimate_ryyi
Estimate the incumbent reliability of variable Y from Y's applicant reliability value, Y's correlation with X, and X's u ratio.

estimate_uy
Estimate the observed-score u ratio for variable Y from Y's applicant and incumbent reliability coefficients.

estimate_up
Estimate the true-score u ratio for variable Y from Y's applicant and incumbent reliability coefficients.

estimate_rxxa(
  rxxi,
  ux,
  ux_observed = TRUE,
  indirect_rr = TRUE,
  rxxi_type = "alpha"
)

estimate_rxxi(
  rxxa,
  ux,
  ux_observed = TRUE,
  indirect_rr = TRUE,
  rxxa_type = "alpha"
)

estimate_ut(ux, rxx, rxx_restricted = TRUE)

estimate_ux(ut, rxx, rxx_restricted = TRUE)

estimate_ryya(ryyi, rxyi, ux)

estimate_ryyi(ryya, rxyi, ux)

estimate_uy(ryyi, ryya, indirect_rr = TRUE, ryy_type = "alpha")

estimate_up(ryyi, ryya)

estimate_rxxa_u(ux, ut)

estimate_rxxi_u(ux, ut)

Arguments

rxxi
Vector of incumbent reliability estimates for X.

ux
Vector of observed-score u ratios for X (if used in the context of estimating a reliability value, a true-score u ratio may be supplied by setting ux_observed to FALSE).

ux_observed
Logical vector determining whether each element of ux is an observed-score u ratio (TRUE) or a true-score u ratio (FALSE).

indirect_rr
Logical vector determining whether each reliability value is associated with indirect range restriction (TRUE) or direct range restriction (FALSE).

rxxi_type, rxxa_type, ryy_type
String vector identifying the types of reliability estimates supplied (e.g., "alpha", "retest", "interrater_r", "splithalf"). See the documentation for ma_r for a full list of acceptable reliability types.

rxxa
Vector of applicant reliability estimates for X.

rxx
Vector of reliability estimates for X (used in the context of estimating ux and ut - specify that reliability is an incumbent value by setting rxx_restricted to FALSE).

rxx_restricted
Logical vector determining whether each element of rxx is an incumbent reliability (TRUE) or an applicant reliability (FALSE).

ut
Vector of true-score u ratios for X.

ryyi
Vector of incumbent reliability estimates for Y.

rxyi
Vector of observed-score incumbent correlations between X and Y.

ryya
Vector of applicant reliability estimates for Y.

Value

A vector of estimated artifact values.

Details

#### Formulas to estimate rxxa ####

Formulas for indirect range restriction: $$\rho_{XX_{a}}=1-u_{X}^{2}\left(1-\rho_{XX_{i}}\right)$$ $$\rho_{XX_{a}}=\frac{\rho_{XX_{i}}}{\rho_{XX_{i}}+u_{T}^{2}-\rho_{XX_{i}}u_{T}^{2}}$$

Formula for direct range restriction: $$\rho_{XX_{a}}=\frac{\rho_{XX_{i}}}{u_{X}^{2}\left[1+\rho_{XX_{i}}\left(\frac{1}{u_{X}^{2}}-1\right)\right]}$$

#### Formulas to estimate rxxi ####

Formulas for indirect range restriction: $$\rho_{XX_{i}}=1-\frac{1-\rho_{XX_{a}}}{u_{X}^{2}}$$ $$\rho_{XX_{i}}=1-\frac{1-\rho_{XX_{a}}}{\rho_{XX_{a}}\left[u_{T}^{2}-\left(1-\frac{1}{\rho_{XX_{a}}}\right)\right]}$$

Formula for direct range restriction: $$\rho_{XX_{i}}=\frac{\rho_{XX_{i}}u_{X}^{2}}{1+\rho_{XX_{i}}\left(u_{X}^{2}-1\right)}$$

#### Formulas to estimate ut ####

$$u_{T}=\sqrt{\frac{\rho_{XX_{i}}u_{X}^{2}}{1+\rho_{XX_{i}}u_{X}^{2}-u_{X}^{2}}}$$ $$u_{T}=\sqrt{\frac{u_{X}^{2}-\left(1-\rho_{XX_{a}}\right)}{\rho_{XX_{a}}}}$$

#### Formulas to estimate ux #### $$u_{X}=\sqrt{\frac{u_{T}^{2}}{\rho_{XX_{i}}\left(1+\frac{u_{T}^{2}}{\rho_{XX_{i}}}-u_{T}^{2}\right)}}$$ $$u_{X}=\sqrt{\rho_{XX_{a}}\left[u_{T}^{2}-\left(1-\frac{1}{\rho_{XX_{a}}}\right)\right]}$$

#### Formula to estimate ryya ####

$$\rho_{YY_{a}}=1-\frac{1-\rho_{YY_{i}}}{1-\rho_{XY_{i}}^{2}\left(1-\frac{1}{u_{X}^{2}}\right)}$$

#### Formula to estimate ryyi $$\rho_{YY_{i}}=1-\left(1-\rho_{YY_{a}}\right)\left[1-\rho_{XY_{i}}^{2}\left(1-\frac{1}{u_{X}^{2}}\right)\right]$$

#### Formula to estimate uy #### $$u_{Y}=\sqrt{\frac{1-\rho_{YY_{a}}}{1-\rho_{YY_{i}}}}$$

#### Formula to estimate up #### $$u_{P}=\sqrt{\frac{\frac{1-\rho_{YY_{a}}}{1-\rho_{YY_{i}}}-\left(1-\rho_{YY_{a}}\right)}{\rho_{YY_{a}}}}$$

References

Schmidt, F. L., & Hunter, J. E. (2015). Methods of meta-analysis: Correcting error and bias in research findings (3rd ed.). Thousand Oaks, CA: Sage. https://doi.org/10/b6mg p. 127.

Le, H., & Schmidt, F. L. (2006). Correcting for indirect range restriction in meta-analysis: Testing a new meta-analytic procedure. Psychological Methods, 11(4), 416–438. https://doi.org/10.1037/1082-989X.11.4.416

Hunter, J. E., Schmidt, F. L., & Le, H. (2006). Implications of direct and indirect range restriction for meta-analysis methods and findings. Journal of Applied Psychology, 91(3), 594–612. https://doi.org/10.1037/0021-9010.91.3.594

Le, H., Oh, I.-S., Schmidt, F. L., & Wooldridge, C. D. (2016). Correction for range restriction in meta-analysis revisited: Improvements and implications for organizational research. Personnel Psychology, 69(4), 975–1008. https://doi.org/10.1111/peps.12122

Examples

estimate_rxxa(rxxi = .8, ux = .8, ux_observed = TRUE)

#> [1] 0.872

estimate_rxxi(rxxa = .8, ux = .8, ux_observed = TRUE)

#> [1] 0.6875

estimate_ut(ux = .8, rxx = .8, rxx_restricted = TRUE)

#> [1] 0.766261

estimate_ux(ut = .8, rxx = .8, rxx_restricted = TRUE)

#> [1] 0.8304548

estimate_ryya(ryyi = .8, rxyi = .3, ux = .8)

#> [1] 0.8096371

estimate_ryyi(ryya = .8, rxyi = .3, ux = .8)

#> [1] 0.789875

estimate_uy(ryyi = c(.5, .7), ryya = c(.7, .8))

#> [1] 0.7745967 0.8164966

estimate_up(ryyi = c(.5, .7), ryya = c(.7, .8))

#> [1] 0.6546537 0.7637626

estimate_rxxa_u(ux = c(.7, .8), ut = c(.65, .75))

#> [1] 0.8831169 0.8228571

estimate_rxxi_u(ux = c(.7, .8), ut = c(.65, .75))

#> [1] 0.7614630 0.7232143