Estimate covariance matrices and mean vectors containing product terms

Estimate covariance matrices and mean vectors containing product terms

estimate_matrix_prods(sigma_mat, mu_vec, prod_list)

Arguments

sigma_mat
Covariance parameter matrix.

mu_vec
Mean parameter matrix.

prod_list
List of 2-element vectors containing the names of variables in sigma_mat.

Value

Augmented covariance matrix and mean vector containing product terms.

Examples

## Establish mean and covariance parameters
mu_vec <- 1:4
sigma_mat <- reshape_vec2mat(c(.1, .2, .3, .4, .5, .6), var_names = LETTERS[1:4])
names(mu_vec) <- colnames(sigma_mat)

## Define a list of variables to be used in estimating products:
prod_list <- list(c("A", "B"),
                  c("A", "C"),
                  c("A", "D"),
                  c("B", "C"),
                  c("B", "D"),
                  c("C", "D"))

## Generate data for the purposes of comparison:
set.seed(1)
dat <- data.frame(MASS::mvrnorm(100000, mu = mu_vec, Sigma = sigma_mat, empirical = TRUE))

## Create product terms in simulated data:
for(i in 1:length(prod_list))
     dat[,paste(prod_list[[i]], collapse = "_x_")] <-
     dat[,prod_list[[i]][1]] * dat[,prod_list[[i]][2]]

## Analytically estimate product variables and compare to simulated data:
estimate_matrix_prods(sigma_mat = sigma_mat, mu_vec = mu_vec, prod_list = prod_list)

#> $sigma
#>         A   B   C   D A_x_B A_x_C A_x_D B_x_C B_x_D C_x_D
#> A     1.0 0.1 0.2 0.3  2.10  3.20  4.30  0.70  1.00  1.70
#> B     0.1 1.0 0.4 0.5  1.20  0.70  0.90  3.80  5.00  3.10
#> C     0.2 0.4 1.0 0.6  0.80  1.60  1.40  3.20  2.80  5.80
#> D     0.3 0.5 0.6 1.0  1.10  1.50  2.20  2.70  4.00  5.40
#> A_x_B 2.1 1.2 0.8 1.1  6.41  7.52 10.03  5.44  7.35  6.72
#> A_x_C 3.2 0.7 1.6 1.5  7.52 12.24 14.96  5.48  5.98 11.32
#> A_x_D 4.3 0.9 1.4 2.2 10.03 14.96 20.49  5.66  8.25 12.58
#> B_x_C 0.7 3.8 3.2 2.7  5.44  5.48  5.66 18.96 21.40 21.64
#> B_x_D 1.0 5.0 2.8 4.0  7.35  5.98  8.25 21.40 29.25 23.90
#> C_x_D 1.7 3.1 5.8 5.4  6.72 11.32 12.58 21.64 23.90 40.76
#> 
#> $mu
#>     A     B     C     D A_x_B A_x_C A_x_D B_x_C B_x_D C_x_D 
#>   1.0   2.0   3.0   4.0   2.1   3.2   4.3   6.4   8.5  12.6 
#> 

round(cov(dat), 2)

#>         A    B    C    D A_x_B A_x_C A_x_D B_x_C B_x_D C_x_D
#> A     1.0 0.10 0.20 0.30  2.10  3.20  4.30  0.70  1.00  1.70
#> B     0.1 1.00 0.40 0.50  1.20  0.70  0.90  3.79  4.99  3.09
#> C     0.2 0.40 1.00 0.60  0.80  1.60  1.40  3.19  2.79  5.80
#> D     0.3 0.50 0.60 1.00  1.10  1.50  2.20  2.69  4.00  5.40
#> A_x_B 2.1 1.20 0.80 1.10  6.40  7.51 10.04  5.42  7.34  6.73
#> A_x_C 3.2 0.70 1.60 1.50  7.51 12.22 14.95  5.49  5.99 11.35
#> A_x_D 4.3 0.90 1.40 2.20 10.04 14.95 20.51  5.67  8.28 12.61
#> B_x_C 0.7 3.79 3.19 2.69  5.42  5.49  5.67 18.89 21.33 21.56
#> B_x_D 1.0 4.99 2.79 4.00  7.34  5.99  8.28 21.33 29.21 23.85
#> C_x_D 1.7 3.09 5.80 5.40  6.73 11.35 12.61 21.56 23.85 40.71

round(apply(dat, 2, mean), 2)

#>     A     B     C     D A_x_B A_x_C A_x_D B_x_C B_x_D C_x_D 
#>   1.0   2.0   3.0   4.0   2.1   3.2   4.3   6.4   8.5  12.6