Estimation of statistics computed from products of random, normal variables

This family of functions computes univariate descriptive statistics for the products of two variables denoted as "x" and "y" (e.g., mean(x * y) or var(x * y)) and the covariance between the products of "x" and "y" and of "u" and "v" (e.g., cov(x * y, u * v) or cor(x * y, u * v)). These functions presume all variables are random normal variables.

Available functions include:

estimate_mean_prod
Estimate the mean of the product of two variables: x * y.

estimate_var_prod
Estimate the variance of the product of two variables: x * y.

estimate_cov_prods
Estimate the covariance between the products of two pairs of variables: x * y and u * v.

estimate_cor_prods
Estimate the correlation between the products of two pairs of variables: x * y and u * v.

estimate_mean_prod(mu_x, mu_y, cov_xy)

estimate_var_prod(mu_x, mu_y, var_x, var_y, cov_xy)

estimate_cov_prods(mu_x, mu_y, mu_u, mu_v, cov_xu, cov_xv, cov_yu, cov_yv)

estimate_cor_prods(
  mu_x,
  mu_y,
  mu_u,
  mu_v,
  var_x,
  var_y,
  var_u,
  var_v,
  cov_xu,
  cov_xv,
  cov_yu,
  cov_yv,
  cov_xy,
  cov_uv
)

Arguments

mu_x
Expected value of variable x.

mu_y
Expected value of variable y.

cov_xy
Covariance between x and y.

var_x
Variance of variable x.

var_y
Variance of variable y.

mu_u
Expected value of variable u.

mu_v
Expected value of variable v.

cov_xu
Covariance between x and u.

cov_xv
Covariance between x and v.

cov_yu
Covariance between y and u.

cov_yv
Covariance between y and v.

var_u
Variance of variable u.

var_v
Variance of variable v.

cov_uv
Covariance between u and v.

Value

An estimated statistic computed from the products of random, normal variables.

References

Bohrnstedt, G. W., & Goldberger, A. S. (1969). On the exact covariance of products of random variables. Journal of the American Statistical Association, 64(328), 1439. https://doi.org/10.2307/2286081

Goodman, L. A. (1960). On the exact variance of products. Journal of the American Statistical Association, 55(292), 708. https://doi.org/10.2307/2281592