Estimate u ratios from available artifact information

Uses information about standard deviations, reliability estimates, and selection ratios to estimate u ratios. Selection ratios are only used to estimate u when no other information is available, but estimates of u computed from SDs and reliabilities will be averaged to reduce error.

estimate_u(
  measure_id = NULL,
  sdi = NULL,
  sda = NULL,
  rxxi = NULL,
  rxxa = NULL,
  item_ki = NULL,
  item_ka = NULL,
  n = NULL,
  meani = NULL,
  sr = NULL,
  rxya_est = NULL,
  data = NULL
)

Arguments

measure_id
Vector of measure identifiers.

sdi
Scalar or vector containing restricted standard deviation(s).

sda
Scalar or vector containing unrestricted standard deviation(s).

rxxi
Scalar or vector containing restricted reliability coefficient(s).

rxxa
Scalar or vector containing unrestricted reliability coefficient(s).

item_ki
Scalar or vector containing the number of items used in measures within samples.

item_ka
Scalar or vector indicating the number of items toward which reliability estimates should be adjusted using the Spearman-Brown formula.

n
Vector of sample sizes.

meani
Vector of sample means.

sr
Vector of selection ratios (used only when no other useable u-ratio inputs are available).

rxya_est
Vector of estimated unrestricted correlations between the selection mechanism and the variable of interest (used only when sr is used).

data
Optional data frame containing any or all information for use in other arguments.

Value

A vector of estimated u ratios.

Examples

sdi <- c(1.4, 1.2, 1.3, 1.4)
sda <- 2
rxxi <- c(.6, .7, .75, .8)
rxxa <- c(.9, .95, .8, .9)
item_ki <- c(12, 12, 12, NA)
item_ka <- NULL
n <- c(200, 200, 200, 200)
meani <- c(2, 1, 2, 3)
sr <- c(.5, .6, .7, .4)
rxya_est <- .5

## Estimate u from standard deviations only:
estimate_u(sdi = sdi, sda = sda)

#> [1] 0.70 0.60 0.65 0.70

## Estimate u from incumbent standard deviations and the
## mixture standard deviation:
estimate_u(sdi = sdi, sda = "mixture", meani = meani, n = n)

#> [1] 0.9319864 0.7988455 0.8654160 0.9319864

## Estimate u from reliability information:
estimate_u(rxxi = rxxi, rxxa = rxxa)

#> [1] 0.5000000 0.4082483 0.8944272 0.7071068

## Estimate u from both standard deviations and reliabilities:
estimate_u(sdi = sdi, sda = sda, rxxi = rxxi, rxxa = rxxa,
           item_ki = item_ki, item_ka = item_ka, n = n,
           meani = meani, sr = sr, rxya_est = rxya_est)

#> [1] 0.70 0.60 0.65 0.70

estimate_u(sdi = sdi, sda = "average", rxxi = rxxi, rxxa = "average",
           item_ki = item_ki, item_ka = item_ka, n = n, meani = meani)

#> [1] 1.0565263 0.9055940 0.9810602 1.0565263

## Estimate u from selection ratios as direct range restriction:
estimate_u(sr = sr)

#> [1] 0.6028103 0.6498123 0.7020069 0.5584071

## Estimate u from selection ratios as indirect range restriction:
estimate_u(sr = sr, rxya_est = rxya_est)

#> [1] 0.8408451 0.8555640 0.8732034 0.8279546