Non-linear estimate of variance of \(\rho\) corrected for psychometric artifacts using numeric integration

Functions to estimate the variance of \(\rho\) corrected for psychometric artifacts. These functions integrate over the residual distribution of correlations from an interactive artifact-distribution meta-analysis to non-linearly estimate the variance of \(\rho\).

Available functions include:

estimate_var_rho_int_meas
Variance of \(\rho\) corrected for measurement error only

estimate_var_rho_int_uvdrr
Variance of \(\rho\) corrected for univariate direct range restriction (i.e., Case II) and measurement error

estimate_var_rho_int_bvdrr
Variance of \(\rho\) corrected for bivariate direct range restriction and measurement error

estimate_var_rho_int_uvirr
Variance of \(\rho\) corrected for univariate indirect range restriction (i.e., Case IV) and measurement error

estimate_var_rho_int_bvirr
Variance of \(\rho\) corrected for bivariate indirect range restriction (i.e., Case V) and measurement error

estimate_var_rho_int_rb
Variance of \(\rho\) corrected using Raju and Burke's correction for direct range restriction and measurement error

estimate_var_rho_int_meas(mean_qx, mean_qy, var_res)

estimate_var_rho_int_uvdrr(
  mean_rxyi,
  mean_rtpa,
  mean_qxa,
  mean_qyi,
  mean_ux,
  var_res
)

estimate_var_rho_int_uvirr(
  mean_rxyi,
  mean_rtpa,
  mean_qxi,
  mean_qyi,
  mean_ut,
  var_res
)

estimate_var_rho_int_bvirr(mean_qxa, mean_qya, mean_ux, mean_uy, var_res)

estimate_var_rho_int_bvdrr(
  mean_rxyi,
  mean_rtpa,
  mean_qxa,
  mean_qya,
  mean_ux,
  mean_uy,
  var_res
)

estimate_var_rho_int_rb(
  mean_rxyi,
  mean_rtpa,
  mean_qx,
  mean_qy,
  mean_ux,
  var_res
)

Arguments

mean_qx
Mean square root of reliability for X.

mean_qy
Mean square root of reliability for Y.

var_res
Residual variance from an interative artifact distribution (i.e., variance of observed correlations minus predicted error variance and predicted artifact variance).

mean_rxyi
Mean observed correlation.

mean_rtpa
Mean corrected correlation.

mean_qxa
Mean square root of unrestricted reliability for X.

mean_qyi
Mean square root of restricted reliability for Y.

mean_ux
Mean observed-score u ratio for X.

mean_qxi
Mean square root of restricted reliability for X.

mean_ut
Mean true-score u ratio for X.

mean_qya
Mean square root of unrestricted reliability for Y.

mean_uy
Mean observed-score u ratio for Y.

Value

A vector of non-linear estimates of the variance of rho.

Notes

estimate_var_rho_int_meas and estimate_var_rho_int_bvirr do not make use of numeric integration because they are linear functions.

References

Law, K. S., Schmidt, F. L., & Hunter, J. E. (1994). Nonlinearity of range corrections in meta-analysis: Test of an improved procedure. Journal of Applied Psychology, 79(3), 425–438. https://doi.org/10.1037/0021-9010.79.3.425