Taylor Series Approximation of variance of \(\rho\) corrected for psychometric artifacts

Functions to estimate the variance of \(\rho\) corrected for psychometric artifacts. These functions use Taylor series approximations (i.e., the delta method) to estimate the variance in observed effect sizes predictable from the variance in artifact distributions based on the partial derivatives.

The available Taylor-series functions include:

estimate_var_rho_tsa_meas
Variance of \(\rho\) corrected for measurement error only

estimate_var_rho_tsa_uvdrr
Variance of \(\rho\) corrected for univariate direct range restriction (i.e., Case II) and measurement error

estimate_var_rho_tsa_bvdrr
Variance of \(\rho\) corrected for bivariate direct range restriction and measurement error

estimate_var_rho_tsa_uvirr
Variance of \(\rho\) corrected for univariate indirect range restriction (i.e., Case IV) and measurement error

estimate_var_rho_tsa_bvirr
Variance of \(\rho\) corrected for bivariate indirect range restriction (i.e., Case V) and measurement error

estimate_var_rho_tsa_rb1
Variance of \(\rho\) corrected using Raju and Burke's TSA1 correction for direct range restriction and measurement error

estimate_var_rho_tsa_rb2
Variance of \(\rho\) corrected using Raju and Burke's TSA2 correction for direct range restriction and measurement error. Note that a typographical error in Raju and Burke's article has been corrected in this function so as to compute appropriate partial derivatives.

estimate_var_rho_tsa_meas(
  mean_rtp,
  var_rxy,
  var_e,
  mean_qx = 1,
  var_qx = 0,
  mean_qy = 1,
  var_qy = 0,
  ...
)

estimate_var_rho_tsa_uvdrr(
  mean_rtpa,
  var_rxyi,
  var_e,
  mean_ux = 1,
  var_ux = 0,
  mean_qxa = 1,
  var_qxa = 0,
  mean_qyi = 1,
  var_qyi = 0,
  ...
)

estimate_var_rho_tsa_bvdrr(
  mean_rtpa,
  var_rxyi,
  var_e = 0,
  mean_ux = 1,
  var_ux = 0,
  mean_uy = 1,
  var_uy = 0,
  mean_qxa = 1,
  var_qxa = 0,
  mean_qya = 1,
  var_qya = 0,
  ...
)

estimate_var_rho_tsa_uvirr(
  mean_rtpa,
  var_rxyi,
  var_e,
  mean_ut = 1,
  var_ut = 0,
  mean_qxa = 1,
  var_qxa = 0,
  mean_qyi = 1,
  var_qyi = 0,
  ...
)

estimate_var_rho_tsa_bvirr(
  mean_rtpa,
  var_rxyi,
  var_e = 0,
  mean_ux = 1,
  var_ux = 0,
  mean_uy = 1,
  var_uy = 0,
  mean_qxa = 1,
  var_qxa = 0,
  mean_qya = 1,
  var_qya = 0,
  sign_rxz = 1,
  sign_ryz = 1,
  ...
)

estimate_var_rho_tsa_rb1(
  mean_rtpa,
  var_rxyi,
  var_e,
  mean_ux = 1,
  var_ux = 0,
  mean_rxx = 1,
  var_rxx = 0,
  mean_ryy = 1,
  var_ryy = 0,
  ...
)

estimate_var_rho_tsa_rb2(
  mean_rtpa,
  var_rxyi,
  var_e,
  mean_ux = 1,
  var_ux = 0,
  mean_qx = 1,
  var_qx = 0,
  mean_qy = 1,
  var_qy = 0,
  ...
)

Arguments

mean_rtp
Mean corrected correlation.

var_rxy
Variance of observed correlations.

var_e
Error variance of observed correlations

mean_qx
Mean square root of reliability for X.

var_qx
Variance of square roots of reliability estimates for X.

mean_qy
Mean square root of reliability for Y.

var_qy
Variance of square roots of reliability estimates for Y.

...
Additional arguments.

mean_rtpa
Mean corrected correlation.

var_rxyi
Variance of observed correlations.

mean_ux
Mean observed-score u ratio for X.

var_ux
Variance of observed-score u ratios for X.

mean_qxa
Mean square root of unrestricted reliability for X.

var_qxa
Variance of square roots of unrestricted reliability estimates for X.

mean_qyi
Mean square root of restricted reliability for Y.

var_qyi
Variance of square roots of restricted reliability estimates for Y.

mean_uy
Mean observed-score u ratio for Y.

var_uy
Variance of observed-score u ratios for Y.

mean_qya
Mean square root of unrestricted reliability for Y.

var_qya
Variance of square roots of unrestricted reliability estimates for Y.

mean_ut
Mean true-score u ratio for X.

var_ut
Variance of true-score u ratios for X.

sign_rxz
Sign of the relationship between X and the selection mechanism.

sign_ryz
Sign of the relationship between Y and the selection mechanism.

mean_rxx
Mean reliability for X.

var_rxx
Variance of reliability estimates for X.

mean_ryy
Mean reliability for Y.

var_ryy
Variance of reliability estimates for Y.

Value

Vector of meta-analytic variances estimated via Taylor series approximation.

Details

######## Measurement error only ########

The attenuation formula for measurement error is

$$\rho_{XY}=\rho_{TP}q_{X}q_{Y}$$ where \(\rho_{XY}\) is an observed correlation, \(\rho_{TP}\) is a true-score correlation, and \(q_{X}\) and \(q_{Y}\) are the square roots of reliability coefficients for X and Y, respectively.

The Taylor series approximation of the variance of \(\rho_{TP}\) can be computed using the following linear equation,

$$var_{\rho_{TP}} \approx \left[var_{r_{XY}}-var_{e}-\left(b_{1}^{2}var_{q_{X}}+b_{2}^{2}var_{q_{Y}}\right)\right]/b_{3}^{2}$$

where \(b_{1}\), \(b_{2}\), and \(b_{3}\) are first-order partial derivatives of the attenuation formula with respect to \(q_{X}\), \(q_{Y}\), and \(\rho_{TP}\), respectively. The first-order partial derivatives of the attenuation formula are:

$$b_{1}=\frac{\partial\rho_{XY}}{\partial q_{X}}=\rho_{TP}q_{Y}$$ $$b_{2}=\frac{\partial\rho_{XY}}{\partial q_{Y}}=\rho_{TP}q_{X}$$ $$b_{3}=\frac{\partial\rho_{XY}}{\partial\rho_{TP}}=q_{X}q_{Y}$$

######## Univariate direct range restriction (UVDRR; i.e., Case II) ########

The UVDRR attenuation procedure may be represented as

$$\rho_{XY_{i}}=\frac{\rho_{TP_{a}}q_{Y_{i}}q_{X_{a}}u_{X}}{\sqrt{\rho_{TP_{a}}^{2}q_{X_{a}}^{2}\left(u_{X}^{2}-1\right)+1}}$$

The attenuation formula can also be represented as:

$$\rho_{XY_{i}}=\rho_{TP_{a}}q_{Y_{i}}q_{X_{a}}u_{X}A$$

where

$$A=\frac{1}{\sqrt{\rho_{TP_{a}}^{2}q_{X_{a}}^{2}\left(u_{X}^{2}-1\right)+1}}$$

The Taylor series approximation of the variance of \(\rho_{TP_{a}}\) can be computed using the following linear equation,

$$var_{\rho_{TP_{a}}} \approx \left[var_{r_{XY_{i}}}-var_{e}-\left(b_{1}^{2}var_{q_{X_{a}}}+b_{2}^{2}var_{q_{Y_{i}}}+b_{3}^{2}var_{u_{X}}\right)\right]/b_{4}^{2}$$

where \(b_{1}\), \(b_{2}\), \(b_{3}\), and \(b_{4}\) are first-order partial derivatives of the attenuation formula with respect to \(q_{X_{a}}\), \(q_{Y_{i}}\), \(u_{X}\), and \(\rho_{TP_{a}}\), respectively. The first-order partial derivatives of the attenuation formula are:

$$b_{1}=\frac{\partial\rho_{XY_{i}}}{\partial q_{X_{a}}}=\rho_{TP_{a}}q_{Y_{i}}u_{X}A^{3}$$ $$b_{2}=\frac{\partial\rho_{XY_{i}}}{\partial q_{Y_{i}}}=\frac{\rho_{XY_{i}}}{q_{Y_{i}}}$$ $$b_{3}=\frac{\partial\rho_{XY_{i}}}{\partial u_{X}}=-\rho_{TP_{a}}q_{Y_{i}}q_{X_{a}}\left(\rho_{TP_{a}}^{2}q_{X_{a}}^{2}-1\right)A^{3}$$ $$b_{4}=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{TP_{a}}}=q_{Y_{i}}q_{X_{a}}u_{X}A^{3}$$

######## Univariate indirect range restriction (UVIRR; i.e., Case IV) ########

Under univariate indirect range restriction, the attenuation formula yielding \(\rho_{XY_{i}}\) is:

$$\rho_{XY_{i}}=\frac{u_{T}q_{X_{a}}}{\sqrt{u_{T}^{2}q_{X_{a}}^{2}+1-q_{X_{a}}^{2}}}\frac{u_{T}\rho_{TP_{a}}}{\sqrt{u_{T}^{2}\rho_{TP_{a}}^{2}+1-\rho_{TP_{a}}^{2}}}$$

The attenuation formula can also be represented as:

$$\rho_{XY_{i}}=q_{X_{a}}q_{Y_{i}}\rho_{TP_{a}}u_{T}^{2}AB$$

where

$$A=\frac{1}{\sqrt{u_{T}^{2}q_{X_{a}}^{2}+1-q_{X_{a}}^{2}}}$$

and

$$B=\frac{1}{\sqrt{u_{T}^{2}\rho_{TP_{a}}^{2}+1-\rho_{TP_{a}}^{2}}}$$

The Taylor series approximation of the variance of \(\rho_{TP_{a}}\) can be computed using the following linear equation,

$$var_{\rho_{TP_{a}}} \approx \left[var_{r_{XY_{i}}}-var_{e}-\left(b_{1}^{2}var_{q_{X_{a}}}+b_{2}^{2}var_{q_{Y_{i}}}+b_{3}^{2}var_{u_{T}}\right)\right]/b_{4}^{2}$$

where \(b_{1}\), \(b_{2}\), \(b_{3}\), and \(b_{4}\) are first-order partial derivatives of the attenuation formula with respect to \(q_{X_{a}}\), \(q_{Y_{i}}\), \(u_{T}\), and \(\rho_{TP_{a}}\), respectively. The first-order partial derivatives of the attenuation formula are:

$$b_{1}=\frac{\partial\rho_{XY_{i}}}{\partial q_{X_{a}}}=\frac{\rho_{XY_{i}}}{q_{X_{a}}}-\rho_{XY_{i}}q_{X_{a}}B^{2}\left(u_{T}^{2}-1\right)$$ $$b_{2}=\frac{\partial\rho_{XY_{i}}}{\partial q_{Y_{i}}}=\frac{\rho_{XY_{i}}}{q_{Y_{i}}}$$ $$b_{3}=\frac{\partial\rho_{XY_{i}}}{\partial u_{T}}=\frac{2\rho_{XY_{i}}}{u_{T}}-\rho_{XY_{i}}u_{T}q_{X_{a}}^{2}B^{2}-\rho_{XY_{i}}u_{T}\rho_{TP_{a}}^{2}A^{2}$$ $$b_{4}=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{TP_{a}}}=\frac{\rho_{XY_{i}}}{\rho_{TP_{a}}}-\rho_{XY_{i}}\rho_{TP_{a}}A^{2}\left(u_{T}^{2}-1\right)$$

######## Bivariate direct range restriction (BVDRR) ########

Under bivariate direct range restriction, the attenuation formula yielding \(\rho_{XY_{i}}\) is:

$$\rho_{XY_{i}}=\frac{A+\rho_{TP_{a}}^{2}q_{X_{a}}q_{Y_{a}}-\frac{1}{q_{X_{a}}q_{Y_{a}}}}{2\rho_{TP_{a}}u_{X}u_{Y}}$$

where

$$A=\sqrt{\left(\frac{1}{q_{X_{a}}q_{Y_{a}}}-\rho_{TP_{a}}^{2}q_{X_{a}}q_{Y_{a}}\right)^{2}+4\rho_{TP_{a}}u_{X}^{2}u_{Y}^{2}}$$

The Taylor series approximation of the variance of \(\rho_{TP_{a}}\) can be computed using the following linear equation,

$$var_{\rho_{TP_{a}}} \approx \left[var_{r_{XY_{i}}}-var_{e}-\left(b_{1}^{2}var_{q_{X_{a}}}+b_{2}^{2}var_{q_{Y_{i}}}+b_{3}^{2}var_{u_{X}}+b_{4}^{2}var_{u_{Y}}\right)\right]/b_{5}^{2}$$

where \(b_{1}\), \(b_{2}\), \(b_{3}\), \(b_{4}\), and \(b_{5}\) are first-order partial derivatives of the attenuation formula with respect to \(q_{X_{a}}\), \(q_{Y_{a}}\), \(u_{X}\), \(u_{Y}\), and \(\rho_{TP_{a}}\), respectively. First, we define terms to simplify the computation of partial derivatives:

$$B=\left(\rho_{TP_{a}}^{2}q_{X_{a}}^{2}q_{Y_{a}}^{2}+q_{X_{a}}q_{Y_{a}}A-1\right)$$

$$C=2\rho_{TP_{a}}q_{X_{a}}^{2}q_{Y_{a}}^{2}u_{X}u_{Y}A$$

The first-order partial derivatives of the attenuation formula are:

$$b_{1}=\frac{\partial\rho_{XY_{i}}}{\partial q_{X_{a}}}=\frac{\left(\rho_{TP_{a}}^{2}q_{X_{a}}^{2}q_{Y_{a}}^{2}+1\right)B}{q_{X_{a}}C}$$ $$b_{2}=\frac{\partial\rho_{XY_{i}}}{\partial q_{Y_{i}}}=\frac{\left(\rho_{TP_{a}}^{2}q_{X_{a}}^{2}q_{Y_{a}}^{2}+1\right)B}{q_{Y_{a}}C}$$ $$b_{3}=\frac{\partial\rho_{XY_{i}}}{\partial u_{X}}=-\frac{\left(\rho_{TP_{a}}q_{X_{a}}q_{Y_{a}}-1\right)\left(\rho_{TP_{a}}q_{X_{a}}q_{Y_{a}}+1\right)B}{u_{X}C}$$ $$b_{4}=\frac{\partial\rho_{XY_{i}}}{\partial u_{Y}}=-\frac{\left(\rho_{TP_{a}}q_{X_{a}}q_{Y_{a}}-1\right)\left(\rho_{TP_{a}}q_{X_{a}}q_{Y_{a}}+1\right)B}{u_{Y}C}$$ $$b_{5}=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{TP_{a}}}=\frac{\left(\rho_{TP_{a}}^{2}q_{X_{a}}^{2}q_{Y_{a}}^{2}+1\right)B}{\rho_{TP_{a}}C}$$

######## Bivariate indirect range restriction (BVIRR; i.e., Case V) ########

Under bivariate indirect range restriction, the attenuation formula yielding \(\rho_{XY_{i}}\) is:

$$\rho_{XY_{i}}=\frac{\rho_{TP_{a}}q_{X_{a}}q_{Y_{a}}-\lambda\sqrt{\left|1-u_{X}^{2}\right|\left|1-u_{Y}^{2}\right|}}{u_{X}u_{Y}}$$

The Taylor series approximation of the variance of \(\rho_{TP_{a}}\) can be computed using the following linear equation,

$$var_{\rho_{TP_{a}}} \approx \left[var_{r_{XY_{i}}}-var_{e}-\left(b_{1}^{2}var_{q_{X_{a}}}+b_{2}^{2}var_{q_{Y_{i}}}+b_{3}^{2}var_{u_{X}}+b_{4}^{2}var_{u_{Y}}\right)\right]/b_{5}^{2}$$

where \(b_{1}\), \(b_{2}\), \(b_{3}\), \(b_{4}\), and \(b_{5}\) are first-order partial derivatives of the attenuation formula with respect to \(q_{X_{a}}\), \(q_{Y_{a}}\), \(u_{X}\), \(u_{Y}\), and \(\rho_{TP_{a}}\), respectively. First, we define terms to simplify the computation of partial derivatives:

$$b_{1}=\frac{\partial\rho_{XY_{i}}}{\partial q_{X_{a}}}=\frac{\rho_{TP_{a}}q_{Y_{a}}}{u_{X}u_{Y}}$$ $$b_{2}=\frac{\partial\rho_{XY_{i}}}{\partial q_{Y_{i}}}=\frac{\rho_{TP_{a}}q_{X_{a}}}{u_{X}u_{Y}}$$ $$b_{3}=\frac{\partial\rho_{XY_{i}}}{\partial u_{X}}=\frac{\lambda\left(1-u_{X}^{2}\right)\sqrt{\left|1-u_{Y}^{2}\right|}}{u_{Y}\left|1-u_{X}^{2}\right|^{1.5}}-\frac{\rho_{XY_{i}}}{u_{X}}$$ $$b_{4}=\frac{\partial\rho_{XY_{i}}}{\partial u_{Y}}=\frac{\lambda\left(1-u_{Y}^{2}\right)\sqrt{\left|1-u_{X}^{2}\right|}}{u_{X}\left|1-u_{Y}^{2}\right|^{1.5}}-\frac{\rho_{XY_{i}}}{u_{Y}}$$ $$b_{5}=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{TP_{a}}}=\frac{q_{X_{a}}q_{Y_{a}}}{u_{X}u_{Y}}$$

######## Raju and Burke's TSA1 procedure ########

Raju and Burke's attenuation formula may be represented as

$$\rho_{XY_{i}}=\frac{\rho_{TP_{a}}u_{X}\sqrt{\rho_{XX_{a}}\rho_{YY_{a}}}}{\sqrt{\rho_{TP_{a}}^{2}\rho_{XX_{a}}\rho_{YY_{a}}u_{X}^{2}-\rho_{TP_{a}}^{2}\rho_{XX_{a}}\rho_{YY_{a}}+1}}$$

The Taylor series approximation of the variance of \(\rho_{TP_{a}}\) can be computed using the following linear equation,

$$var_{\rho_{TP_{a}}} \approx \left[var_{r_{XY_{i}}}-var_{e}-\left(B^{2}var_{\rho_{YY_{a}}}+C^{2}var_{\rho_{XX_{a}}}+D^{2}var_{u_{X}}\right)\right]/A^{2}$$

where A, B, C, and D are first-order partial derivatives of the attenuation formula with respect to \(\rho_{TP_{a}}\), \(\rho_{XX_{a}}\), \(\rho_{YY_{a}}\), and \(u_{X}\), respectively. The first-order partial derivatives of the attenuation formula are:

$$A=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{TP_{a}}}=\frac{\rho_{XY_{i}}}{\rho_{TP_{a}}}+\frac{\rho_{XY_{i}\left(1-u_{X}^{2}\right)}^{3}}{\rho_{TP_{a}}u_{X}^{2}}$$ $$B=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{YY_{a}}}=\frac{1}{2}\left(\frac{\rho_{XY_{i}}}{\rho_{YY_{a}}}+\frac{\rho_{XY_{i}\left(1-u_{X}^{2}\right)}^{3}}{\rho_{YY_{a}}u_{X}^{2}}\right)$$ $$C=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{XX_{a}}}=\frac{1}{2}\left(\frac{\rho_{XY_{i}}}{\rho_{XX_{a}}}+\frac{\rho_{XY_{i}\left(1-u_{X}^{2}\right)}^{3}}{\rho_{XX_{a}}u_{X}^{2}}\right)$$ $$D=\frac{\partial\rho_{XY_{i}}}{\partial u_{X}}=\frac{\rho_{XY_{i}}-\rho_{XY_{i}}^{3}}{u_{X}}$$

######## Raju and Burke's TSA2 procedure ########

Raju and Burke's attenuation formula may be represented as

$$\rho_{XY_{i}}=\frac{\rho_{TP_{a}}q_{X_{a}}q_{Y_{a}}u_{X}}{\sqrt{\rho_{TP_{a}}^{2}q_{X_{a}}^{2}q_{Y_{a}}^{2}u_{X}^{2}-\rho_{TP_{a}}^{2}q_{X_{a}}^{2}q_{Y_{a}}^{2}+1}}$$

The Taylor series approximation of the variance of \(\rho_{TP_{a}}\) can be computed using the following linear equation,

$$var_{\rho_{TP_{a}}} \approx \left[var_{r_{XY_{i}}}-var_{e}-\left(F^{2}var_{q_{Y_{a}}}+G^{2}var_{q_{X_{a}}}+H^{2}var_{u_{X}}\right)\right]/E^{2}$$

where E, F, G, and H are first-order partial derivatives of the attenuation formula with respect to \(\rho_{TP_{a}}\), \(q_{X_{a}}\), \(q_{Y_{a}}\), and \(u_{X}\), respectively. The first-order partial derivatives of the attenuation formula (with typographic errors in the original article corrected) are:

$$E=\frac{\partial\rho_{XY_{i}}}{\partial\rho_{TP_{a}}}=\frac{\rho_{XY_{i}}}{\rho_{TP_{a}}}+\frac{\rho_{XY_{i}\left(1-u_{X}^{2}\right)}^{3}}{\rho_{TP_{a}}u_{X}^{2}}$$ $$F=\frac{\partial\rho_{XY_{i}}}{\partial q_{Y_{a}}}=\frac{\rho_{XY_{i}}}{q_{Y_{a}}}+\frac{\rho_{XY_{i}\left(1-u_{X}^{2}\right)}^{3}}{q_{Y_{a}}u_{X}^{2}}$$ $$G=\frac{\partial\rho_{XY_{i}}}{\partial q_{X_{a}}}=\frac{\rho_{XY_{i}}}{q_{X_{a}}}+\frac{\rho_{XY_{i}\left(1-u_{X}^{2}\right)}^{3}}{q_{X_{a}}u_{X}^{2}}$$ $$H=\frac{\partial\rho_{XY_{i}}}{\partial u_{X}}=\frac{\rho_{XY_{i}}-\rho_{XY_{i}}^{3}}{u_{X}}$$

Notes

A typographical error in Raju and Burke's article has been corrected in estimate_var_rho_tsa_rb2 so as to compute appropriate partial derivatives.

References

Dahlke, J. A., & Wiernik, B. M. (2019). Not restricted to selection research: Accounting for indirect range restriction in organizational research. Organizational Research Methods. Advance online publication. https://doi.org/10.1177/1094428119859398

Hunter, J. E., Schmidt, F. L., & Le, H. (2006). Implications of direct and indirect range restriction for meta-analysis methods and findings. Journal of Applied Psychology, 91(3), 594–612. https://doi.org/10.1037/0021-9010.91.3.594

Raju, N. S., & Burke, M. J. (1983). Two new procedures for studying validity generalization. Journal of Applied Psychology, 68(3), 382–395. https://doi.org/10.1037/0021-9010.68.3.382

Examples

estimate_var_rho_tsa_meas(mean_rtp = .5, var_rxy = .02, var_e = .01,
                 mean_qx = .8, var_qx = .005,
                 mean_qy = .8, var_qy = .005)

#>   var_art var_pre var_res    var_rho
#> 1  0.0016  0.0116  0.0084 0.02050781

estimate_var_rho_tsa_uvdrr(mean_rtpa = .5, var_rxyi = .02, var_e = .01,
                  mean_ux = .8, var_ux = .005,
                  mean_qxa = .8, var_qxa = .005,
                  mean_qyi = .8, var_qyi = .005)

#>       var_art    var_pre     var_res    var_rho
#> 1 0.001586671 0.01158667 0.008413329 0.02686172

estimate_var_rho_tsa_bvdrr(mean_rtpa = .5, var_rxyi = .02, var_e = .01,
                  mean_ux = .8, var_ux = .005,
                  mean_uy = .8, var_uy = .005,
                  mean_qxa = .8, var_qxa = .005,
                  mean_qya = .8, var_qya = .005)

#>       var_art    var_pre     var_res    var_rho
#> 1 0.001532889 0.01153289 0.008467111 0.03588116

estimate_var_rho_tsa_uvirr(mean_rtpa = .5, var_rxyi = .02, var_e = .01,
                  mean_ut = .8, var_ut = .005,
                  mean_qxa = .8, var_qxa = .005,
                  mean_qyi = .8, var_qyi = .005)

#>       var_art    var_pre     var_res    var_rho
#> 1 0.002038837 0.01203884 0.007961163 0.02751984

estimate_var_rho_tsa_bvirr(mean_rtpa = .5, var_rxyi = .02, var_e = .01,
                  mean_ux = .8, var_ux = .005,
                  mean_uy = .8, var_uy = .005,
                  mean_qxa = .8, var_qxa = .005,
                  mean_qya = .8, var_qya = .005,
                  sign_rxz = 1, sign_ryz = 1)

#>      var_art    var_pre     var_res     var_rho
#> 1 0.02154541 0.03154541 -0.01154541 -0.01154541

estimate_var_rho_tsa_rb1(mean_rtpa = .5, var_rxyi = .02, var_e = .01,
                mean_ux = .8, var_ux = .005,
                mean_rxx = .8, var_rxx = .005,
                mean_ryy = .8, var_ryy = .005)

#>       var_art    var_pre     var_res    var_rho
#> 1 0.001152358 0.01115236 0.008847642 0.01807896

estimate_var_rho_tsa_rb2(mean_rtpa = .5, var_rxyi = .02, var_e = .01,
                mean_ux = .8, var_ux = .005,
                mean_qx = .8, var_qx = .005,
                mean_qy = .8, var_qy = .005)

#>       var_art    var_pre     var_res    var_rho
#> 1 0.001607852 0.01160785 0.008392148 0.02860198