Taylor Series Approximation of effect-size variances corrected for psychometric artifacts

Functions to estimate the variances corrected for psychometric artifacts. These functions use Taylor series approximations (i.e., the delta method) to estimate the corrected variance of an effect-size distribution.

The available Taylor-series functions include:

estimate_var_tsa_meas
Variance of \(\rho\) corrected for measurement error only

estimate_var_tsa_uvdrr
Variance of \(\rho\) corrected for univariate direct range restriction (i.e., Case II) and measurement error

estimate_var_tsa_bvdrr
Variance of \(\rho\) corrected for bivariate direct range restriction and measurement error

estimate_var_tsa_uvirr
Variance of \(\rho\) corrected for univariate indirect range restriction (i.e., Case IV) and measurement error

estimate_var_tsa_bvirr
Variance of \(\rho\) corrected for bivariate indirect range restriction (i.e., Case V) and measurement error

estimate_var_tsa_rb1
Variance of \(\rho\) corrected using Raju and Burke's TSA1 correction for direct range restriction and measurement error

estimate_var_tsa_rb2
Variance of \(\rho\) corrected using Raju and Burke's TSA2 correction for direct range restriction and measurement error. Note that a typographical error in Raju and Burke's article has been corrected in this function so as to compute appropriate partial derivatives.

estimate_var_tsa_meas(mean_rtp, var = 0, mean_qx = 1, mean_qy = 1, ...)

estimate_var_tsa_uvdrr(
  mean_rtpa,
  var = 0,
  mean_ux = 1,
  mean_qxa = 1,
  mean_qyi = 1,
  ...
)

estimate_var_tsa_bvdrr(
  mean_rtpa,
  var = 0,
  mean_ux = 1,
  mean_uy = 1,
  mean_qxa = 1,
  mean_qya = 1,
  ...
)

estimate_var_tsa_uvirr(
  mean_rtpa,
  var = 0,
  mean_ut = 1,
  mean_qxa = 1,
  mean_qyi = 1,
  ...
)

estimate_var_tsa_bvirr(
  mean_rtpa,
  var = 0,
  mean_ux = 1,
  mean_uy = 1,
  mean_qxa = 1,
  mean_qya = 1,
  sign_rxz = 1,
  sign_ryz = 1,
  ...
)

estimate_var_tsa_rb1(
  mean_rtpa,
  var = 0,
  mean_ux = 1,
  mean_rxx = 1,
  mean_ryy = 1,
  ...
)

estimate_var_tsa_rb2(
  mean_rtpa,
  var = 0,
  mean_ux = 1,
  mean_qx = 1,
  mean_qy = 1,
  ...
)

Arguments

mean_rtp
Mean corrected correlation.

var
Variance to be corrected for artifacts.

mean_qx
Mean square root of reliability for X.

mean_qy
Mean square root of reliability for Y.

...
Additional arguments.

mean_rtpa
Mean corrected correlation.

mean_ux
Mean observed-score u ratio for X.

mean_qxa
Mean square root of unrestricted reliability for X.

mean_qyi
Mean square root of restricted reliability for Y.

mean_uy
Mean observed-score u ratio for Y.

mean_qya
Mean square root of unrestricted reliability for Y.

mean_ut
Mean true-score u ratio for X.

sign_rxz
Sign of the relationship between X and the selection mechanism.

sign_ryz
Sign of the relationship between Y and the selection mechanism.

mean_rxx
Mean reliability for X.

mean_ryy
Mean reliability for Y.

Value

Vector of variances corrected for mean artifacts via Taylor series approximation.

Notes

A typographical error in Raju and Burke's article has been corrected in estimate_var_tsa_rb2 so as to compute appropriate partial derivatives.

References

Dahlke, J. A., & Wiernik, B. M. (2019). Not restricted to selection research: Accounting for indirect range restriction in organizational research. Organizational Research Methods. Advance online publication. https://doi.org/10.1177/1094428119859398

Hunter, J. E., Schmidt, F. L., & Le, H. (2006). Implications of direct and indirect range restriction for meta-analysis methods and findings. Journal of Applied Psychology, 91(3), 594–612. https://doi.org/10.1037/0021-9010.91.3.594

Raju, N. S., & Burke, M. J. (1983). Two new procedures for studying validity generalization. Journal of Applied Psychology, 68(3), 382–395. https://doi.org/10.1037/0021-9010.68.3.382

Examples

estimate_var_tsa_meas(mean_rtp = .5, var = .02,
                 mean_qx = .8,
                 mean_qy = .8)

#> [1] 0.04882812

estimate_var_tsa_uvdrr(mean_rtpa = .5, var = .02,
                  mean_ux = .8,
                  mean_qxa = .8,
                  mean_qyi = .8)

#> [1] 0.06385515

estimate_var_tsa_bvdrr(mean_rtpa = .5, var = .02,
                  mean_ux = .8,
                  mean_uy = .8,
                  mean_qxa = .8,
                  mean_qya = .8)

#> [1] 0.08475419

estimate_var_tsa_uvirr(mean_rtpa = .5, var = .02,
                  mean_ut = .8,
                  mean_qxa = .8,
                  mean_qyi = .8)

#> [1] 0.06913522

estimate_var_tsa_bvirr(mean_rtpa = .5, var = .02,
                  mean_ux = .8,
                  mean_uy = .8,
                  mean_qxa = .8,
                  mean_qya = .8,
                  sign_rxz = 1, sign_ryz = 1)

#> [1] 0.02

estimate_var_tsa_rb1(mean_rtpa = .5, var = .02,
                mean_ux = .8,
                mean_rxx = .8,
                mean_ryy = .8)

#> [1] 0.04086729

estimate_var_tsa_rb2(mean_rtpa = .5, var = .02,
                mean_ux = .8,
                mean_qx = .8,
                mean_qy = .8)

#> [1] 0.06816366