Filter meta-analyses

Filter psychmeta meta-analysis objects based on specified criteria.

filter_ma(
  ma_obj,
  analyses = "all",
  match = c("all", "any"),
  case_sensitive = TRUE,
  ...
)

filter_meta(
  ma_obj,
  analyses = "all",
  match = c("all", "any"),
  case_sensitive = TRUE,
  ...
)

Arguments

ma_obj
A psychmeta meta-analysis object.

analyses
Which analyses to extract? Can be either "all" to extract all meta-analyses in the object (default) or a list containing one or more of the following arguments:

construct: A list or vector of construct names to search for.

construct_pair: A list of vectors of construct pairs to search for.
(e.g., list(c("X", "Y"), c("X", "Z")) ).

pair_id: A list or vector of numeric construct pair IDs (unique construct-pair indices).

analysis_id: A list or vector of numeric analysis IDs (unique analysis indexes).

k_min: A numeric value specifying the minimum k for extracted meta-analyses.

N_min: A numeric value specifying the minimum N for extracted meta-analyses.

match
Should extracted meta-analyses match all (default) or any of the criteria given in analyses?

case_sensitive
Logical scalar that determines whether character values supplied in analyses should be treated as case sensitive (TRUE, default) or not (FALSE).

...
Additional arguments.

Value

A psychmeta meta-analysis object with analyses matching the specified criteria.

Examples

ma_obj <- ma_r(ma_method = "ic", rxyi = rxyi, n = n, rxx = rxxi, ryy = ryyi,
               construct_x = x_name, construct_y = y_name, sample_id = sample_id, citekey = NULL,
               moderators = moderator, data = data_r_meas_multi,
               impute_artifacts = FALSE, clean_artifacts = FALSE)

#>  **** Running ma_r: Meta-analysis of correlations **** 

ma_obj <- ma_r_ad(ma_obj, correct_rr_x = FALSE, correct_rr_y = FALSE)

#> Artifact-distribution meta-analyses have been added to 'ma_obj'

filter_ma(ma_obj, analyses="all")

#> Overview tibble of psychmeta meta-analysis of correlations  
#> ---------------------------------------------------------------------- 
#> # A tibble: 9 x 9
#>   analysis_id pair_id construct_x construct_y analysis_type    moderator 
#> *       <int>   <int> <chr>       <chr>       <fct>            <fct>     
#> 1           1       1 X           Y           Overall          All Levels
#> 2           2       1 X           Y           Simple Moderator 1         
#> 3           3       1 X           Y           Simple Moderator 2         
#> 4           4       2 X           Z           Overall          All Levels
#> 5           5       2 X           Z           Simple Moderator 1         
#> 6           6       2 X           Z           Simple Moderator 2         
#> 7           7       3 Y           Z           Overall          All Levels
#> 8           8       3 Y           Z           Simple Moderator 1         
#> 9           9       3 Y           Z           Simple Moderator 2         
#> # … with 3 more variables: meta_tables <named list>, escalc <named list>,
#> #   ad <named list>
#> 
#> To extract results, try summary() or the get_stuff functions (run ?get_stuff for help). 

filter_ma(ma_obj, analyses=list(construct="X"), match="all")

#> Overview tibble of psychmeta meta-analysis of correlations  
#> ---------------------------------------------------------------------- 
#> # A tibble: 6 x 9
#>   analysis_id pair_id construct_x construct_y analysis_type    moderator 
#> *       <int>   <int> <chr>       <chr>       <fct>            <fct>     
#> 1           1       1 X           Y           Overall          All Levels
#> 2           2       1 X           Y           Simple Moderator 1         
#> 3           3       1 X           Y           Simple Moderator 2         
#> 4           4       2 X           Z           Overall          All Levels
#> 5           5       2 X           Z           Simple Moderator 1         
#> 6           6       2 X           Z           Simple Moderator 2         
#> # … with 3 more variables: meta_tables <named list>, escalc <named list>,
#> #   ad <named list>
#> 
#> To extract results, try summary() or the get_stuff functions (run ?get_stuff for help). 

filter_ma(ma_obj, analyses=list(construct="X", k_min=21), match="any")

#> Overview tibble of psychmeta meta-analysis of correlations  
#> ---------------------------------------------------------------------- 
#> # A tibble: 9 x 9
#>   analysis_id pair_id construct_x construct_y analysis_type    moderator 
#> *       <int>   <int> <chr>       <chr>       <fct>            <fct>     
#> 1           1       1 X           Y           Overall          All Levels
#> 2           2       1 X           Y           Simple Moderator 1         
#> 3           3       1 X           Y           Simple Moderator 2         
#> 4           4       2 X           Z           Overall          All Levels
#> 5           5       2 X           Z           Simple Moderator 1         
#> 6           6       2 X           Z           Simple Moderator 2         
#> 7           7       3 Y           Z           Overall          All Levels
#> 8           8       3 Y           Z           Simple Moderator 1         
#> 9           9       3 Y           Z           Simple Moderator 2         
#> # … with 3 more variables: meta_tables <named list>, escalc <named list>,
#> #   ad <named list>
#> 
#> To extract results, try summary() or the get_stuff functions (run ?get_stuff for help). 

filter_ma(ma_obj, analyses=list(construct="X", k_min=21), match="all")

#> Overview tibble of psychmeta meta-analysis of correlations  
#> ---------------------------------------------------------------------- 
#> # A tibble: 2 x 9
#>   analysis_id pair_id construct_x construct_y analysis_type moderator 
#> *       <int>   <int> <chr>       <chr>       <fct>         <fct>     
#> 1           1       1 X           Y           Overall       All Levels
#> 2           4       2 X           Z           Overall       All Levels
#> # … with 3 more variables: meta_tables <named list>, escalc <named list>,
#> #   ad <named list>
#> 
#> To extract results, try summary() or the get_stuff functions (run ?get_stuff for help).