Supplemental heterogeneity statistics for meta-analyses

This function computes a variety of supplemental statistics for meta-analyses. The statistics here are included for interested users. It is strongly recommended that heterogeneity in meta-analysis be interpreted using the \(SD_{res}\), \(SD_{\rho}\), and \(SD_{\delta}\) statistics, along with corresponding credibility intervals, which are reported in the default ma_obj output (Wiernik et al., 2017).

heterogeneity(
  ma_obj,
  es_failsafe = NULL,
  conf_level = attributes(ma_obj)$inputs$conf_level,
  var_res_ci_method = c("profile_var_es", "profile_Q", "normal_logQ"),
  ...
)

Arguments

ma_obj
Meta-analysis object.

es_failsafe
Failsafe effect-size value for file-drawer analyses.

conf_level
Confidence level to define the width of confidence intervals (default is conf_level specified in ma_obj).

var_res_ci_method
Which method to use to estimate the limits. Options are profile_var_es for a profile-likelihood interval assuming \(\sigma^{2}_es ~ \chi^{2}(k-1)\), profile_Q for a profile-likelihood interval assuming \(Q ~ \chi^{2}(k-1, \lambda), \lambda = \sum_{i=1}^{k} w_i(\theta - \bar{\theta})^{2}\), and normal_logQ for a delta method assuming log(Q) follows a standard normal distribution.

...
Additional arguments.

Value

ma_obj with heterogeneity statistics added. Included statistics include:

es_type

The effect size metric used.

percent_var_accounted

Percent variance accounted for statistics (by sampling error, by other artifacts, and total). These statistics are widely reported, but not recommended, as they tend to be misinterpreted as suggesting only a small portion of the observed variance is accounted for by sampling error and other artifacts (Schmidt, 2010; Schmidt & Hunter, 2015, p. 15, 425). The square roots of these values are more interpretable and appropriate indices of the relations between observed effect sizes and statistical artifacts (see cor(es, perturbations)).

cor(es, perturbations)

The correlation between observed effect sizes and statistical artifacts in each sample (with sampling error, with other artifacts, and with artifacts in total), computed as \(\sqrt{percent\;var\;accounted}\). These indices are more interpretable and appropriate indices of the relations between observed effect sizes and statistical artifacts than percent_var_accounted.

rel_es_obs

\(1-\frac{var_{pre}}{var_{es}}\), the reliability of observed effect size differences as indicators of true effect sizes differences in the sampled studies. This value is useful for correcting correlations between moderators and effect sizes in meta-regression.

H_squared

The ratio of the observed effect size variance to the predicted (error) variance. Also the square root of Q divided by its degrees of freedom.

H

The ratio of the observed effect size standard deviation to the predicted (error) standard deviation.

I_squared

The estimated percent variance not accounted for by sampling error or other artifacts (attributable to moderators and uncorrected artifacts). This statistic is simply rel_es_obs expressed as a percentage rather than a decimal.

Q

Cochran's \(\chi^{2}\) statistic. Significance tests using this statistic are strongly discouraged; heterogeneity should instead be determined by examining the width of the credibility interval and the practical differences between effect sizes contained within it (Wiernik et al., 2017). This value is not accurate when artifact distribution methods are used for corrections.

tau_squared

\(\tau^{2}\), an estimator of the random effects variance component (analogous to the Hunter-Schmidt \(SD_{res}^{2}\), \(SD_{\rho}^{2}\), or \(SD_{\delta}^{2}\) statistics), with its confidence interval. This value is not accurate when artifact distribution methods are used for corrections.

tau

\(\sqrt{\tau^{2}}\), analogous to the Hunter-Schmidt \(SD_{res}\), \(SD_{\rho}\), and \(SD_{\delta}\) statistics, with its confidence interval. This value is not accurate when artifact distribution methods are used for corrections.

Q_r, H_r_squared, H_r, I_r_squared, tau_r_squared, tau_r

Outlier-robust versions of these statistics, computed based on absolute deviations from the weighted mean effect size (see Lin et al., 2017). These values are not accurate when artifact distribution methods are used for corrections.

Q_m, H_m_squared, H_m, I_m_squared, tau_m_squared, tau_m

Outlier-robust versions of these statistics, computed based on absolute deviations from the weighted median effect size (see Lin et al., 2017). These values are not accurate when artifact distribution methods are used for corrections.

file_drawer

Fail-safe N and k statistics (file-drawer analyses). These statistics should not be used to evaluate publication bias, as they counterintuitively suggest less when publication bias is strong (Becker, 2005). However, in the absence of publication bias, they can be used as an index of second-order sampling error (how likely is a mean effect to reduce to the specified value with additional studies?). The confidence interval around the mean effect can be used more directly for the same purpose.

Results are reported using computation methods described by Schmidt and Hunter. For barebones and indivdiual-correction meta-analyses, results are also reported using computation methods described by DerSimonian and Laird, outlier-robust computation methods, and, if weights from \pkg{metafor} are used, heterogeneity results from \pkg{metafor}.
References

Becker, B. J. (2005). Failsafe N or file-drawer number. In H. R. Rothstein, A. J. Sutton, & M. Borenstein (Eds.), Publication bias in meta-analysis: Prevention, assessment and adjustments (pp. 111–125). Hoboken, NJ: Wiley. https://doi.org/10.1002/0470870168.ch7

Higgins, J. P. T., & Thompson, S. G. (2002). Quantifying heterogeneity in a meta-analysis. Statistics in Medicine, 21(11), 1539–1558. https://doi.org/10.1002/sim.1186

Lin, L., Chu, H., & Hodges, J. S. (2017). Alternative measures of between-study heterogeneity in meta-analysis: Reducing the impact of outlying studies. Biometrics, 73(1), 156–166. https://doi.org/10.1111/biom.12543

Schmidt, F. (2010). Detecting and correcting the lies that data tell. Perspectives on Psychological Science, 5(3), 233–242. https://doi.org/10.1177/1745691610369339

Schmidt, F. L., & Hunter, J. E. (2015). Methods of meta-analysis: Correcting error and bias in research findings (3rd ed.). Thousand Oaks, CA: Sage. https://doi.org/10/b6mg. pp. 15, 414, 426, 533–534.

Wiernik, B. M., Kostal, J. W., Wilmot, M. P., Dilchert, S., & Ones, D. S. (2017). Empirical benchmarks for interpreting effect size variability in meta-analysis. Industrial and Organizational Psychology, 10(3). https://doi.org/10.1017/iop.2017.44

Examples

## Correlations
ma_obj <- ma_r_ic(rxyi = rxyi, n = n, rxx = rxxi, ryy = ryyi, ux = ux,
                  correct_rr_y = FALSE, data = data_r_uvirr)
ma_obj <- ma_r_ad(ma_obj, correct_rr_y = FALSE)

#> Artifact-distribution meta-analyses have been added to 'ma_obj'

ma_obj <- heterogeneity(ma_obj = ma_obj)

#> Heterogeneity analyses have been added to 'ma_obj' - use get_heterogeneity() to retrieve them.

ma_obj$heterogeneity[[1]]$barebones

#> 
#> Heterogeneity results for r 
#> ---------------------------
#> 
#> Accounted for a total of 57.076% of variance
#> 
#> Correlation between r values and artifactual perturbations: 0.755
#> 
#> The reliability of observed effect sizes is: 0.429
#> 
#> 
#> Random effects variance estimates
#> ---------------------------------
#> Hunter-Schmidt method:
#>   sd_res  (tau):   0.060, SE = 0.022, 95% CI = [0.008, 0.114] 
#>   var_res (tau^2): 0.004, SE = 0.003, 95% CI = [0.000, 0.013] 
#> 
#>   Q statistic: 33.289 (df = 19, p = 0.022) 
#>   H: 1.324   H^2: 1.752   I^2: 42.924
#> 
#> DerSimonian-Laird method:
#>   sd_res  (tau):   0.061
#>   var_res (tau^2): 0.004
#> 
#>   Q statistic: 33.334
#>   H: 1.325   H^2: 1.754   I^2: 43.000
#> 
#> Outlier-robust method (absolute deviation from mean):
#>   sd_res  (tau_r):   0.062
#>   var_res (tau_r^2): 0.004
#> 
#>   Q_r statistic: 20.714
#>   H_r: 1.332   H_r^2: 1.774   I_r^2: 0.436
#> 
#> Outlier-robust method (absolute deviation from median):
#>   sd_res  (tau_m):   0.059
#>   var_res (tau_m^2): 0.003
#> 
#>   Q_m statistic: 20.816
#>   H_m: 1.304   H_m^2: 1.702   I_m^2: 0.412
#> 

ma_obj$heterogeneity[[1]]$individual_correction$true_score

#> 
#> Heterogeneity results for r 
#> ---------------------------
#> 
#> Accounted for a total of 49.443% of variance
#> 
#> Correlation between r values and artifactual perturbations: 0.703
#> 
#> The reliability of observed effect sizes is: 0.506
#> 
#> 
#> Random effects variance estimates
#> ---------------------------------
#> Hunter-Schmidt method:
#>   sd_rho  (tau):   0.152, SE = 0.049, 95% CI = [0.062, 0.273] 
#>   var_rho (tau^2): 0.023, SE = 0.015, 95% CI = [0.004, 0.075] 
#> 
#>   Q statistic: 38.428 (df = 19, p = 0.005) 
#>   H: 1.422   H^2: 2.023   I^2: 50.557
#> 
#> DerSimonian-Laird method:
#>   sd_rho  (tau):   0.156
#>   var_rho (tau^2): 0.024
#> 
#>   Q statistic: 38.780
#>   H: 1.429   H^2: 2.041   I^2: 51.005
#> 
#> Outlier-robust method (absolute deviation from mean):
#>   sd_rho  (tau_r):   0.172
#>   var_rho (tau_r^2): 0.030
#> 
#>   Q_r statistic: 23.108
#>   H_r: 1.486   H_r^2: 2.207   I_r^2: 0.547
#> 
#> Outlier-robust method (absolute deviation from median):
#>   sd_rho  (tau_m):   0.164
#>   var_rho (tau_m^2): 0.027
#> 
#>   Q_m statistic: 23.168
#>   H_m: 1.452   H_m^2: 2.108   I_m^2: 0.526
#> 

ma_obj$heterogeneity[[1]]$artifact_distribution$true_score

#> 
#> Heterogeneity results for r 
#> ---------------------------
#> 
#> Accounted for a total of 62.658% of variance
#>    Due to sampling error:  57.076%
#>    Due to other artifacts: 5.582%
#> 
#> Correlation between r values and artifactual perturbations: 0.792
#>    Between r values and sampling error values: 0.755
#>    Between r values and other artifact values: 0.236
#> 
#> The reliability of observed effect sizes is: 0.373
#> 
#> 
#> Random effects variance estimates
#> ---------------------------------
#> Hunter-Schmidt method:
#>   sd_rho  (tau):   0.123, SE = 0.052, 95% CI = [0.000, 0.247] 
#>   var_rho (tau^2): 0.015, SE = 0.013, 95% CI = [0.000, 0.061] 
#> 
#>   Q statistic: 30.323 (df = 19, p = 0.048) 
#>   H: 1.263   H^2: 1.596   I^2: 37.342
#> 

## d values
ma_obj <- ma_d_ic(d = d, n1 = n1, n2 = n2, ryy = ryyi,
                  data = data_d_meas_multi)

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

ma_obj <- ma_d_ad(ma_obj)

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

#> Artifact-distribution meta-analyses have been added to 'ma_obj'

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

ma_obj <- heterogeneity(ma_obj = ma_obj)

#> Heterogeneity analyses have been added to 'ma_obj' - use get_heterogeneity() to retrieve them.

ma_obj$heterogeneity[[1]]$barebones

#> 
#> Heterogeneity results for d 
#> ---------------------------
#> 
#> Accounted for a total of 27.053% of variance
#> 
#> Correlation between d values and artifactual perturbations: 0.520
#> 
#> The reliability of observed effect sizes is: 0.729
#> 
#> 
#> Random effects variance estimates
#> ---------------------------------
#> Hunter-Schmidt method:
#>   sd_res  (tau):   0.245, SE = 0.024, 95% CI = [0.203, 0.297] 
#>   var_res (tau^2): 0.060, SE = 0.012, 95% CI = [0.041, 0.088] 
#> 
#>   Q statistic: 365.954 (df = 99, p = 0.000) 
#>   H: 1.923   H^2: 3.697   I^2: 72.947
#> 
#> DerSimonian-Laird method:
#>   sd_res  (tau):   0.279
#>   var_res (tau^2): 0.078
#> 
#>   Q statistic: 471.390
#>   H: 2.182   H^2: 4.762   I^2: 78.998
#> 
#> Outlier-robust method (absolute deviation from mean):
#>   sd_res  (tau_r):   0.271
#>   var_res (tau_r^2): 0.073
#> 
#>   Q_r statistic: 167.748
#>   H_r: 2.113   H_r^2: 4.465   I_r^2: 0.776
#> 
#> Outlier-robust method (absolute deviation from median):
#>   sd_res  (tau_m):   0.273
#>   var_res (tau_m^2): 0.074
#> 
#>   Q_m statistic: 169.618
#>   H_m: 2.126   H_m^2: 4.519   I_m^2: 0.779
#> 

ma_obj$heterogeneity[[1]]$individual_correction$latentGroup_latentY

#> 
#> Heterogeneity results for d 
#> ---------------------------
#> 
#> Accounted for a total of 30.416% of variance
#> 
#> Correlation between d values and artifactual perturbations: 0.552
#> 
#> The reliability of observed effect sizes is: 0.696
#> 
#> 
#> Random effects variance estimates
#> ---------------------------------
#> Hunter-Schmidt method:
#>   sd_delta  (tau):   0.275, SE = 0.028, 95% CI = [0.225, 0.337] 
#>   var_delta (tau^2): 0.075, SE = 0.016, 95% CI = [0.051, 0.113] 
#> 
#>   Q statistic: 325.485 (df = 99, p = 0.000) 
#>   H: 1.813   H^2: 3.288   I^2: 69.584
#> 
#> DerSimonian-Laird method:
#>   sd_delta  (tau):   0.350
#>   var_delta (tau^2): 0.122
#> 
#>   Q statistic: 489.826
#>   H: 2.224   H^2: 4.948   I^2: 79.789
#> 
#> Outlier-robust method (absolute deviation from mean):
#>   sd_delta  (tau_r):   0.329
#>   var_delta (tau_r^2): 0.108
#> 
#>   Q_r statistic: 166.360
#>   H_r: 2.096   H_r^2: 4.391   I_r^2: 0.772
#> 
#> Outlier-robust method (absolute deviation from median):
#>   sd_delta  (tau_m):   0.332
#>   var_delta (tau_m^2): 0.110
#> 
#>   Q_m statistic: 168.498
#>   H_m: 2.112   H_m^2: 4.460   I_m^2: 0.776
#> 

ma_obj$heterogeneity[[1]]$artifact_distribution$latentGroup_latentY

#> 
#> Heterogeneity results for d 
#> ---------------------------
#> 
#> Accounted for a total of 27.402% of variance
#>    Due to sampling error:  27.053%
#>    Due to other artifacts: 0.349%
#> 
#> Correlation between d values and artifactual perturbations: 0.523
#>    Between d values and sampling error values: 0.520
#>    Between d values and other artifact values: 0.059
#> 
#> The reliability of observed effect sizes is: 0.726
#> 
#> 
#> Random effects variance estimates
#> ---------------------------------
#> Hunter-Schmidt method:
#>   sd_delta  (tau):   0.303, SE = 0.029, 95% CI = [0.250, 0.368] 
#>   var_delta (tau^2): 0.092, SE = 0.018, 95% CI = [0.063, 0.136] 
#> 
#>   Q statistic: 361.288 (df = 99, p = 0.000) 
#>   H: 1.910   H^2: 3.649   I^2: 72.598
#>