Compute linear regression models and generate "lm" objects from covariance matrices.

Compute linear regression models and generate "lm" objects from covariance matrices.

lm_mat(
  formula,
  cov_mat,
  mean_vec = rep(0, ncol(cov_mat)),
  n = Inf,
  se_beta_method = c("lm", "normal"),
  ...
)

matrixreg(
  formula,
  cov_mat,
  mean_vec = rep(0, ncol(cov_mat)),
  n = Inf,
  se_beta_method = c("lm", "normal"),
  ...
)

matreg(
  formula,
  cov_mat,
  mean_vec = rep(0, ncol(cov_mat)),
  n = Inf,
  se_beta_method = c("lm", "normal"),
  ...
)

lm_matrix(
  formula,
  cov_mat,
  mean_vec = rep(0, ncol(cov_mat)),
  n = Inf,
  se_beta_method = c("lm", "normal"),
  ...
)

Arguments

formula
Regression formula with a single outcome variable on the left-hand side and one or more predictor variables on the right-hand side (e.g., Y ~ X1 + X2).

cov_mat
Covariance matrix containing the variables to be used in the regression.

mean_vec
Vector of means corresponding to the variables in cov_mat.

n
Sample size to be used in significance testing

se_beta_method
Method to use to estimate the standard errors of standardized regression (beta) coefficients. Current options include "lm" (estimate standard errors using conventional regression formulas) and "normal" (use the Jones-Waller normal-theory approach from the fungible::seBeta() and fungible::seBetaCor() functions)

...
Additional arguments.

Value

An object with the class "lm_mat" that can be used with summary, print, predict, and anova methods.

Examples

## Generate data
S <- reshape_vec2mat(cov = c(.3 * 2 * 3,
                             .4 * 2 * 4,
                             .5 * 3 * 4),
                     var = c(2, 3, 4)^2,
                     var_names = c("X", "Y", "Z"))
mean_vec <- setNames(c(1, 2, 3), colnames(S))
dat <- data.frame(MASS::mvrnorm(n = 100, mu = mean_vec,
                                Sigma = S, empirical = TRUE))

## Compute regression models with lm
lm_out1 <- lm(Y ~ X, data = dat)
lm_out2 <- lm(Y ~ X + Z, data = dat)

## Compute regression models with lm_mat
matreg_out1 <- lm_mat(formula = Y ~ X, cov_mat = S, mean_vec = mean_vec, n = nrow(dat))
matreg_out2 <- lm_mat(formula = Y ~ X + Z, cov_mat = S, mean_vec = mean_vec, n = nrow(dat))

## Compare results of lm and lm_mat with one predictor
lm_out1

#> 
#> Call:
#> lm(formula = Y ~ X, data = dat)
#> 
#> Coefficients:
#> (Intercept)            X  
#>        1.55         0.45  
#> 

matreg_out1

#> 
#> Call:
#> lm_mat(formula = Y ~ X, cov_mat = S, mean_vec = mean_vec, n = nrow(dat))
#> 
#> Coefficients:
#> (Intercept)            X  
#>        1.55         0.45  
#> 

## Compare summaries of lm and lm_mat with one predictor
summary(lm_out1)

#> 
#> Call:
#> lm(formula = Y ~ X, data = dat)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -7.6940 -1.5727  0.2675  1.9246  5.7867 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   1.5500     0.3219   4.815 5.36e-06 ***
#> X             0.4500     0.1445   3.113  0.00243 ** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 2.876 on 98 degrees of freedom
#> Multiple R-squared:   0.09,	Adjusted R-squared:  0.08071 
#> F-statistic: 9.692 on 1 and 98 DF,  p-value: 0.002426
#> 

summary(matreg_out1)

#> 
#> Call:
#> lm_mat(formula = Y ~ X, cov_mat = S, mean_vec = mean_vec, n = nrow(dat))
#> 
#> Residuals:
#>    Min     1Q Median     3Q    Max 
#> -10.70  -1.94   0.00   1.94  10.70 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   1.5500     0.3219   4.815 5.36e-06 ***
#> X             0.4500     0.1445   3.113  0.00243 ** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 2.876 on 98 degrees of freedom
#> Multiple R-squared:   0.09,	Adjusted R-squared:  0.08071 
#> F-statistic: 9.692 on 1 and 98 DF,  p-value: 0.002426
#> 

## Compare results of lm and lm_mat with two predictors
lm_out2

#> 
#> Call:
#> lm(formula = Y ~ X + Z, data = dat)
#> 
#> Coefficients:
#> (Intercept)            X            Z  
#>      0.8036       0.1786       0.3393  
#> 

matreg_out2

#> 
#> Call:
#> lm_mat(formula = Y ~ X + Z, cov_mat = S, mean_vec = mean_vec, 
#>     n = nrow(dat))
#> 
#> Coefficients:
#> (Intercept)            X            Z  
#>      0.8036       0.1786       0.3393  
#> 

## Compare summaries of lm and lm_mat with two predictors
summary(lm_out2)

#> 
#> Call:
#> lm(formula = Y ~ X + Z, data = dat)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -7.3988 -1.2806 -0.0327  1.3721  6.6489 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  0.80357    0.33103   2.427    0.017 *  
#> X            0.17857    0.14277   1.251    0.214    
#> Z            0.33929    0.07138   4.753 6.95e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 2.604 on 97 degrees of freedom
#> Multiple R-squared:  0.2619,	Adjusted R-squared:  0.2467 
#> F-statistic: 17.21 on 2 and 97 DF,  p-value: 4.013e-07
#> 

summary(matreg_out2)

#> 
#> Call:
#> lm_mat(formula = Y ~ X + Z, cov_mat = S, mean_vec = mean_vec, 
#>     n = nrow(dat))
#> 
#> Residuals:
#>    Min     1Q Median     3Q    Max 
#> -9.684 -1.756  0.000  1.756  9.684 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  0.80357    0.33103   2.427    0.017 *  
#> X            0.17857    0.14277   1.251    0.214    
#> Z            0.33929    0.07138   4.753 6.95e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 2.604 on 97 degrees of freedom
#> Multiple R-squared:  0.2619,	Adjusted R-squared:  0.2467 
#> F-statistic: 17.21 on 2 and 97 DF,  p-value: 4.013e-07
#> 

## Compare predictions made with lm and lm_mat
predict(object = matreg_out1, newdata = data.frame(X = 1:5))

#>    1    2    3    4    5 
#> 2.00 2.45 2.90 3.35 3.80 

predict(object = summary(matreg_out1), newdata = data.frame(X = 1:5))

#>    1    2    3    4    5 
#> 2.00 2.45 2.90 3.35 3.80 

predict(lm_out1, newdata = data.frame(X = 1:5))

#>    1    2    3    4    5 
#> 2.00 2.45 2.90 3.35 3.80 

## Compare predictions made with lm and lm_mat (with confidence intervals)
predict(object = matreg_out1, newdata = data.frame(X = 1:5),
        se.fit = TRUE, interval = "confidence")

#> $fit
#>    fit      lwr      upr
#> 1 2.00 1.429191 2.570809
#> 2 2.45 1.811172 3.088828
#> 3 2.90 2.090719 3.709281
#> 4 3.35 2.317368 4.382632
#> 5 3.80 2.518487 5.081513
#> 
#> $se.fit
#>         1         2         3         4         5 
#> 0.2876382 0.3219139 0.4078077 0.5203570 0.6457720 
#> 
#> $df
#> [1] 98
#> 
#> $residual.scale
#> [1] 2.876382
#> 

predict(lm_out1, newdata = data.frame(X = 1:5),
        se.fit = TRUE, interval = "confidence")

#> $fit
#>    fit      lwr      upr
#> 1 2.00 1.429191 2.570809
#> 2 2.45 1.811172 3.088828
#> 3 2.90 2.090719 3.709281
#> 4 3.35 2.317368 4.382632
#> 5 3.80 2.518487 5.081513
#> 
#> $se.fit
#>         1         2         3         4         5 
#> 0.2876382 0.3219139 0.4078077 0.5203570 0.6457720 
#> 
#> $df
#> [1] 98
#> 
#> $residual.scale
#> [1] 2.876382
#> 

## Compare predictions made with lm and lm_mat (with prediction intervals)
predict(object = matreg_out1, newdata = data.frame(X = 1:5),
        se.fit = TRUE, interval = "prediction")

#> $fit
#>    fit       lwr      upr
#> 1 2.00 -3.736555 7.736555
#> 2 2.45 -3.293722 8.193722
#> 3 2.90 -2.865169 8.665169
#> 4 3.35 -2.450739 9.150739
#> 5 3.80 -2.050173 9.650173
#> 
#> $se.fit
#>         1         2         3         4         5 
#> 0.2876382 0.3219139 0.4078077 0.5203570 0.6457720 
#> 
#> $df
#> [1] 98
#> 
#> $residual.scale
#> [1] 2.876382
#> 

predict(lm_out1, newdata = data.frame(X = 1:5),
        se.fit = TRUE, interval = "prediction")

#> $fit
#>    fit       lwr      upr
#> 1 2.00 -3.736555 7.736555
#> 2 2.45 -3.293722 8.193722
#> 3 2.90 -2.865169 8.665169
#> 4 3.35 -2.450739 9.150739
#> 5 3.80 -2.050173 9.650173
#> 
#> $se.fit
#>         1         2         3         4         5 
#> 0.2876382 0.3219139 0.4078077 0.5203570 0.6457720 
#> 
#> $df
#> [1] 98
#> 
#> $residual.scale
#> [1] 2.876382
#> 

## Compare model comparisons computed using lm and lm_mat objects
anova(lm_out1, lm_out2)

#> Analysis of Variance Table
#> 
#> Model 1: Y ~ X
#> Model 2: Y ~ X + Z
#>   Res.Df    RSS Df Sum of Sq      F    Pr(>F)    
#> 1     98 810.81                                  
#> 2     97 657.64  1    153.17 22.592 6.952e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

anova(matreg_out1, matreg_out2)

#> Analysis of Variance Table
#> 
#> Model 1: Y ~ X
#> Model 2: Y ~ X + Z
#> 
#>   Res.Df    RSS Df Sum of Sq      F    Pr(>F)    
#> 1     98 810.81  0      0.00                     
#> 2     97 657.64  1    153.17 22.592 6.952e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

## Model comparisons can be run on lm_mat summaries, too:
anova(summary(matreg_out1), summary(matreg_out2))

#> Analysis of Variance Table
#> 
#> Model 1: Y ~ X
#> Model 2: Y ~ X + Z
#> 
#>   Res.Df    RSS Df Sum of Sq      F    Pr(>F)    
#> 1     98 810.81  0      0.00                     
#> 2     97 657.64  1    153.17 22.592 6.952e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

## Or summaries and raw models can be mixed:
anova(matreg_out1, summary(matreg_out2))

#> Analysis of Variance Table
#> 
#> Model 1: Y ~ X
#> Model 2: Y ~ X + Z
#> 
#>   Res.Df    RSS Df Sum of Sq      F    Pr(>F)    
#> 1     98 810.81  0      0.00                     
#> 2     97 657.64  1    153.17 22.592 6.952e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

anova(summary(matreg_out1), matreg_out2)

#> Analysis of Variance Table
#> 
#> Model 1: Y ~ X
#> Model 2: Y ~ X + Z
#> 
#>   Res.Df    RSS Df Sum of Sq      F    Pr(>F)    
#> 1     98 810.81  0      0.00                     
#> 2     97 657.64  1    153.17 22.592 6.952e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

## Compare confidence intervals computed using lm and lm_mat objects
confint(object = lm_out1)

#>                 2.5 %    97.5 %
#> (Intercept) 0.9111723 2.1888277
#> X           0.1631579 0.7368421

confint(object = matreg_out1)

#>                 2.5 %    97.5 %
#> (Intercept) 0.9111723 2.1888277
#> X           0.1631579 0.7368421

confint(object = summary(matreg_out1))

#>                 2.5 %    97.5 %
#> (Intercept) 0.9111723 2.1888277
#> X           0.1631579 0.7368421

confint(object = lm_out2)

#>                  2.5 %    97.5 %
#> (Intercept)  0.1465679 1.4605749
#> X           -0.1047777 0.4619206
#> Z            0.1976111 0.4809603

confint(object = matreg_out2)

#>                  2.5 %    97.5 %
#> (Intercept)  0.1466526 1.4604903
#> X           -0.1047412 0.4618841
#> Z            0.1976294 0.4809420

confint(object = summary(matreg_out2))

#>                  2.5 %    97.5 %
#> (Intercept)  0.1466526 1.4604903
#> X           -0.1047412 0.4618841
#> Z            0.1976294 0.4809420