Bare-bones meta-analysis of generic effect sizes

This function computes bare-bones meta-analyses of any effect size using user-supplied effect error variances.

ma_generic(
  es,
  n,
  var_e,
  sample_id = NULL,
  citekey = NULL,
  construct_x = NULL,
  construct_y = NULL,
  group1 = NULL,
  group2 = NULL,
  wt_type = c("sample_size", "inv_var", "DL", "HE", "HS", "SJ", "ML", "REML", "EB",
    "PM"),
  moderators = NULL,
  cat_moderators = TRUE,
  moderator_type = c("simple", "hierarchical", "none"),
  data = NULL,
  control = control_psychmeta(),
  weights = NULL,
  ...
)

Arguments

es
Vector or column name of observed effect sizes.

n
Vector or column name of sample sizes.

var_e
Vector or column name of error variances.

sample_id
Optional vector of identification labels for samples/studies in the meta-analysis.

citekey
Optional vector of bibliographic citation keys for samples/studies in the meta-analysis (if multiple citekeys pertain to a given effect size, combine them into a single string entry with comma delimiters (e.g., "citkey1,citekey2"). When TRUE, program will use sample-size weights, error variances estimated from the mean effect size, maximum likelihood variances, and normal-distribution confidence and credibility intervals.

construct_x, construct_y
Vector of construct names for constructs designated as "X" and as "Y".

group1, group2
Vector of groups' names associated with effect sizes that represent pairwise contrasts.

wt_type
Type of weight to use in the meta-analysis: native options are "sample_size" and "inv_var" (inverse error variance). Supported options borrowed from metafor are "DL", "HE", "HS", "SJ", "ML", "REML", "EB", and "PM" (see metafor documentation for details about the metafor methods).

moderators
Matrix of moderator variables to be used in the meta-analysis (can be a vector in the case of one moderator).

cat_moderators
Logical scalar or vector identifying whether variables in the moderators argument are categorical variables (TRUE) or continuous variables (FALSE).

moderator_type
Type of moderator analysis ("none", "simple", or "hierarchical").

data
Data frame containing columns whose names may be provided as arguments to vector arguments and/or moderators.

control
Output from the control_psychmeta() function or a list of arguments controlled by the control_psychmeta() function. Ellipsis arguments will be screened for internal inclusion in control.

weights
Optional vector of weights to be used. When weights is non-NULL, these weights override the argument supplied to wt_type.

...
Further arguments to be passed to functions called within the meta-analysis.

Value

A nested tabular object of the class "ma_psychmeta".

Examples

es <- c(.3, .5, .8)
n <- c(100, 200, 150)
var_e <- 1 / n
ma_obj <- ma_generic(es = es, n = n, var_e = var_e)
ma_obj

#> Overview tibble of psychmeta meta-analysis of generic effect sizes  
#> ---------------------------------------------------------------------- 
#> # A tibble: 1 x 4
#>   analysis_id analysis_type meta_tables      escalc          
#> *       <int> <chr>         <named list>     <named list>    
#> 1           1 Overall       <named list [1]> <named list [2]>
#> 
#> To extract results, try summary() or the get_stuff functions (run ?get_stuff for help). 

summary(ma_obj)

#> Bare-bones meta-analysis results 
#> ---------------------------------------------------------------------- 
#>   analysis_id analysis_type k   N mean_es sd_es se_es sd_res CI_LL_95 CI_UL_95
#> 1           1       Overall 3 450   0.556 0.232 0.134  0.217  -0.0201     1.13
#>   CR_LL_80 CR_UL_80
#> 1    0.147    0.965
#> 
#> 
#> Information available in the meta-analysis object includes:
#>  - meta_tables   [ access using get_metatab() ]
#>  - escalc        [ access using get_escalc() ]