Second-order meta-analysis function for correlations

This function computes second-order meta-analysis function for correlations. It supports second-order analyses of bare-bones, artifact-distribution, and individual-correction meta-analyses.

ma_r_order2(
  k,
  N = NULL,
  r = NULL,
  rho = NULL,
  var_r = NULL,
  var_r_c = NULL,
  ma_type = c("bb", "ic", "ad"),
  sample_id = NULL,
  citekey = NULL,
  moderators = NULL,
  moderator_type = "simple",
  construct_x = NULL,
  construct_y = NULL,
  construct_order = NULL,
  data = NULL,
  control = control_psychmeta(),
  ...
)

Arguments

k
Vector or column name of meta-analyses' k values.

N
Vector or column name of meta-analyses' total sample sizes (optional).

r
Vector or column name of mean observed correlations.

rho
Vector or column name of mean corrected correlations.

var_r
Vector or column name of observed variances of observed correlations.

var_r_c
Vector or column name of observed variances of corrected correlations.

ma_type
Type of meta-analyses being analyzed: "bb" (barebones), "ic" (individual correction), or "ad" (artifact distribution).

sample_id
Vector or column name of study ID labels.

citekey
Optional vector of bibliographic citation keys for samples/studies in the meta-analysis (if multiple citekeys pertain to a given effect size, combine them into a single string entry with comma delimiters (e.g., "citkey1,citekey2").

moderators
Matrix or column names of moderator variables to be used in the meta-analysis (can be a vector in the case of one moderator).

moderator_type
Type of moderator analysis ("none", "simple", or "hierarchical").

construct_x
Vector or column name of construct names for X.

construct_y
Vector or column name of construct names for Y.

construct_order
Vector indicating the order in which variables should be arranged, with variables listed earlier in the vector being preferred for designation as X.

data
Data frame containing columns whose names may be provided as arguments to vector arguments and/or moderators.

control
Output from the control_psychmeta() function or a list of arguments controlled by the control_psychmeta() function. Ellipsis arguments will be screened for internal inclusion in control.

...
Further arguments to be passed to functions called within the meta-analysis.

Value

A nested tabular object of the class "ma_psychmeta".

Examples

## Analysis of the validity of conscientiousness as a predictor of job performance in East Asia
out <- ma_r_order2(k = k, r = r_bar_i, rho = rho_bar_i, var_r = var_r,
                   var_r_c = NULL, ma_type = c("bb", "ad"),
                   sample_id = NULL, moderators = NULL,
                   construct_x = NULL, construct_y = NULL,
                   data = dplyr::filter(data_r_oh_2009, Predictor == "Conscientiousness"))
summary(out)

#> 
#> Second-order artifact-distribution meta-analysis results 
#> ---------------------------------------------------------------------- 
#>   analysis_id analysis_type L  k  N mean_rho_bar sd_rho_bar se_rho_bar
#> 1           1       Overall 4 24 NA        0.208     0.0369    0.00753
#>   sd_rho_bar_res CI_LL_95 CI_UL_95 CR_LL_80 CR_UL_80
#> 1        0.00879    0.192    0.223    0.196    0.219
#> 
#> 
#> Information available in the meta-analysis object includes:
#>  - meta_tables   [ access using get_metatab() ]
#>  - escalc        [ access using get_escalc() ]

## Analysis of the validity of the Big Five traits as predictors of job performance in East Asia
out <- ma_r_order2(k = k, r = r_bar_i, rho = rho_bar_i, var_r = var_r,
                   var_r_c = NULL, ma_type = c("bb", "ad"),
                   sample_id = NULL, moderators = NULL, construct_x = Predictor,
                   data = data_r_oh_2009)
summary(out)

#> 
#> Second-order artifact-distribution meta-analysis results 
#> ---------------------------------------------------------------------- 
#>   analysis_id         construct_x analysis_type L  k  N mean_rho_bar sd_rho_bar
#> 1           1       Agreeableness       Overall 4 22 NA       0.0981     0.0785
#> 2           2   Conscientiousness       Overall 4 24 NA       0.2077     0.0369
#> 3           3 Emotional Stability       Overall 4 23 NA       0.1610     0.0926
#> 4           4        Extraversion       Overall 4 24 NA       0.1764     0.1197
#> 5           5            Openness       Overall 4 22 NA       0.1323     0.1179
#>   se_rho_bar sd_rho_bar_res CI_LL_95 CI_UL_95 CR_LL_80 CR_UL_80
#> 1    0.01673        0.06996   0.0633    0.133  0.00555    0.191
#> 2    0.00753        0.00879   0.1921    0.223  0.19612    0.219
#> 3    0.01931        0.08055   0.1210    0.201  0.05460    0.267
#> 4    0.02443        0.11191   0.1259    0.227  0.02876    0.324
#> 5    0.02514        0.10499   0.0800    0.185 -0.00659    0.271
#> 
#> 
#> Information available in the meta-analysis object includes:
#>  - meta_tables   [ access using get_metatab() ]
#>  - escalc        [ access using get_escalc() ]

## Analysis of the average validity of the Big Five traits as predictors of
## job performance by Eastern Asian country
out <- ma_r_order2(k = k, r = r_bar_i, rho = rho_bar_i, var_r = var_r,
                   var_r_c = NULL, ma_type = c("bb", "ad"),
                   sample_id = NULL, moderators = "Country", data = data_r_oh_2009)
summary(out)

#> 
#> Second-order artifact-distribution meta-analysis results 
#> ---------------------------------------------------------------------- 
#>   analysis_id    analysis_type    Country  L   k  N mean_rho_bar sd_rho_bar
#> 1           1          Overall All Levels 20 115 NA       0.1560     0.0891
#> 2           2 Simple Moderator      China  5  15 NA       0.2295     0.0308
#> 3           3 Simple Moderator      Korea  5  69 NA       0.0992     0.0617
#> 4           4 Simple Moderator  Singapore  5  18 NA       0.2016     0.0731
#> 5           5 Simple Moderator     Taiwan  5  13 NA       0.2683     0.1283
#>   se_rho_bar sd_rho_bar_res CI_LL_95 CI_UL_95 CR_LL_80 CR_UL_80
#> 1    0.00831         0.0791   0.1395    0.172   0.0541    0.258
#> 2    0.00795         0.0000   0.2125    0.247   0.2295    0.230
#> 3    0.00743         0.0544   0.0844    0.114   0.0289    0.170
#> 4    0.01723         0.0656   0.1653    0.238   0.1141    0.289
#> 5    0.03558         0.0991   0.1908    0.346   0.1339    0.403
#> 
#> 
#> Information available in the meta-analysis object includes:
#>  - meta_tables   [ access using get_metatab() ]
#>  - escalc        [ access using get_escalc() ]