Compute meta-regressions

This function is a wrapper for metafor's rma function that computes meta-regressions for all bare-bones and individual-correction meta-analyses within an object. It makes use of both categorical and continuous moderator information stored in the meta-analysis object and allows for interaction effects to be included in the regression model. Output from this function will be added to the meta-analysis object in a list called follow_up_analyses. If using this function with a multi-construct meta-analysis object from ma_r or ma_d, note that the follow_up_analyses list is appended to the meta-analysis object belonging to a specific construct pair within the construct_pairs list.

metareg(ma_obj, formula_list = NULL, ...)

Arguments

ma_obj
Meta-analysis object.

formula_list
Optional list of regression formulas to evaluate. NOTE: If there are spaces in your moderator names, replace them with underscores (i.e., "_") so that the formula(s) will perform properly. The function will remove spaces in the data, you only have to account for this in formula_list when you supply your own formula(s).

...
Additional arguments.

Value

ma_obj with meta-regression results added (see ma_obj$follow_up_analyses$metareg).

Examples

## Meta-analyze the data from Gonzalez-Mule et al. (2014)
## Note: These are corrected data and we have confirmed with the author that
## these results are accurate:
ma_obj <- ma_r_ic(rxyi = rxyi, n = n, hs_override = TRUE, data = data_r_gonzalezmule_2014,
                  rxx = rxxi, ryy = ryyi, ux = ux, indirect_rr_x = TRUE,
                  correct_rr_x = TRUE, moderators = Complexity)

## Pass the meta-analysis object to the meta-regression function:
ma_obj <- metareg(ma_obj)

#> Meta-regressions have been added to 'ma_obj' - use get_metareg() to retrieve them.

## Examine the meta-regression results for the bare-bones and corrected data:
ma_obj$metareg[[1]]$barebones$`Main Effects`

#> 
#> Mixed-Effects Model (k = 43; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0098 (SE = 0.0033)
#> tau (square root of estimated tau^2 value):             0.0989
#> I^2 (residual heterogeneity / unaccounted variability): 73.48%
#> H^2 (unaccounted variability / sampling variability):   3.77
#> R^2 (amount of heterogeneity accounted for):            0.00%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 40) = 129.4707, p-val < .0001
#> 
#> Test of Moderators (coefficients 2:3):
#> QM(df = 2) = 1.9904, p-val = 0.3696
#> 
#> Model Results:
#> 
#>                   estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt             0.1736  0.0285   6.0812  <.0001   0.1176  0.2295  *** 
#> ComplexityMedium   -0.0356  0.0391  -0.9105  0.3626  -0.1123  0.0410      
#> ComplexityHigh     -0.1068  0.0842  -1.2695  0.2043  -0.2718  0.0581      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

ma_obj$metareg[[1]]$individual_correction$true_score$`Main Effects`

#> 
#> Mixed-Effects Model (k = 43; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0343 (SE = 0.0111)
#> tau (square root of estimated tau^2 value):             0.1852
#> I^2 (residual heterogeneity / unaccounted variability): 76.33%
#> H^2 (unaccounted variability / sampling variability):   4.22
#> R^2 (amount of heterogeneity accounted for):            0.00%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 40) = 153.8340, p-val < .0001
#> 
#> Test of Moderators (coefficients 2:3):
#> QM(df = 2) = 1.9983, p-val = 0.3682
#> 
#> Model Results:
#> 
#>                   estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt             0.3340  0.0519   6.4343  <.0001   0.2323  0.4358  *** 
#> ComplexityMedium   -0.0717  0.0714  -1.0042  0.3153  -0.2115  0.0682      
#> ComplexityHigh     -0.1874  0.1549  -1.2099  0.2263  -0.4909  0.1162      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

## Meta-analyze simulated d-value data
dat <- data_d_meas_multi
## Simulate a random moderator
set.seed(100)
dat$moderator <- sample(1:2, nrow(dat), replace = TRUE)
ma_obj <- ma_d(ma_method = "ic", d = d, n1 = n1, n2 = n2, ryy = ryyi,
               construct_y = construct, sample_id = sample_id,
               moderators = moderator, data = dat)

#>  **** Running ma_d: Meta-analysis of d values **** 

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

#> Warning: You have removed the following critical columns: analysis_id, analysis_type, meta_tables, escalc

## Pass the meta-analysis object to the meta-regression function:
ma_obj <- metareg(ma_obj)

#> Meta-regressions have been added to 'ma_obj' - use get_metareg() to retrieve them.

## Examine the meta-regression results for the bare-bones and corrected data:
ma_obj$metareg[[1]]$barebones$`Main Effects`

#> 
#> Mixed-Effects Model (k = 50; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0325 (SE = 0.0113)
#> tau (square root of estimated tau^2 value):             0.1804
#> I^2 (residual heterogeneity / unaccounted variability): 59.74%
#> H^2 (unaccounted variability / sampling variability):   2.48
#> R^2 (amount of heterogeneity accounted for):            6.90%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 48) = 118.8375, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 2.9297, p-val = 0.0870
#> 
#> Model Results:
#> 
#>             estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt       0.9087  0.0507  17.9410  <.0001   0.8095  1.0080  *** 
#> moderator2   -0.1155  0.0675  -1.7116  0.0870  -0.2478  0.0168    . 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

ma_obj$metareg[[1]]$individual_correction$latentGroup_latentY$`Main Effects`

#> 
#> Mixed-Effects Model (k = 50; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0453 (SE = 0.0167)
#> tau (square root of estimated tau^2 value):             0.2128
#> I^2 (residual heterogeneity / unaccounted variability): 56.50%
#> H^2 (unaccounted variability / sampling variability):   2.30
#> R^2 (amount of heterogeneity accounted for):            8.13%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 48) = 109.7234, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 3.2015, p-val = 0.0736
#> 
#> Model Results:
#> 
#>             estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt       1.1085  0.0615  18.0360  <.0001   0.9880  1.2290  *** 
#> moderator2   -0.1467  0.0820  -1.7893  0.0736  -0.3075  0.0140    . 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>