Retrieve the NEWS file for the psychmeta package

This function gives a shortcut to the utils::news(package = "psychmeta") function and displays psychmeta's NEWS file, which contains version information, outlines additions and changes to the package, and describes other updates.

psychmeta_news()

Examples

psychmeta_news()

#> Version: 2.4.0
#> Date: 2020-07-11
#> Text: Bug fix in convert_es().
#> 
#> Version: 2.4.0
#> Date: 2020-07-11
#> Text: convert_es() now returns a simple frame with converted effect
#>         sizes, variances, and (effective) sample sizes.
#> 
#> Version: 2.4.0
#> Date: 2020-07-11
#> Text: Bug fix in moderator reconciliation.
#> 
#> Version: 2.4.0
#> Date: 2020-07-11
#> Text: Added robustness to variable selection and second-order
#>         meta-analysis functions.
#> 
#> Version: 2.3.10
#> Date: 2020-06-07
#> Text: Bug fix when compositing facets into overall construct variables.
#> 
#> Version: 2.3.10
#> Date: 2020-06-07
#> Text: Bug fix in ma_r_ic.
#> 
#> Version: 2.3.10
#> Date: 2020-06-07
#> Text: Moved packages `MASS`, `tmvtnorm`, and `nor1mix` to "Suggests" to
#>         avoid errors some users have encountered when loading
#>         `psychmeta` in R 4.0.0.
#> 
#> Version: 2.3.9
#> Date: 2020-06-04
#> Text: Bug fix when compositing facets into overall construct variables.
#> 
#> Version: 2.3.9
#> Date: 2020-06-04
#> Text: Remove unnecessary S4 class exports.
#> 
#> Version: 2.3.9
#> Date: 2020-06-04
#> Text: Bug fix in get_matrix().
#> 
#> Version: 2.3.9
#> Date: 2020-06-04
#> Text: Fix extraneous warnings in ma_r() when ma_method="ic" since dplyr
#>         1.0.0.
#> 
#> Version: 2.3.9
#> Date: 2020-06-04
#> Text: Print all columns of meta-analysis results tables.
#> 
#> Version: 2.3.8
#> Date: 2020-04-21
#> Text: Bug fix to correct an issue that arose in version 2.3.7 in which
#>         factor-type variables passed to ma_r() and ma_d() as
#>         construct/facet/measure arguments were coerced to numeric
#>         variables.
#> 
#> Version: 2.3.7
#> Date: 2020-04-09
#> Text: Updates to ensure consistency with upcoming R version 4.0.0.
#> 
#> Version: 2.3.7
#> Date: 2020-04-09
#> Text: Bug fixes in ma_r() moderated meta-analyses that involve
#>         dependency consolidation.
#> 
#> Version: 2.3.6
#> Date: 2020-02-19
#> Text: Bug fix in dependent sample consolidation.
#> 
#> Version: 2.3.6
#> Date: 2020-02-19
#> Text: A vector of moderator variable names can now be specified quoted
#>         or unquoted, like other arguments.
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Improved print methods for tibble-like output.
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Improved small-sample bias correction for d values; new
#>         correction uses the gamma function, consistent with metafor.
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Replaced deprecated functions from dplyr.
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Made stability improvements to compute_dmod().
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Added moderator information to the output from get_escalc().
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Added var_error_R() and var_error_Rsq() functions for the error
#>         variance of multiple correlations from multiple regression.
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Updated the URL used version checking to avoid error messages
#>         experienced by a subset of users.
#> 
#> Version: 2.3.5
#> Date: 2020-02-13
#> Text: Additional bug fixes in convert_es(), create_ad_int(),
#>         anova.ma_psychmeta(), metabulate(), generate_bib(); typo
#>         corrections in documentation.
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: Support for contour-enhanced funnel plots added to plot_funnel()
#>         and a new plot_cefp() function.
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: Bug fixes for uncertainty intervals in second-order
#>         meta-analyses.
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: New anova() method for comparing subgroup moderator levels.
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: Abbreviation for credibility interval changed to "CR" instead of
#>         "CV".
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: Support for construct_order argument added to ma_r_order2().
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: Added an error message for impossible composites. An error
#>         message is now displayed when a sample's composite effect size
#>         is statistically impossible. Prior to this update, impossible
#>         effect sizes produced generic errors that did not provide users
#>         with useful feedback on the source of the problem.
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: Updated the URL used version checking to avoid error messages
#>         experienced by a subset of users.
#> 
#> Version: 2.3.4
#> Date: 2019-12-18
#> Text: Additional bug fixes in get_matrix(), reshape_wide2long(), and
#>         ma_r_order2().
#> 
#> Version: 2.3.3
#> Date: 2019-04-11
#> Text: Bug fix for the metabulate() function's use of R Markdown.
#> 
#> Version: 2.3.3
#> Date: 2019-04-11
#> Text: Bug fix for retrieving artifact distributions with the get_ad()
#>         function.
#> 
#> Version: 2.3.3
#> Date: 2019-04-11
#> Text: Bug fix to accommodate truly missing artifacts in the
#>         artifact-imputation procedure.
#> 
#> Version: 2.3.3
#> Date: 2019-04-11
#> Text: Bug fix for leave-one-out and cumulative meta-analyses for d
#>         values.
#> 
#> Version: 2.3.3
#> Date: 2019-04-11
#> Text: Update to sample from beta distributions with SD = 0 in
#>         simulate_r_database and simulate_d_database.
#> 
#> Version: 2.3.2
#> Date: 2019-02-26
#> Text: Bug fix for using categorical moderators in meta-regressions.
#> 
#> Version: 2.3.2
#> Date: 2019-02-26
#> Text: Bug fix for using moderators in ma_generic() when there are no
#>         construct or group-identification variables in the analysis.
#> 
#> Version: 2.3.2
#> Date: 2019-02-26
#> Text: Added an optional "weights" argument to ma_generic(), which
#>         allows users to supply their own custom weights for effect
#>         sizes.
#> 
#> Version: 2.3.2
#> Date: 2019-02-26
#> Text: Added alternative methods for calculating confidence limits for
#>         tau-squared.
#> 
#> Version: 2.3.1
#> Date: 2019-02-15
#> Text: Increased support for heterogeneity analyses conducted using
#>         weights imported from the metafor package.
#> 
#> Version: 2.3.1
#> Date: 2019-02-15
#> Text: Updated estimates for small-sample-bias corrected error variances
#>         to be computed from the bias-corrected effect size rather than
#>         applying a post-hoc linear variance adjustment.
#> 
#> Version: 2.3.1
#> Date: 2019-02-15
#> Text: Changed default weighting method for d values to "n_effective."
#>         This method is superior to sample-size weights because it
#>         accounts for both total sample size and deviations from equal
#>         subgroup representation.
#> 
#> Version: 2.3.1
#> Date: 2019-02-15
#> Text: Updated compatibility with dplyr version 0.8.0.
#> 
#> Version: 2.3.1
#> Date: 2019-02-15
#> Text: Updated the labeling of output from the mix_dist function.
#> 
#> Version: 2.3.1
#> Date: 2019-02-15
#> Text: Assorted bug fixes affecting mean d value estimates (from the
#>         ma_d and ma_d_ic functions), artifact distributions, and
#>         weighted correlations and covariances.
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Individual-correction validity generalization meta-analyses have
#>         been corrected to remove variation due to between-sample
#>         reliability differences prior to adding back measurement error.
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Improved compatibility with dplyr and tibble packages.
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Bug fixes for `var_error_d` and `var_error_g` to accommodate
#>         inputs of differing lengths.
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Bug fix for cumulative meta-analyses.
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Bug fix for handling null interactive ADs
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Bug fix for meta-analyzing correlations between global constructs
#>         and facets.
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Bug fix for estimating correction-implied subgroup proportions in
#>         correct_d.
#> 
#> Version: 2.3.0
#> Date: 2019-01-08
#> Text: Bug fix for using correct_d when n2 is left unspecified.
#> 
#> Version: 2.2.1
#> Date: 2018-10-12
#> Text: Added option to suppress progress indicators in the console. Use
#>         options(psychmeta.show_progress = FALSE) to suppress progress
#>         indicators.
#> 
#> Version: 2.2.1
#> Date: 2018-10-12
#> Text: Increased support for common wide database formats in the
#>         reshape_wide2long() function.
#> 
#> Version: 2.2.1
#> Date: 2018-10-12
#> Text: Bug fixes for managing dependent observations in artifact
#>         distributions created within ma_r() and ma_d().
#> 
#> Version: 2.2.1
#> Date: 2018-10-12
#> Text: Bug fixes for resolving dependency in effect sizes in
#>         meta-analyses with moderators.
#> 
#> Version: 2.2.1
#> Date: 2018-10-12
#> Text: Added error-variance estimates to the output of convert_es() for
#>         increased compatibility with the ma_generic() function.
#> 
#> Version: 2.2.0
#> Date: 2018-09-18
#> Text: Added support for separately compositing measures of constructs
#>         and the constructs' facets. Arguments for 'facet_x' and
#>         'facet_y' have been added to ma_r() and ma_d() so that more
#>         comprehensive output tables can be generated.
#> 
#> Version: 2.2.0
#> Date: 2018-09-18
#> Text: Stability fixes for data management within the ma_r() function.
#> 
#> Version: 2.2.0
#> Date: 2018-09-18
#> Text: Added checks for missing and infinite values to the ma_generic()
#>         function.
#> 
#> Version: 2.1.12
#> Date: 2018-09-07
#> Text: Bug fix for managing reliability types that have been
#>         incidentally converted from characters into factors.
#> 
#> Version: 2.1.11
#> Date: 2018-08-30
#> Text: Bug fix for create_ad() with interactive method when all sample
#>         sizes are NA.
#> 
#> Version: 2.1.11
#> Date: 2018-08-30
#> Text: Bug fix for ma_r_ad() with individual artifact distributions
#>         supplied.
#> 
#> Version: 2.1.10
#> Date: 2018-08-24
#> Text: Bug fix for using output of metabulate() in Rmarkdown documents.
#> 
#> Version: 2.1.10
#> Date: 2018-08-24
#> Text: Bug fix for meta-analyses of effect sizes that have been averaged
#>         or composited.
#> 
#> Version: 2.1.9
#> Date: 2018-08-22
#> Text: Increased the stability of artifact-distribution meta-analyses
#>         with moderators computed using the ma_r() function.
#> 
#> Version: 2.1.9
#> Date: 2018-08-22
#> Text: Bug fix for the get_ad() function to appropriately handle
#>         artifact-distribution objects that contain no artifact
#>         information.
#> 
#> Version: 2.1.8
#> Date: 2018-08-21
#> Text: Updated and corrected the text displayed in table notes produced
#>         by the metabulate() function.
#> 
#> Version: 2.1.8
#> Date: 2018-08-21
#> Text: Bug fix for handling meta-analyses with only one effect size.
#> 
#> Version: 2.1.7
#> Date: 2018-07-16
#> Text: Removed package 'fungible' as a dependency, as 'fungible' was
#>         archived and removed from CRAN on 2018-07-15.
#> 
#> Version: 2.1.7
#> Date: 2018-07-16
#> Text: Bug fixes and improvements to format_num() and metabulate() for
#>         handling leading zeros.
#> 
#> Version: 2.1.6
#> Date: 2018-07-12
#> Text: Bug fix for ma_generic() analyses that include construct names
#>         and/or group names, but no moderators.
#> 
#> Version: 2.1.5
#> Date: 2018-07-09
#> Text: Bug fix for printing the residual summary from lm_mat so that
#>         console output will show a properly scaled vector of
#>         theoretical residual quantiles.
#> 
#> Version: 2.1.4
#> Date: 2018-06-28
#> Text: Bug fix for defining the p values of negative coefficients in
#>         lm_mat.
#> 
#> Version: 2.1.3
#> Date: 2018-06-27
#> Text: Changed unicode entities to HTML entities for increased support
#>         across locales.
#> 
#> Version: 2.1.3
#> Date: 2018-06-27
#> Text: Bug fix for artifact imputation without a moderator.
#> 
#> Version: 2.1.2
#> Date: 2018-06-26
#> Text: Added grouping variables and construct variables to ma_generic().
#>         This helps to streamline meta-analyses of generic effect sizes
#>         that involve multiple group-wise contrasts and/or multiple
#>         constructs.
#> 
#> Version: 2.1.2
#> Date: 2018-06-26
#> Text: Bug fixes for format_num() and metabulate() to allow missing
#>         values in formatted numeric strings.
#> 
#> Version: 2.1.2
#> Date: 2018-06-26
#> Text: Updated lm_mat()to use fungible::seBetaCor to estimate standard
#>         errors when se_beta_method is "normal" and cov_mat is a
#>         correlation matrix.
#> 
#> Version: 2.1.1
#> Date: 2018-06-10
#> Text: Updates and bug fixes for the metabulate() function. Fixed issues
#>         with file-writing protocols. Increased support for unicode
#>         characters on Windows systems. Added arguments to control
#>         whether table headers are rendered in bolded font
#>         (bold_headers) and whether redundant construct-pair labels are
#>         collapsed in meta-analysis tables (collapse_construct_labels).
#> 
#> Version: 2.1.1
#> Date: 2018-06-10
#> Text: Updated the “summary” attributes of artifact-distribution objects
#>         for greater consistency between interactive and Taylor-series
#>         objects.
#> 
#> Version: 2.1.1
#> Date: 2018-06-10
#> Text: Added third-order sample sizes (i.e., the number of meta-analyses
#>         meta-analyzed; denoted as “L”) to the output from second-order
#>         meta-analyses.
#> 
#> Version: 2.1.1
#> Date: 2018-06-10
#> Text: Added get_stuff() function that serves as a wrapper for other
#>         functions in the “get_stuff” family of functions.
#> 
#> Version: 2.1.0
#> Date: 2018-06-01
#> Text: Updated "metabulate" function based on RMarkdown. The metabulate
#>         function can now output to Word, PDF, HTML, Markdown, ODT, and
#>         plain text and accepts arguments to include a bibliography in
#>         the same output as results tables. Output can be rendered
#>         directly into the selected format or placed into a larger
#>         RMarkdown document. The function also includes improvements to
#>         the typesetting of numeric results and column headings (e.g., ρ̅
#>         instead of Mean<sub>ρ</sub>).
#> 
#> Version: 2.1.0
#> Date: 2018-06-01
#> Text: New "num_format" function to format numbers to a specified number
#>         of digits and allowing control of the characters used for
#>         decimal points, thousands and decimal digit group separators,
#>         leading zeros, and positive and negative signs. It's useful to
#>         avoid unpleasant surprises during publisher typesetting of
#>         tables.
#> 
#> Version: 2.0.2
#> Date: 2018-05-30
#> Category: Hotfix for handling NULL moderator objects in dependency
#>         resolution procedure
#> Text:
#> 
#> Version: 2.0.1
#> Date: 2018-05-29
#> Category: Hotfix for missing object in the dependency resolution
#>         procedure when no moderators are present
#> Text:
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: Major overhauls to the format of meta-analysis objects to improve
#>         the user experience. The outputs of meta-analysis functions are
#>         now nested tibbles rather than nested lists. This provides
#>         users with a clearer overview of the structure of the object,
#>         streamlines the manipulation of results objects (e.g.,
#>         selecting/subsetting columns, filtering, and arranging), and
#>         makes the objects more flexible in terms of performing follow
#>         up analyses and performing artifact-distribution corrections
#>         (because AD corrections can now by done separately by moderator
#>         level). As the print methods for meta-analyses no longer report
#>         grand tables of results, we recommend using summary() or
#>         get_metatab() to obtain this information.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: Arguments for specifying methodological parameters (e.g.,
#>         conf_method, conf_level, var_unbiased) in meta-analyses have
#>         been migrated to a new control_psychmeta() function - the
#>         output of this function can be passed to the “control” argument
#>         of meta-analysis functions to select the methodological
#>         parameters. This is meant to reduce the number of arguments
#>         native to meta-analysis functions and simplify the
#>         documentation of psychmeta’s core functions. All of the
#>         arguments to control_psychmeta can still be passed directly to
#>         meta-analysis functions and all function calls accepted by
#>         previous versions of psychmeta should continue to work in this
#>         version of the package.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: The dependency resolution process in ma_r() and ma_d() in which
#>         dependent effect sizes are automatically combined into
#>         composites has been enhanced. The “intercor” argument now
#>         accepts a variety of types of information. The user can now set
#>         a default scalar value to use as well as supply a named vector
#>         of values (the names can refer to construct names and/or
#>         sample_id and construct pairings) and/or a database from which
#>         to harvest intercorrelation information. This allows the
#>         compositing procedure to use sample-specific information to
#>         compute composites when that information is available.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: “summary” methods have been added for classes exported from
#>         psychmeta.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: The overview vignette has been updated to provide instructions on
#>         how to navigate the output objects from meta-analyses.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: The class structure of meta-analysis objects has been simplified.
#>         There is now only one class for meta-analyses ("ma_psychmeta").
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: Data management has been streamlined within the ma_r() and ma_d()
#>         functions to permit more efficient cleaning and imputation of
#>         artifact information.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: To complement the new structure of meta-analysis objects, the
#>         create_ad_list function now produces tibbles of artifact
#>         distributions and the function is now an alias of
#>         create_ad_tibbles. This function can now create artifact
#>         distributions for construct pairs and/or specific moderator
#>         combinations.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: The behavior of the supplemental_ads argument to meta-analysis
#>         functions has been expanded to allow artifact information to be
#>         supplied as individual artifact distributions, lists of
#>         artifact distributions, or tibbles of artifact distributions.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: "print" methods have been updated for a variety of object
#>         classes.
#> 
#> Version: 2.0.0
#> Date: 2018-05-28
#> Text: Bug fixes and stability improvements have been made to follow-up
#>         analyses, plotting functions, dependency resolution,
#>         convert_ma(), and dMod().
#> 
#> Version: 1.0.3
#> Date: 2018-05-08
#> Text: The performance of the "hs_override" argument has been improved
#>         in ma_r and ma_d by correcting a bug that caused the argument
#>         to not be evaluated when running an artifact-distribution
#>         meta-analysis or only a bare-bones meta-analysis.
#> 
#> Version: 1.0.3
#> Date: 2018-05-08
#> Text: The default value for the "use_all_arts" argument in ma_r and
#>         ma_d has been changed from FALSE to TRUE. This is meant to
#>         improve the ease with which the most robust
#>         artifact-distribution corrections can be applied.
#> 
#> Version: 1.0.2
#> Date: 2018-05-01
#> Text: Bug fixes have been made for defining confidence-interval method
#>         in bare-bones and artifact-distribution meta-analyses computed
#>         with the ma_r function.
#> 
#> Version: 1.0.2
#> Date: 2018-05-01
#> Text: A new “seed” argument has been added to ma_r and ma_d to set the
#>         seed value when imputing artifacts for greater reproducibility.
#> 
#> Version: 1.0.2
#> Date: 2018-05-01
#> Text: New console messages have been introduced for when follow-up
#>         results are added to a meta-analysis object.
#> 
#> Version: 1.0.1
#> Date: 2018-04-25
#> Text: Bug fixes for using true-score u ratios in individual-correction
#>         meta-analyses.
#> 
#> Version: 1.0.1
#> Date: 2018-04-25
#> Text: Bug fix for appropriately returning results from the compute_dmod
#>         function.
#> 
#> Version: 1.0.0
#> Date: 2018-04-17
#> Text: New “citekey” arguments have been added to meta-analysis function
#>         that allow users to supply citation keys for use in
#>         reference-management programs. The new generate_bib() function
#>         can harvest the citation keys from studies that are included in
#>         meta-analysis objects and create formatted reference lists for
#>         use in publications.
#> 
#> Version: 1.0.0
#> Date: 2018-04-17
#> Text: Plotting functions have been added. The plot_forest() and
#>         plot_funnel() functions use ggplot2 to create forest and funnel
#>         plots, respectively. These functions take a meta-analysis
#>         object as an input and add a list of plots to the object.
#> 
#> Version: 1.0.0
#> Date: 2018-04-17
#> Text: New functions to retrieve results from within meta-analysis
#>         objects have been added. The “get_stuff” family of functions
#>         provide easy ways to extract meta-analysis tables, follow-up
#>         analyses (e.g., cumulative meta-analyses, leave-one-out
#>         meta-analyses, bootstrap results), plots, and more.
#> 
#> Version: 1.0.0
#> Date: 2018-04-17
#> Text: Assorted bug fixes have been implemented to improve the stability
#>         of the dependency resolution process, the handling of artifact
#>         distributions, and the accommodation of continuous moderators.
#> 
#> Version: 0.2.5
#> Date: 2018-03-21
#> Text: Hot fix for consolidating dependent effect sizes.
#> 
#> Version: 0.2.4
#> Date: 2018-03-20
#> Text: A new method to account for sampling error in interactive
#>         artifact distributions has been added (NOTE: this is now the
#>         default methods used in interactive AD meta-analyses).
#> 
#> Version: 0.2.4
#> Date: 2018-03-20
#> Text: Stability improvements for dMod functions.
#> 
#> Version: 0.2.4
#> Date: 2018-03-20
#> Text: Artifacts harvested from rows of a database that do not contain
#>         effects sizes in the ma_r and ma_d functions are now organized
#>         to avoid redundancy with observations from the same sample that
#>         are included in the meta-analytic computations.
#> 
#> Version: 0.2.4
#> Date: 2018-03-20
#> Text: Bug fixes to metabulate().
#> 
#> Version: 0.2.3
#> Date: 2018-02-09
#> Text: Bug fix to the interactive artifact-distribution method for
#>         bivariate indirect range restriction (bvirr) to properly
#>         compute SD rho.
#> 
#> Version: 0.2.3
#> Date: 2018-02-09
#> Text: Bug fix to allow interactive artifact distributions to be rounded
#>         to the desired number of decimal places.
#> 
#> Version: 0.2.3
#> Date: 2018-02-09
#> Text: Bug fix for computing the mean reliability indices in the
#>         Taylor-series estimate of the corrected error variance for
#>         correlations corrected with the bivariate indirect range
#>         restriction (bvirr) correction.
#> 
#> Version: 0.2.2
#> Date: 2018-01-24
#> Text: A new function (called matreg, with aliases of matrixreg, lm_mat,
#>         and lm_matrix) has been added that computes regression models
#>         from covariance matrices and vectors of means and generates
#>         output that resembles that of the "lm" function. Output from
#>         matreg can be used with the summary(), predict(), and confint()
#>         functions.
#> 
#> Version: 0.2.2
#> Date: 2018-01-24
#> Text: The ma_r and mad_d functions have been modified to allow for more
#>         flexibility in computing multiple meta-analyses in a single
#>         function call, especially when those meta-analyses use
#>         artifact-distribution corrections. Several new arguments have
#>         been added to allow users to name the types artifact
#>         corrections they would like to apply to each construct - see
#>         function documentation for argument details.
#> 
#> Version: 0.2.2
#> Date: 2018-01-24
#> Text: The correct_rxx and correct_ryy arguments across meta-analysis
#>         routines (excluding ma_r_ad and ma_d_ad) can now support both
#>         scalar and vector arguments.
#> 
#> Version: 0.2.2
#> Date: 2018-01-24
#> Text: Efficiency improvements and improvements to the accuracy of
#>         progress-bar feedback have been made to the simulate_r_database
#>         and simulate_d_database functions.
#> 
#> Version: 0.2.1
#> Date: 2018-01-07
#> Text: Hotfixes for bugs in “simulate_d_sample” function affecting
#>         composite variables.
#> 
#> Version: 0.2.1
#> Date: 2018-01-07
#> Text: A “k_items_vec” argument is now included in the “simulate_psych”
#>         function.
#> 
#> Version: 0.2.1
#> Date: 2018-01-07
#> Text: The “simulate_d_database” function has been improved for faster
#>         computations.
#> 
#> Version: 0.2.0
#> Date: 2018-01-05
#> Text: Bug fixes for identifying independent artifacts in
#>         artifact-distribution meta-analyses using ma_r and ma_d.
#> 
#> Version: 0.2.0
#> Date: 2018-01-05
#> Text: The simulation functions (simulate_r_sample, simulate_r_database,
#>         simulate_d_sample, simulate_d_database) can now simulate alpha
#>         reliabilities for variables when the number of items in a scale
#>         is indicated with the “k_items_vec” or “k_items_params”
#>         arguments.
#> 
#> Version: 0.2.0
#> Date: 2018-01-05
#> Text: A new function called “metabulate" has been added that writes
#>         meta-analytic tables as rich text files with near
#>         publication-quality formatting.
#> 
#> Version: 0.1.3
#> Date: 2017-12-13
#> Text: Bug fixes and stability improvements for moderator analyses, the
#>         creation of artifact distributions, and computation of
#>         follow-up analyses.
#> 
#> Version: 0.1.3
#> Date: 2017-12-13
#> Text: Added reliability types arguments (e.g., “rxx_type”, “ryy_type”)
#>         to functions requiring reliability estimates to be corrected
#>         for range restriction. String vectors of reliability labels
#>         (e.g., “alpha”, “retest”, “parallel”; see documentation for the
#>         ma_r function for a complete list of supported labels) can now
#>         be supplied to many functions. When direct range restriction
#>         occurs, different corrections for range restriction are applied
#>         to internal-consistency reliability estimates and reliability
#>         estimates computed by correlating data from different testing
#>         occasions.
#> 
#> Version: 0.1.3
#> Date: 2017-12-13
#> Text: The ma_r and ma_d functions now allow users to supply lists of
#>         external artifact information using the “supplemental_ads”
#>         argument (the ma_r_ic and ma_d_ic functions also allow this
#>         with the “supplemental_ads_x” and “supplemental_ads_y”
#>         arguments). These functions can also now harvest artifact
#>         information from studies in a database with invalid or missing
#>         effect sizes by setting the “use_all_arts” argument to TRUE.
#> 
#> Version: 0.1.3
#> Date: 2017-12-13
#> Text: “rho_params” arguments to the simulate_r_database function can
#>         now be supplied as a correlation matrix rather than a list (for
#>         situations in which samples are drawn from a single
#>         population). Similarly, elements in the “rho_params” list
#>         argument for the simulate_d_database can include matrices.
#> 
#> Version: 0.1.3
#> Date: 2017-12-13
#> Text: Progress bars have been added to potentially time-consuming
#>         functions to provide feedback about the percentage of progress
#>         and the estimated time remaining.
#> 
#> Version: 0.1.2
#> Date: 2017-11-16
#> Text: Bug fixes for moderator analyses.
#> 
#> Version: 0.1.2
#> Date: 2017-11-16
#> Text: Bug fixes for specifying the order of constructs and selecting a
#>         subset of constructs in a database.
#> 
#> Version: 0.1.2
#> Date: 2017-11-16
#> Text: New function to correct for scale coarseness.
#> 
#> Version: 0.1.2
#> Date: 2017-11-16
#> Text: Improved vectorization support.
#> 
#> Version: 0.1.1
#> Date: 2017-09-21
#> Text: We have added a suite of simulation function for d values (see
#>         the "simulate_d_sample" and "simulate_d_database" functions).
#> 
#> Version: 0.1.1
#> Date: 2017-09-21
#> Text: Methods for meta-analyzing d values have been updated to better
#>         account for subgroup proportions when converting between the r
#>         and d metrics. The escalc objects that accompany a
#>         meta-analysis of d values now include more detailed information
#>         about incumbent and applicant subgroup proportions and the
#>         default for the "pa" argument (i.e., applicant proportions) is
#>         now NULL rather than .5 to prevent unintended corrections for
#>         range restriction.
#> 
#> Version: 0.1.1
#> Date: 2017-09-21
#> Text: A new "ma_generic" function has been added that allow users to do
#>         a psychmeta-style meta-analysis for any effect size for which
#>         the user has error-variance estimates. This function is
#>         supported by psychmeta's sensitivity() and heterogeneity()
#>         functions.
#> 
#> Version: 0.1.1
#> Date: 2017-09-21
#> Text: Taylor series methods for estimating the variance of converted
#>         artifact values (e.g., a rxxi distribution converted to a rxxa
#>         distribution) have been expanded to account for the correlation
#>         between artifact distributions when such information is
#>         available. Correlations among artifacts' sampling distributions
#>         can also be passed to the var_error_r_bvirr() function when it
#>         is called outside of a meta-analysis routine.
#> 
#> Version: 0.1.1
#> Date: 2017-09-21
#> Text: The implementation of corrections of bivariate range restriction
#>         in individual-correction meta-analyses have been updated.
#>         Corrections for bivariate direct range restriction now use
#>         conventional compound attenuation factors, while the
#>         attenuation factors for bivariate indirect range restriction
#>         are now computed using mean effect sizes and artifacts to
#>         estimate sampling variances.
#> 
#> Version: 0.1.1
#> Date: 2017-09-21
#> Text: Assorted bugs have been fixed (e.g., an error that occurred when
#>         running multi-construct meta-analyses without specifying
#>         moderators; filters to retain valid correlations did not screen
#>         construct names, sample ids, or measure names; subgroup
#>         proportions were not being retained when d values were
#>         meta-analyzed as correlations).
#> 
#> Version: 0.1.0
#> Date: 2017-08-15
#> Text: psychmeta has officially launched for public beta!
#> 
#> Version: 0.1.0
#> Date: 2017-08-15
#> Text: Please see the "psychmeta-package" entry in the psychmeta manual
#>         for an overview of the package and its applications.