Simulate a sample of psychometric d value data with measurement error, direct range restriction, and/or indirect range restriction

This function generates a simulated psychometric sample consisting of any number of groups and computes the d values that result after introducing measurement error and/or range restriction.

simulate_d_sample(
  n_vec,
  rho_mat_list,
  mu_mat,
  sigma_mat = 1,
  rel_mat = 1,
  sr_vec = 1,
  k_items_vec = 1,
  wt_mat = NULL,
  sr_composites = NULL,
  group_names = NULL,
  var_names = NULL,
  composite_names = NULL,
  diffs_as_obs = FALSE
)

Arguments

n_vec
Vector of sample sizes (or a vector of proportions, if parameters are to be estimated).

rho_mat_list
List of true-score correlation matrices.

mu_mat
Matrix of mean parameters, with groups on the rows and variables on the columns.

sigma_mat
Matrix of standard-deviation parameters, with groups on the rows and variables on the columns.

rel_mat
Matrix of reliability parameters, with groups on the rows and variables on the columns.

sr_vec
Vector of selection ratios.

k_items_vec
Number of test items comprising each of the variables to be simulated (all are single-item variables by default).

wt_mat
Optional matrix of weights to use in forming a composite of the variables in rho_mat. Matrix should have as many rows (or vector elements) as there are variables in rho_mat.

sr_composites
Optional vector selection ratios for composite variables. If not NULL, sr_composites must have as many elements as there are columns in wt_mat.

group_names
Optional vector of group names.

var_names
Optional vector of variable names.

composite_names
Optional vector of names for composite variables.

diffs_as_obs
Logical scalar that determines whether standard deviation parameters represent standard deviations of observed scores (TRUE) or of true scores (FALSE; default).

Value

A sample of simulated mean differences.

Examples

## Simulate statistics by providing integers as "n_vec":
simulate_d_sample(n_vec = c(200, 100), rho_mat_list = list(reshape_vec2mat(.5),
                                                           reshape_vec2mat(.4)),
                  mu_mat = rbind(c(1, .5), c(0, 0)), sigma_mat = rbind(c(1, 1), c(1, 1)),
                  rel_mat = rbind(c(.8, .7), c(.7, .7)), sr_vec = c(1, .5),
                  group_names = c("A", "B"))

#> Results of Simulated Study with 2 Groups (Statistics) 
#> --------------------------
#> 
#> Simulated 300 applicant cases and selected 150 incumbent cases for an overall selection ratio of 50%.
#> 
#> # A tibble: 2 x 53
#>   group1 group2 y_name   ni1   ni2   na1   na2   dyi   dya    pi    pa
#>   <chr>  <chr>  <chr>  <int> <int> <int> <int> <dbl> <dbl> <dbl> <dbl>
#> 1 A      B      y1       111    39   200   100 1.05  0.829  0.74 0.667
#> 2 A      B      y2       111    39   200   100 0.451 0.503  0.74 0.667
#> # … with 42 more variables: parallel_ryyi_pooled <dbl>,
#> #   parallel_ryyi_total <dbl>, parallel_ryyi1 <dbl>, parallel_ryyi2 <dbl>,
#> #   raw_alpha_yi_pooled <lgl>, raw_alpha_yi_total <lgl>, raw_alpha_yi1 <dbl>,
#> #   raw_alpha_yi2 <dbl>, std_alpha_yi_pooled <lgl>, std_alpha_yi_total <lgl>,
#> #   std_alpha_yi1 <dbl>, std_alpha_yi2 <dbl>, parallel_ryya_pooled <dbl>,
#> #   parallel_ryya_total <dbl>, parallel_ryya1 <dbl>, parallel_ryya2 <dbl>,
#> #   raw_alpha_ya_pooled <lgl>, raw_alpha_ya_total <lgl>, raw_alpha_ya1 <dbl>,
#> #   raw_alpha_ya2 <dbl>, std_alpha_ya_pooled <lgl>, std_alpha_ya_total <lgl>,
#> #   std_alpha_ya1 <dbl>, std_alpha_ya2 <dbl>, uy_pooled <dbl>, uy_total <dbl>,
#> #   uy1 <dbl>, uy2 <dbl>, meanyi_total <dbl>, meanyi1 <dbl>, meanyi2 <dbl>,
#> #   meanya_total <dbl>, meanya1 <dbl>, meanya2 <dbl>, sdyi_pooled <dbl>,
#> #   sdyi_total <dbl>, sdyi1 <dbl>, sdyi2 <dbl>, sdya_pooled <dbl>,
#> #   sdya_total <dbl>, sdya1 <dbl>, sdya2 <dbl>

## Simulate parameters by providing proportions as "n_vec":
simulate_d_sample(n_vec = c(2/3, 1/3), rho_mat_list = list(reshape_vec2mat(.5),
                                                           reshape_vec2mat(.4)),
                  mu_mat = rbind(c(1, .5), c(0, 0)), sigma_mat = rbind(c(1, 1), c(1, 1)),
                  rel_mat = rbind(c(.8, .7), c(.7, .7)), sr_vec = c(1, .5),
                  group_names = c("A", "B"))

#> Error in .simulate_d_sample_params(p_vec = args$n_vec/sum(args$n_vec),     rho_mat_list = args$rho_mat_list, mu_mat = args$mu_mat, sigma_mat = args$sigma_mat,     rel_mat = args$rel_mat, sr_vec = args$sr_vec, k_items_vec = args$k_items_vec,     group_names = args$group_names, var_names = args$var_names,     sr_composites = args$sr_composites, composite_names = args$composite_names,     wt_mat = args$wt_mat, show_applicant = TRUE, diffs_as_obs = args$diffs_as_obs): The package 'nor1mix' is not installed.
#>   'nor1mix' is required to estimate population parameters under multivariate selection.
#>   Please install 'nor1mix'.