Generate a list of simulated sample matrices sampled from the Wishart distribution

This function generates simulated sample matrices based on a population matrix and a sample size. It uses the Wishart distribution (i.e., the multivariate \(\chi^2\) distribution) to obtain data, rescales the data into the input metric, and can be standardized into a correlation matrix by setting as_cor to TRUE. The function can produce a list of matrices for any number of samples.

simulate_matrix(sigma, n, k = 1, as_cor = FALSE)

Arguments

sigma
Population covariance matrix. May be standardized or unstandardized.

n
Sample size for simulated sample matrices.

k
Number of sample matrices to generate.

as_cor
Should the simulated matrices be standardized (TRUE) or unstandardized (FALSE)?

Value

A list of simulated sample matrices.

Examples

## Define a hypothetical matrix:
sigma <- reshape_vec2mat(cov = .4, order = 5)

## Simualte a list of unstandardized covariance matrices:
simulate_matrix(sigma = sigma, n = 50, k = 10, as_cor = FALSE)

#> $`Sample 1`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0683364 0.3473643 0.4838062 0.4323141 0.2290617
#> Var2 0.3473643 0.8056304 0.3338216 0.2336477 0.2771231
#> Var3 0.4838062 0.3338216 1.1043261 0.4362124 0.2794521
#> Var4 0.4323141 0.2336477 0.4362124 0.8446489 0.2183814
#> Var5 0.2290617 0.2771231 0.2794521 0.2183814 0.9628289
#> 
#> $`Sample 2`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 0.9345340 0.4159207 0.3561820 0.2759175 0.1603988
#> Var2 0.4159207 1.2591666 0.6478552 0.3512885 0.3632822
#> Var3 0.3561820 0.6478552 1.2859956 0.4082996 0.4153353
#> Var4 0.2759175 0.3512885 0.4082996 0.6018280 0.2063425
#> Var5 0.1603988 0.3632822 0.4153353 0.2063425 0.8694426
#> 
#> $`Sample 3`
#>           Var1       Var2       Var3       Var4       Var5
#> Var1 0.8607192 0.35328775 0.11030555 0.24158927 0.28796245
#> Var2 0.3532878 0.74437854 0.08241279 0.31451568 0.14557528
#> Var3 0.1103055 0.08241279 0.49803594 0.09739064 0.04866251
#> Var4 0.2415893 0.31451568 0.09739064 0.67059418 0.18371540
#> Var5 0.2879625 0.14557528 0.04866251 0.18371540 0.84299559
#> 
#> $`Sample 4`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 0.7533902 0.3193048 0.4226988 0.4039051 0.3042532
#> Var2 0.3193048 0.9952307 0.4273556 0.3876090 0.4385034
#> Var3 0.4226988 0.4273556 0.8476912 0.4111993 0.3954772
#> Var4 0.4039051 0.3876090 0.4111993 0.9698303 0.3422877
#> Var5 0.3042532 0.4385034 0.3954772 0.3422877 0.9365318
#> 
#> $`Sample 5`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 0.8355127 0.3160253 0.3095169 0.2742366 0.4392838
#> Var2 0.3160253 0.7546977 0.4553412 0.3206004 0.4305910
#> Var3 0.3095169 0.4553412 1.0510751 0.3637123 0.5857690
#> Var4 0.2742366 0.3206004 0.3637123 0.8648559 0.4165791
#> Var5 0.4392838 0.4305910 0.5857690 0.4165791 1.0969169
#> 
#> $`Sample 6`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.3727987 0.4608626 0.3430505 0.6138211 0.4812637
#> Var2 0.4608626 1.0660014 0.4510476 0.5765529 0.5548843
#> Var3 0.3430505 0.4510476 0.7897439 0.4981923 0.4268082
#> Var4 0.6138211 0.5765529 0.4981923 1.2175105 0.7462946
#> Var5 0.4812637 0.5548843 0.4268082 0.7462946 1.3507669
#> 
#> $`Sample 7`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 0.8062211 0.5190346 0.2004088 0.2275253 0.3630753
#> Var2 0.5190346 1.0376631 0.2702969 0.4092047 0.5237935
#> Var3 0.2004088 0.2702969 0.8232760 0.2720221 0.1609581
#> Var4 0.2275253 0.4092047 0.2720221 0.6643051 0.2449200
#> Var5 0.3630753 0.5237935 0.1609581 0.2449200 1.0468297
#> 
#> $`Sample 8`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0459203 0.3216995 0.4114503 0.2833668 0.4652524
#> Var2 0.3216995 0.6075279 0.3727611 0.2595439 0.2651401
#> Var3 0.4114503 0.3727611 0.9080435 0.2325271 0.1533405
#> Var4 0.2833668 0.2595439 0.2325271 0.7373389 0.1534255
#> Var5 0.4652524 0.2651401 0.1533405 0.1534255 0.9891237
#> 
#> $`Sample 9`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.3948231 0.6926985 0.3333452 0.5078066 0.5225731
#> Var2 0.6926985 1.0649326 0.2830735 0.3680131 0.4232601
#> Var3 0.3333452 0.2830735 0.7342637 0.3236488 0.3462927
#> Var4 0.5078066 0.3680131 0.3236488 0.8435914 0.3290280
#> Var5 0.5225731 0.4232601 0.3462927 0.3290280 0.9189670
#> 
#> $`Sample 10`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.1069973 0.6978940 0.4756483 0.2150381 0.3601556
#> Var2 0.6978940 0.8916490 0.4104308 0.2868680 0.3010647
#> Var3 0.4756483 0.4104308 0.7118901 0.2678608 0.2182475
#> Var4 0.2150381 0.2868680 0.2678608 0.7575259 0.2559988
#> Var5 0.3601556 0.3010647 0.2182475 0.2559988 0.6892784
#> 

## Simualte a list of correlation matrices:
simulate_matrix(sigma = sigma, n = 50, k = 10, as_cor = TRUE)

#> $`Sample 1`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.5282607 0.5251390 0.4805475 0.3744626
#> Var2 0.5282607 1.0000000 0.4240135 0.4257780 0.2983002
#> Var3 0.5251390 0.4240135 1.0000000 0.4636513 0.4934199
#> Var4 0.4805475 0.4257780 0.4636513 1.0000000 0.5195001
#> Var5 0.3744626 0.2983002 0.4934199 0.5195001 1.0000000
#> 
#> $`Sample 2`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.3492999 0.3346629 0.3074550 0.3610874
#> Var2 0.3492999 1.0000000 0.3848689 0.3727775 0.3947252
#> Var3 0.3346629 0.3848689 1.0000000 0.3561384 0.4467760
#> Var4 0.3074550 0.3727775 0.3561384 1.0000000 0.3135572
#> Var5 0.3610874 0.3947252 0.4467760 0.3135572 1.0000000
#> 
#> $`Sample 3`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.4762906 0.5628236 0.3867214 0.4540104
#> Var2 0.4762906 1.0000000 0.6120597 0.4552400 0.4956264
#> Var3 0.5628236 0.6120597 1.0000000 0.3850610 0.4133984
#> Var4 0.3867214 0.4552400 0.3850610 1.0000000 0.3857532
#> Var5 0.4540104 0.4956264 0.4133984 0.3857532 1.0000000
#> 
#> $`Sample 4`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.5378410 0.3387209 0.3956402 0.5362296
#> Var2 0.5378410 1.0000000 0.3480725 0.5374853 0.6043317
#> Var3 0.3387209 0.3480725 1.0000000 0.4579758 0.4384663
#> Var4 0.3956402 0.5374853 0.4579758 1.0000000 0.6361192
#> Var5 0.5362296 0.6043317 0.4384663 0.6361192 1.0000000
#> 
#> $`Sample 5`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.4908583 0.4568105 0.3680510 0.2610223
#> Var2 0.4908583 1.0000000 0.4812614 0.4138212 0.2882620
#> Var3 0.4568105 0.4812614 1.0000000 0.4497032 0.2753690
#> Var4 0.3680510 0.4138212 0.4497032 1.0000000 0.4099602
#> Var5 0.2610223 0.2882620 0.2753690 0.4099602 1.0000000
#> 
#> $`Sample 6`
#>             Var1      Var2      Var3      Var4        Var5
#> Var1  1.00000000 0.2751139 0.3557941 0.3253125 -0.01847821
#> Var2  0.27511389 1.0000000 0.1426664 0.1960688  0.13811250
#> Var3  0.35579406 0.1426664 1.0000000 0.2832819  0.27762262
#> Var4  0.32531248 0.1960688 0.2832819 1.0000000  0.34414229
#> Var5 -0.01847821 0.1381125 0.2776226 0.3441423  1.00000000
#> 
#> $`Sample 7`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.5071588 0.3113188 0.5057508 0.5029372
#> Var2 0.5071588 1.0000000 0.3604981 0.3709960 0.4071177
#> Var3 0.3113188 0.3604981 1.0000000 0.5785768 0.4184082
#> Var4 0.5057508 0.3709960 0.5785768 1.0000000 0.5764721
#> Var5 0.5029372 0.4071177 0.4184082 0.5764721 1.0000000
#> 
#> $`Sample 8`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.4038451 0.6262862 0.4920425 0.4757643
#> Var2 0.4038451 1.0000000 0.5093327 0.4694287 0.3345657
#> Var3 0.6262862 0.5093327 1.0000000 0.6252262 0.4155766
#> Var4 0.4920425 0.4694287 0.6252262 1.0000000 0.1908506
#> Var5 0.4757643 0.3345657 0.4155766 0.1908506 1.0000000
#> 
#> $`Sample 9`
#>           Var1      Var2      Var3      Var4      Var5
#> Var1 1.0000000 0.1828855 0.3259527 0.3981729 0.1710999
#> Var2 0.1828855 1.0000000 0.4757279 0.3967386 0.2550562
#> Var3 0.3259527 0.4757279 1.0000000 0.2716189 0.3291192
#> Var4 0.3981729 0.3967386 0.2716189 1.0000000 0.3011394
#> Var5 0.1710999 0.2550562 0.3291192 0.3011394 1.0000000
#> 
#> $`Sample 10`
#>           Var1      Var2      Var3       Var4       Var5
#> Var1 1.0000000 0.1088464 0.2156792 0.24490946 0.39186392
#> Var2 0.1088464 1.0000000 0.5264771 0.27527218 0.40679912
#> Var3 0.2156792 0.5264771 1.0000000 0.17097322 0.36078538
#> Var4 0.2449095 0.2752722 0.1709732 1.00000000 0.08012169
#> Var5 0.3918639 0.4067991 0.3607854 0.08012169 1.00000000
#>