Simulate Monte Carlo psychometric data (observed, true, and error scores)

Simulate Monte Carlo psychometric data (observed, true, and error scores)

simulate_psych(
  n,
  rho_mat,
  mu_vec = rep(0, ncol(rho_mat)),
  sigma_vec = rep(1, ncol(rho_mat)),
  rel_vec = rep(1, ncol(rho_mat)),
  sr_vec = rep(1, ncol(rho_mat)),
  k_items_vec = rep(1, ncol(rho_mat)),
  wt_mat = NULL,
  sr_composites = NULL,
  var_names = NULL,
  composite_names = NULL
)

Arguments

n
Number of cases to simulate before performing selection.

rho_mat
Matrix of true-score correlations.

mu_vec
Vector of means.

sigma_vec
Vector of observed-score standard deviations.

rel_vec
Vector of reliabilities corresponding to the variables in rho_mat.

sr_vec
Vector of selection ratios corresponding to the variables in rho_mat. (set selection ratios to 1 for variables that should not be used in selection).

k_items_vec
Number of test items comprising each of the variables to be simulated (all are single-item variables by default).

wt_mat
Optional matrix of weights to use in forming a composite of the variables in rho_mat. Matrix should have as many rows (or vector elements) as there are variables in rho_mat.

sr_composites
Optional vector selection ratios for composite variables. If not NULL, sr_composites must have as many elements as there are columns in wt_mat.

var_names
Vector of variable names corresponding to the variables in rho_mat.

composite_names
Optional vector of names for composite variables.

Value

A list of observed-score, true-score, and error-score data frames. If selection is requested, the data frames will include logical variables indicating whether each case would be selected on the basis of observed scores, true scores, or error scores.

Examples

## Generate data for a simple sample with two variables without selection:
simulate_psych(n = 1000, rho_mat = matrix(c(1, .5, .5, 1), 2, 2), sigma_vec = c(1, 1),
          rel_vec = c(.8, .8), var_names = c("Y", "X"))

#> Data from a Simulated Study of 1000 Cases
#> --------------------------
#> 
#> Observed scores:
#> # A tibble: 1,000 x 2
#>          Y       X
#>      <dbl>   <dbl>
#>  1  0.451   0.333 
#>  2  0.416  -1.21  
#>  3 -0.558   1.10  
#>  4 -0.0636 -0.295 
#>  5  0.353   0.0923
#>  6 -1.45   -1.27  
#>  7 -0.459  -0.501 
#>  8 -0.361   1.04  
#>  9  0.137  -1.20  
#> 10  0.870   0.952 
#> # … with 990 more rows
#> 
#> True scores:
#> # A tibble: 1,000 x 2
#>          Y      X
#>      <dbl>  <dbl>
#>  1  0.292   0.861
#>  2  0.576  -0.862
#>  3 -0.286   1.79 
#>  4  0.355  -0.434
#>  5  0.226  -0.665
#>  6 -1.01   -1.05 
#>  7 -0.745  -0.163
#>  8  0.0139  0.437
#>  9  0.321  -1.18 
#> 10  0.852   0.509
#> # … with 990 more rows
#> 
#> Error scores:
#> # A tibble: 1,000 x 2
#>          Y       X
#>      <dbl>   <dbl>
#>  1  0.159  -0.528 
#>  2 -0.160  -0.345 
#>  3 -0.272  -0.685 
#>  4 -0.419   0.139 
#>  5  0.127   0.757 
#>  6 -0.435  -0.218 
#>  7  0.286  -0.338 
#>  8 -0.375   0.598 
#>  9 -0.183  -0.0167
#> 10  0.0181  0.443 
#> # … with 990 more rows

## Generate data for a simple sample with two variables with selection:
simulate_psych(n = 1000, rho_mat = matrix(c(1, .5, .5, 1), 2, 2), sigma_vec = c(1, 1),
          rel_vec = c(.8, .8), sr_vec = c(1, .5), var_names = c("Y", "X"))

#> Data from a Simulated Study of 1000 Cases
#> --------------------------
#> 
#> Observed scores:
#> # A tibble: 1,000 x 5
#>         Y      X selected_obs selected_true selected_error
#>     <dbl>  <dbl> <lgl>        <lgl>         <lgl>         
#>  1 -0.519 -2.04  FALSE        FALSE         FALSE         
#>  2 -1.60  -0.554 FALSE        FALSE         TRUE          
#>  3 -0.356 -0.288 FALSE        FALSE         TRUE          
#>  4 -0.671 -0.775 FALSE        FALSE         TRUE          
#>  5 -1.52  -1.80  FALSE        FALSE         FALSE         
#>  6  0.209 -0.834 FALSE        FALSE         FALSE         
#>  7  0.843 -0.890 FALSE        FALSE         FALSE         
#>  8 -0.172 -2.38  FALSE        FALSE         FALSE         
#>  9  1.13   0.317 TRUE         TRUE          FALSE         
#> 10  0.162  0.958 TRUE         TRUE          TRUE          
#> # … with 990 more rows
#> 
#> True scores:
#> # A tibble: 1,000 x 5
#>         Y      X selected_obs selected_true selected_error
#>     <dbl>  <dbl> <lgl>        <lgl>         <lgl>         
#>  1 -0.291 -1.38  FALSE        FALSE         FALSE         
#>  2 -1.45  -0.901 FALSE        FALSE         TRUE          
#>  3 -0.113 -0.514 FALSE        FALSE         TRUE          
#>  4 -0.529 -1.74  FALSE        FALSE         TRUE          
#>  5 -1.76  -1.13  FALSE        FALSE         FALSE         
#>  6  0.839 -0.316 FALSE        FALSE         FALSE         
#>  7  0.693 -0.579 FALSE        FALSE         FALSE         
#>  8  0.221 -1.38  FALSE        FALSE         FALSE         
#>  9  1.69   0.404 TRUE         TRUE          FALSE         
#> 10  0.742  0.733 TRUE         TRUE          TRUE          
#> # … with 990 more rows
#> 
#> Error scores:
#> # A tibble: 1,000 x 5
#>         Y       X selected_obs selected_true selected_error
#>     <dbl>   <dbl> <lgl>        <lgl>         <lgl>         
#>  1 -0.228 -0.660  FALSE        FALSE         FALSE         
#>  2 -0.151  0.348  FALSE        FALSE         TRUE          
#>  3 -0.243  0.226  FALSE        FALSE         TRUE          
#>  4 -0.141  0.970  FALSE        FALSE         TRUE          
#>  5  0.239 -0.671  FALSE        FALSE         FALSE         
#>  6 -0.631 -0.518  FALSE        FALSE         FALSE         
#>  7  0.150 -0.311  FALSE        FALSE         FALSE         
#>  8 -0.393 -0.992  FALSE        FALSE         FALSE         
#>  9 -0.558 -0.0876 TRUE         TRUE          FALSE         
#> 10 -0.580  0.226  TRUE         TRUE          TRUE          
#> # … with 990 more rows

## Generate data for samples with five variables, of which subsets are used to form composites:
rho_mat <- matrix(.5, 5, 5)
diag(rho_mat) <- 1
simulate_psych(n = 1000, rho_mat = rho_mat,
                rel_vec = rep(.8, 5), sr_vec = c(1, 1, 1, 1, .5),
                wt_mat = cbind(c(0, 0, 0, .3, 1), c(1, .3, 0, 0, 0)), sr_composites = c(.7, .5))

#> Data from a Simulated Study of 1000 Cases
#> --------------------------
#> 
#> Observed scores:
#> # A tibble: 1,000 x 10
#>        x1      x2      x3      x4     x5 composite1 composite2 selected_obs
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>      <dbl>      <dbl> <lgl>       
#>  1 -0.567 -0.958  -1.20   -0.392  -0.809    -0.926      -0.855 FALSE       
#>  2  1.66   0.978   0.0935  0.594   0.248     0.426       1.95  TRUE        
#>  3 -0.917 -0.560  -0.581  -0.0567 -0.685    -0.702      -1.09  FALSE       
#>  4 -0.193 -0.0113 -1.12   -0.677  -0.691    -0.894      -0.196 FALSE       
#>  5 -0.682  0.179  -0.597  -0.872  -0.699    -0.960      -0.629 FALSE       
#>  6 -0.412 -1.57    0.877   0.369  -0.177    -0.0664     -0.884 FALSE       
#>  7  0.431 -0.604   1.13    0.314   0.684     0.778       0.250 TRUE        
#>  8 -1.72  -2.01   -3.72   -2.75   -0.580    -1.40       -2.32  FALSE       
#>  9  1.41   0.793   0.638   2.00    2.03      2.63        1.65  TRUE        
#> 10 -0.246 -0.200  -0.222   0.714  -0.290    -0.0755     -0.306 FALSE       
#> # … with 990 more rows, and 2 more variables: selected_true <lgl>,
#> #   selected_error <lgl>
#> 
#> True scores:
#> # A tibble: 1,000 x 10
#>         x1      x2      x3     x4     x5 composite1 composite2 selected_obs
#>      <dbl>   <dbl>   <dbl>  <dbl>  <dbl>      <dbl>      <dbl> <lgl>       
#>  1 -0.790  -1.03   -0.517  -0.783 -0.851     -1.09     -1.10   FALSE       
#>  2  1.44    0.690   0.578   0.317  0.791      0.886     1.65   TRUE        
#>  3  0.0179 -0.0122 -0.320   0.189 -0.448     -0.392     0.0142 FALSE       
#>  4  0.386   0.209   0.0658 -0.801 -0.573     -0.814     0.448  FALSE       
#>  5 -0.355   0.462  -0.968  -1.84  -0.775     -1.33     -0.216  FALSE       
#>  6 -0.545  -1.02   -0.0423 -0.405 -0.155     -0.277    -0.851  FALSE       
#>  7  0.966  -0.966   1.21    0.126  1.04       1.08      0.677  TRUE        
#>  8 -1.37   -1.94   -3.42   -2.82  -0.477     -1.32     -1.96   FALSE       
#>  9  1.58    0.644   0.939   1.44   1.72       2.15      1.77   TRUE        
#> 10 -0.0863  0.205  -0.142   0.905 -0.456     -0.184    -0.0249 FALSE       
#> # … with 990 more rows, and 2 more variables: selected_true <lgl>,
#> #   selected_error <lgl>
#> 
#> Error scores:
#> # A tibble: 1,000 x 10
#>        x1      x2      x3      x4      x5 composite1 composite2 selected_obs
#>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>      <dbl>      <dbl> <lgl>       
#>  1  0.223  0.0683 -0.685   0.392   0.0420     0.160      0.243  FALSE       
#>  2  0.218  0.288  -0.484   0.277  -0.543     -0.460      0.304  TRUE        
#>  3 -0.935 -0.548  -0.261  -0.246  -0.237     -0.311     -1.10   FALSE       
#>  4 -0.578 -0.220  -1.18    0.124  -0.117     -0.0802    -0.644  FALSE       
#>  5 -0.328 -0.283   0.371   0.968   0.0767     0.367     -0.413  FALSE       
#>  6  0.133 -0.554   0.919   0.774  -0.0220     0.210     -0.0334 FALSE       
#>  7 -0.535  0.362  -0.0884  0.188  -0.354     -0.298     -0.427  TRUE        
#>  8 -0.343 -0.0643 -0.303   0.0712 -0.103     -0.0814    -0.363  FALSE       
#>  9 -0.162  0.149  -0.301   0.563   0.307      0.476     -0.117  TRUE        
#> 10 -0.160 -0.405  -0.0798 -0.191   0.166      0.109     -0.281  FALSE       
#> # … with 990 more rows, and 2 more variables: selected_true <lgl>,
#> #   selected_error <lgl>

## Generate data for similar scenario as above, but with scales consisting of 1-5 items:
rho_mat <- matrix(.5, 5, 5)
diag(rho_mat) <- 1
simulate_psych(n = 1000, rho_mat = rho_mat,
                rel_vec = rep(.8, 5), sr_vec = c(1, 1, 1, 1, .5),
                k_items_vec = 1:5,
                wt_mat = cbind(c(0, 0, 0, .3, 1), c(1, .3, 0, 0, 0)), sr_composites = c(.7, .5))

#> Data from a Simulated Study of 1000 Cases
#> --------------------------
#> 
#> Observed scores:
#> # A tibble: 1,000 x 25
#>        x1      x2      x3      x4     x5 composite1 composite2 x1_item1 x2_item1
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>      <dbl>      <dbl>    <dbl>    <dbl>
#>  1  0.454  0.251   0.923   0.102  -1.94      -1.91       0.529    0.454  -0.419 
#>  2  1.32   0.438  -1.27    0.413  -0.553     -0.429      1.45     1.32    0.0177
#>  3  1.42  -0.0613  0.0693  0.365  -0.223     -0.113      1.40     1.42   -0.103 
#>  4  1.07  -0.905  -0.649   0.875  -0.799     -0.536      0.802    1.07   -0.453 
#>  5  0.507 -0.678   2.60    2.48    1.19       1.94       0.304    0.507  -0.239 
#>  6  0.136  0.905  -0.176   0.231   0.843      0.912      0.407    0.136   0.409 
#>  7  0.251 -2.15   -1.39   -1.27   -0.493     -0.874     -0.393    0.251  -1.28  
#>  8  1.91  -0.0869  0.466   0.522   0.729      0.886      1.88     1.91    0.484 
#>  9 -1.13  -0.384   0.253  -0.0958 -0.381     -0.410     -1.24    -1.13   -0.609 
#> 10  0.294  1.26    1.15    0.954  -0.567     -0.281      0.673    0.294   0.770 
#> # … with 990 more rows, and 16 more variables: x2_item2 <dbl>, x3_item1 <dbl>,
#> #   x3_item2 <dbl>, x3_item3 <dbl>, x4_item1 <dbl>, x4_item2 <dbl>,
#> #   x4_item3 <dbl>, x4_item4 <dbl>, x5_item1 <dbl>, x5_item2 <dbl>,
#> #   x5_item3 <dbl>, x5_item4 <dbl>, x5_item5 <dbl>, selected_obs <lgl>,
#> #   selected_true <lgl>, selected_error <lgl>
#> 
#> True scores:
#> # A tibble: 1,000 x 25
#>        x1      x2     x3     x4     x5 composite1 composite2 x1_item1 x2_item1
#>     <dbl>   <dbl>  <dbl>  <dbl>  <dbl>      <dbl>      <dbl>    <dbl>    <dbl>
#>  1  0.695 -0.0730  1.10  -0.412 -1.41     -1.53       0.673     0.695  -0.0365
#>  2  0.609  0.433  -0.540  0.464 -0.461    -0.322      0.739     0.609   0.217 
#>  3  0.705 -0.345  -0.681  0.153  0.162     0.207      0.601     0.705  -0.173 
#>  4  1.11  -1.03   -0.394  0.470 -0.194    -0.0534     0.804     1.11   -0.517 
#>  5  0.872 -0.679   2.38   1.70   1.12      1.63       0.669     0.872  -0.339 
#>  6 -0.183  0.843  -0.285  0.222  1.02      1.08       0.0695   -0.183   0.422 
#>  7 -0.309 -1.54   -1.10  -0.893 -0.744    -1.01      -0.772    -0.309  -0.772 
#>  8  1.44   0.420  -0.139  0.890  1.59      1.86       1.57      1.44    0.210 
#>  9 -0.649 -0.383  -0.219  0.302 -1.25     -1.16      -0.764    -0.649  -0.191 
#> 10  0.489  1.41    0.805  0.905  0.479     0.750      0.910     0.489   0.703 
#> # … with 990 more rows, and 16 more variables: x2_item2 <dbl>, x3_item1 <dbl>,
#> #   x3_item2 <dbl>, x3_item3 <dbl>, x4_item1 <dbl>, x4_item2 <dbl>,
#> #   x4_item3 <dbl>, x4_item4 <dbl>, x5_item1 <dbl>, x5_item2 <dbl>,
#> #   x5_item3 <dbl>, x5_item4 <dbl>, x5_item5 <dbl>, selected_obs <lgl>,
#> #   selected_true <lgl>, selected_error <lgl>
#> 
#> Error scores:
#> # A tibble: 1,000 x 25
#>         x1        x2     x3       x4      x5 composite1 composite2 x1_item1
#>      <dbl>     <dbl>  <dbl>    <dbl>   <dbl>      <dbl>      <dbl>    <dbl>
#>  1 -0.241   0.324    -0.173  0.514   -0.535      -0.381   -0.144    -0.241 
#>  2  0.712   0.00490  -0.734 -0.0509  -0.0918     -0.107    0.713     0.712 
#>  3  0.713   0.284     0.750  0.213   -0.384      -0.320    0.798     0.713 
#>  4 -0.0405  0.129    -0.255  0.406   -0.605      -0.483   -0.00185  -0.0405
#>  5 -0.365   0.000697  0.226  0.775    0.0721      0.305   -0.365    -0.365 
#>  6  0.319   0.0618    0.110  0.00841 -0.174      -0.171    0.338     0.319 
#>  7  0.561  -0.606    -0.290 -0.376    0.250       0.138    0.379     0.561 
#>  8  0.465  -0.507     0.606 -0.368   -0.865      -0.975    0.313     0.465 
#>  9 -0.478  -0.00110   0.472 -0.398    0.871       0.752   -0.478    -0.478 
#> 10 -0.195  -0.141     0.346  0.0487  -1.05       -1.03    -0.237    -0.195 
#> # … with 990 more rows, and 17 more variables: x2_item1 <dbl>, x2_item2 <dbl>,
#> #   x3_item1 <dbl>, x3_item2 <dbl>, x3_item3 <dbl>, x4_item1 <dbl>,
#> #   x4_item2 <dbl>, x4_item3 <dbl>, x4_item4 <dbl>, x5_item1 <dbl>,
#> #   x5_item2 <dbl>, x5_item3 <dbl>, x5_item4 <dbl>, x5_item5 <dbl>,
#> #   selected_obs <lgl>, selected_true <lgl>, selected_error <lgl>