Simulation of data with measurement error and range-restriction artifacts

This function simulates a psychometric sample and produces correlation matrices, artifact information, and other descriptive statistics that have been affected by measurement error and/or range restriction. It allows the formation of composite variables within the simulation and allows selection to be performed on any or all variables, including composites. By setting the sample size to n = Inf, users can explore the effects of measurement error and/or range restriction on parameters without the influence of sampling error. To generate multiple samples and compile a database of simulated statistics, see the simulate_r_database function.

simulate_r_sample(
  n,
  rho_mat,
  rel_vec = rep(1, ncol(rho_mat)),
  mu_vec = rep(0, ncol(rho_mat)),
  sigma_vec = rep(1, ncol(rho_mat)),
  sr_vec = rep(1, ncol(rho_mat)),
  k_items_vec = rep(1, ncol(rho_mat)),
  wt_mat = NULL,
  sr_composites = NULL,
  var_names = NULL,
  composite_names = NULL,
  n_as_ni = FALSE,
  ...
)

Arguments

n
Number of cases to simulate before performing selection. If Inf, function will simulate parameter values.

rho_mat
Matrix of true-score correlations.

rel_vec
Vector of reliabilities corresponding to the variables in rho_mat.

mu_vec
Vector of means.

sigma_vec
Vector of observed-score standard deviations.

sr_vec
Vector of selection ratios corresponding to the variables in rho_mat (set selection ratios to 1 for variables that should not be used in selection).

k_items_vec
Number of test items comprising each of the variables to be simulated (all are single-item variables by default).

wt_mat
Optional matrix of weights to use in forming a composite of the variables in rho_mat. Matrix should have as many rows (or vector elements) as there are variables in rho_mat.

sr_composites
Optional vector selection ratios for composite variables. If not NULL, sr_composites must have as many elements as there are columns in wt_mat.

var_names
Vector of variable names corresponding to the variables in rho_mat.

composite_names
Optional vector of names for composite variables.

n_as_ni
Logical argument determining whether n specifies the incumbent sample size (TRUE) or the applicant sample size (FALSE; default). This can only be TRUE when only one variable is involved in selection.

...
Further arguments.

Value

A list of study information, including correlations, reliabilities, standard deviations, means, and u ratios for true scores and for observed scores.

Examples

## Generate data for a simple sample with two variables:
simulate_r_sample(n = 1000, rho_mat = matrix(c(1, .5, .5, 1), 2, 2),
          rel_vec = c(.8, .8), sr_vec = c(1, .5), var_names = c("Y", "X"))

#> Results of Simulated Study (Statistics) 
#> --------------------------
#> 
#> Simulated 1000 applicant cases and selected 500 incumbent cases for a selection ratio of 50%.
#> 
#> Observed Applicant Correlations:
#>       Y     X
#> Y 1.000 0.403
#> X 0.403 1.000
#> 
#> Observed Incumbent Correlations:
#>       Y     X
#> Y 1.000 0.239
#> X 0.239 1.000
#> 
#> Observed Descriptive Statistics:
#>                                          Y     X
#> Applicant parallel-forms reliability 0.795 0.792
#> Applicant unstandardized alpha          NA    NA
#> Applicant standardized alpha            NA    NA
#> Incumbent parallel-forms reliability 0.777 0.553
#> Incumbent unstandardized alpha          NA    NA
#> Incumbent standardized alpha            NA    NA
#> u ratio                              0.944 0.604
#> Applicant SD                         1.017 1.023
#> Incumbent SD                         0.960 0.619
#> Applicant mean                       0.000 0.027
#> Incumbent mean                       0.319 0.844

## Generate data for samples with five variables, of which subsets are used to form composites:
rho_mat <- matrix(.5, 5, 5)
diag(rho_mat) <- 1

## Simulate parameters by supply n = Inf
simulate_r_sample(n = Inf, rho_mat = rho_mat,
                rel_vec = rep(.8, 5), sr_vec = c(1, 1, 1, 1, .5),
                wt_mat = cbind(c(0, 0, 0, .3, 1), c(1, .3, 0, 0, 0)), sr_composites = c(.7, .5))

#> Error in .simulate_r_sample_params(n = args$n, rho_mat = args$rho_mat,     mu_vec = args$mu_vec, sigma_vec = args$sigma_vec, rel_vec = args$rel_vec,     sr_vec = args$sr_vec, k_items_vec = args$k_items_vec, wt_mat = args$wt_mat,     sr_composites = args$sr_composites, var_names = args$var_names,     composite_names = args$composite_names): The package 'mvtnorm' is not installed.
#>   'mvtnorm' is required to estimate population parameters under multivariate selection.
#>   Please install 'mvtnorm'.

## Finite sample sizes allow the generation of sample data
simulate_r_sample(n = 1000, rho_mat = rho_mat,
                rel_vec = rep(.8, 5), sr_vec = c(1, 1, 1, 1, .5),
                wt_mat = cbind(c(0, 0, 0, .3, 1), c(1, .3, 0, 0, 0)), sr_composites = c(.7, .5))

#> Results of Simulated Study (Statistics) 
#> --------------------------
#> 
#> Simulated 1000 applicant cases and selected 324 incumbent cases for a selection ratio of 32.4%.
#> 
#> Observed Applicant Correlations:
#>               x1    x2    x3    x4    x5 composite1 composite2
#> x1         1.000 0.394 0.404 0.379 0.394      0.439      0.969
#> x2         0.394 1.000 0.410 0.404 0.424      0.472      0.609
#> x3         0.404 0.410 1.000 0.392 0.435      0.479      0.459
#> x4         0.379 0.404 0.392 1.000 0.396      0.598      0.436
#> x5         0.394 0.424 0.435 0.396 1.000      0.973      0.454
#> composite1 0.439 0.472 0.479 0.598 0.973      1.000      0.506
#> composite2 0.969 0.609 0.459 0.436 0.454      0.506      1.000
#> 
#> Observed Incumbent Correlations:
#>               x1    x2    x3    x4    x5 composite1 composite2
#> x1         1.000 0.073 0.197 0.165 0.163      0.203      0.930
#> x2         0.073 1.000 0.285 0.327 0.240      0.330      0.435
#> x3         0.197 0.285 1.000 0.174 0.302      0.327      0.283
#> x4         0.165 0.327 0.174 1.000 0.166      0.515      0.270
#> x5         0.163 0.240 0.302 0.166 1.000      0.931      0.236
#> composite1 0.203 0.330 0.327 0.515 0.931      1.000      0.305
#> composite2 0.930 0.435 0.283 0.270 0.236      0.305      1.000
#> 
#> Observed Descriptive Statistics:
#>                                          x1    x2     x3     x4    x5
#> Applicant parallel-forms reliability  0.779 0.793  0.801  0.814 0.797
#> Applicant unstandardized alpha           NA    NA     NA     NA    NA
#> Applicant standardized alpha             NA    NA     NA     NA    NA
#> Incumbent parallel-forms reliability  0.612 0.711  0.795  0.779 0.570
#> Incumbent unstandardized alpha           NA    NA     NA     NA    NA
#> Incumbent standardized alpha             NA    NA     NA     NA    NA
#> u ratio                               0.682 0.894  0.896  0.917 0.623
#> Applicant SD                          0.973 1.010  1.001  0.975 1.009
#> Incumbent SD                          0.663 0.904  0.897  0.894 0.629
#> Applicant mean                       -0.084 0.000 -0.013 -0.005 0.006
#> Incumbent mean                        0.730 0.636  0.523  0.487 0.921
#>                                      composite1 composite2
#> Applicant parallel-forms reliability      0.831      0.819
#> Applicant unstandardized alpha               NA         NA
#> Applicant standardized alpha                 NA         NA
#> Incumbent parallel-forms reliability      0.630      0.661
#> Incumbent unstandardized alpha               NA         NA
#> Incumbent standardized alpha                 NA         NA
#> u ratio                                   0.626      0.652
#> Applicant SD                              1.157      1.127
#> Incumbent SD                              0.724      0.735
#> Applicant mean                            0.005     -0.084
#> Incumbent mean                            1.067      0.920