Truncation function for means

This function computes the mean of a normal distributions that has been truncated at one or both ends.

truncate_mean(a = -Inf, b = Inf, mean = 0, sd = 1)

Arguments

a
Quantile (i.e., cut score) below which scores should be censored from the distribution.

b
Quantile (i.e., cut score) above which scores should be censored from the distribution.

mean
Scalar mean or vector of means.

sd
Scalar standard deviation or vector of standard deviations.

Value

A vector of truncated means.

Examples

truncate_mean(a = -1, b = 3, mean = 0, sd = 1)

#> [1] 0.2827861

truncate_mean(a = 1, b = Inf, mean = 0, sd = 1)

#> [1] 1.525135

truncate_mean(a = c(-1, 1), b = c(3, Inf), mean = 0, sd = 1)

#> [1] 0.2827861 1.5251353