Estimate the error variance of multiple correlations and squared multiple correlations for linear regressions

Estimate the error variance of multiple correlations and squared multiple correlations for linear regressions

var_error_mult_R(R, n, p)

var_error_mult_Rsq(Rsq, n, p)

var_error_R(R, n, p)

var_error_Rsq(Rsq, n, p)

Arguments

R
Vector of multiple correlation coefficients.

n
Vector of sample sizes.

p
Vector of numbers of predictors in the model.

Rsq
Vector of squared multiple correlation coefficients.

Value

A vector of sampling-error variances.

Details

The sampling variance of a multiple correlation is approximately:

$$var_{e}=\frac{(1-R^{2})^{2}(n-p-1)^{2}}{(n^{2}-1)(n+3)}$$

The sampling variance of a squared multiple correlation is approximately:

$$var_{e}=\frac{4R^{2}(1-R^{2})^{2}(n-p-1)^{2}}{(n^{2}-1)(n+3)}$$

References

Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2003). Applied multiple regression/correlation analysis for the behavioral sciences (3rd ed.). Mahwah, NJ: Erlbaum. https://doi.org/10/crtf. p. 88.

Olkin, I., & Finn, J. D. (1995). Correlations redux. Psychological Bulletin, 118(1), 155–164. https://doi.org/10/bt48vd

Examples

var_error_mult_R(R = .5, n = 30, p = 4)

#> [1] 0.01185029

var_error_mult_R(R = .5, n = 30, p = 4)

#> [1] 0.01185029

var_error_mult_Rsq(Rsq = .25, n = 30, p = 4)

#> [1] 0.01185029

var_error_mult_Rsq(Rsq = .25, n = 30, p = 4)

#> [1] 0.01185029