Estimate the error variance of reliability estimates

Estimate the error variance of reliability estimates

var_error_rel(rel, n, rel_type = "alpha", k_items = NULL)

Arguments

rel
Vector of reliability estimates.

n
Vector of sample sizes.

rel_type
Character vector indicating the type(s) of reliabilities being analyzed. See documentation for ma_r for a full list of acceptable reliability types. NOTE: Currently, only alpha has its own dedicated error-variance estimate; the error variance of other reliability types is estimated using the generic definition of reliability as the squared correlation between observed scores and true scores.

k_items
Optional numeric vector indicating the number of items in each scale for which reliabilities are being analyzed.

Value

A vector of sampling-error variances.

Details

The sampling variance of a reliability coefficient is:

$$var_{e}=\frac{4r_{XX}(1-r_{XX})^{2}}{n-1}$$

For the equation to estimate the variance of coefficient alpha, see Duhachek and Iacobucci (2004).

References

Dahlke, J. A., & Wiernik, B. M. (2019). Not restricted to selection research: Accounting for indirect range restriction in organizational research. Organizational Research Methods. Advance online publication. https://doi.org/10.1177/1094428119859398

Duhachek, A., & Iacobucci, D. (2004). Alpha’s standard error (ASE): An accurate and precise confidence interval estimate. Journal of Applied Psychology, 89(5), 792–808. https://doi.org/10.1037/0021-9010.89.5.792

Examples

var_error_rel(rel = .8, n = 100)

#> [1] 0.001292929

var_error_rel(rel = .8, n = 100, rel_type = "alpha", k_items = 10)

#> [1] 0.0008888889