Estimate the error variance of u ratios

Estimate the error variance of u ratios

var_error_u(u, ni, na = NA, dependent_sds = FALSE)

Arguments

u
Vector of u ratios.

ni
Vector of incumbent-group sample sizes.

na
Vector of applicant-group sample sizes.

dependent_sds
Logical vector identifying whether each u ratio is based on standard deviations from independent samples (FALSE) or based on standard deviations from an applicant sample and an incumbent sample that is a subset of that applicant sample (TRUE).

Value

A vector of sampling-error variances.

Details

The sampling variance of a u ratio is computed differently for independent samples (i.e., settings where the referent unrestricted standard deviation comes from an different sample than the range-restricted standard deviation) than for dependent samples (i.e., unrestricted samples from which a subset of individuals are selected to be in the incumbent sample).

The sampling variance for independent samples (the more common case) is:

$$var_{e}=\frac{u^{2}}{2}\left(\frac{1}{n_{i}-1}+\frac{1}{n_{a}-1}\right)$$

and the sampling variance for dependent samples is:

$$var_{e}=\frac{u^{2}}{2}\left(\frac{1}{n_{i}-1}-\frac{1}{n_{a}-1}\right)$$

where u is the u ratio, \(n_{i}\) is the incumbent sample size, and \(n_{a}\) is the applicant sample size.

References

Dahlke, J. A., & Wiernik, B. M. (2019). Not restricted to selection research: Accounting for indirect range restriction in organizational research. Organizational Research Methods. Advance online publication. https://doi.org/10.1177/1094428119859398

Examples

var_error_u(u = .8, ni = 100, na = 200)

#> [1] 0.004840363

var_error_u(u = .8, ni = 100, na = NA)

#> [1] 0.003232323